### 2023-08-28 c1b23b8
Bugfix in syncronization. Now all records are removed for which there are no observations!
