### 2023-08-28 67cf5db
bugfix to remove nan lines when reading the template
