### 2023-08-28 734c612
return plotting axes for heatmap
