+++ pyproject.toml
@@ -21,7 +21,7 @@ pyproj = "~3.4"
 mapclassify = "^2.4.0"
 earthengine-api = "^0.1.340"
 geemap = '^0.20.0'
-titanlib = "^0.3" #can give errors due to C-compiler in windows ....
+#titanlib = "^0.3" #can give errors due to C-compiler in windows ....
 
 
 
