+++ .github/workflows/os_istall_test.yml
@@ -15,12 +15,8 @@ jobs:
             uses: actions/setup-python@v4
             with:
               python-version: '3.9'
-
-          - name: PyPi install on ${{matrix.os}}
-            run: python pip install metobs-toolkit
           - name: install this branch-version on ${{matrix.os}}
             run: |
-                pip install poetry
                 poetry update
                 poetry install --all-extras
                 poetry build
@@ -29,6 +25,8 @@ jobs:
                 cd tests
                 cd push_test
                 poetry run python import_test.py
+          - name: PyPi install on ${{matrix.os}}
+            run: pip install metobs-toolkit
 
 
 
