+++ .github/workflows/os_istall_test.yml
@@ -10,7 +10,6 @@ jobs:
             matrix:
                 os: [ubuntu-latest, windows-latest, macos-latest]
         steps:
-        steps:
           - uses: actions/checkout@v3
           - uses: actions/setup-python@v4
             with:
