+++ .github/workflows/os_istall_test.yml
@@ -47,11 +47,11 @@ jobs:
       #----------------------------------------------
       #          install and run linters
       #----------------------------------------------
-      - run: python -m pip install black flake8 isort
-      - run: |
-          flake8 .
-          black . --check
-          isort .
+       - run: python -m pip install black flake8 isort
+       - run: |
+           flake8 metobs_toolkit
+          # black . --check
+          # isort .
   test:
     needs: linting
     strategy:
