+++ .github/workflows/os_istall_test.yml
@@ -49,9 +49,9 @@ jobs:
       #----------------------------------------------
        - run: python -m pip install black flake8 isort
        - run: |
-           flake8 metobs_toolkit
-          # black . --check
-          # isort .
+           flake8 metobs_toolkit/.
+           black . --check
+           isort .
   test:
     needs: linting
     strategy:
