+++ .github/workflows/os_istall_test.yml
@@ -4,9 +4,6 @@ jobs:
   linting:
     runs-on: ubuntu-latest
     steps:
-      #----------------------------------------------
-      #       check-out repo and set-up python
-      #----------------------------------------------
       - uses: actions/checkout@v3
       - uses: actions/setup-python@v4
       #----------------------------------------------
