+++ .github/workflows/os_istall_test.yml
@@ -1,9 +1,14 @@
 name: Installation testing
-on: [push]
+
+on: push
+
 jobs:
   linting:
     runs-on: ubuntu-latest
     steps:
+      #----------------------------------------------
+      #       check-out repo and set-up python
+      #----------------------------------------------
       - uses: actions/checkout@v3
       - uses: actions/setup-python@v4
       #----------------------------------------------
@@ -17,11 +22,11 @@ jobs:
       #----------------------------------------------
       #          install and run linters
       #----------------------------------------------
-       - run: python -m pip install black flake8 isort
-       - run: |
-           flake8 metobs_toolkit/.
-           black . --check
-           isort .
+      - run: python -m pip install black flake8 isort
+      - run: |
+          flake8 metobs_toolkit
+          black . --check
+          isort .
   test:
     needs: linting
     strategy:
@@ -29,7 +34,7 @@ jobs:
       matrix:
         os: [ "ubuntu-latest", "macos-latest" ]
         python-version: [ "3.7", "3.8", "3.9", "3.10", "3.11" ]
-        #django-version: ["3", "4" ]
+        django-version: ["3", "4" ]
     runs-on: ${{ matrix.os }}
     steps:
       #----------------------------------------------
@@ -70,13 +75,17 @@ jobs:
       #----------------------------------------------
       - name: Install library
         run: poetry install --no-interaction
+      #----------------------------------------------
+      #    add matrix specifics and run test suite
+      #----------------------------------------------
+      - name: Install django ${{ matrix.django-version }}
+        run: |
+          source .venv/bin/activate
+          pip install "Django==${{ matrix.django-version }}"
       - name: Run tests
         run: |
           source .venv/bin/activate
           pytest tests/push_test/import_test.py
           coverage report
-
-
-
 
 
