+++ .github/workflows/os_istall_test.yml
@@ -24,7 +24,7 @@ jobs:
       #----------------------------------------------
       - run: python -m pip install black flake8 isort
       - run: |
-          flake8 metobs_toolkit
+          flake8 --output-file=flake8_output.txt metobs_toolkit/
           black . --check
           isort .
   test:
