+++ .github/workflows/os_istall_test.yml
@@ -3,26 +3,26 @@ name: Installation testing
 on: push
 
 jobs:
-  # linting:
-  #   runs-on: ubuntu-latest
-  #   steps:
-  #     #----------------------------------------------
-  #     #       check-out repo and set-up python
-  #     #----------------------------------------------
-  #     - uses: actions/checkout@v3
-  #     - uses: actions/setup-python@v4
-  #     #----------------------------------------------
-  #     #        load pip cache if cache exists
-  #     #----------------------------------------------
-  #     - uses: actions/cache@v3
-  #       with:
-  #         path: ~/.cache/pip
-  #         key: ${{ runner.os }}-pip
-  #         restore-keys: ${{ runner.os }}-pip
-  #     #----------------------------------------------
-  #     #          install and run linters
-  #     #----------------------------------------------
-  #     - run: python -m pip install black flake8 isort
+   linting:
+     runs-on: ubuntu-latest
+     steps:
+       #----------------------------------------------
+       #       check-out repo and set-up python
+       #----------------------------------------------
+       - uses: actions/checkout@v3
+       - uses: actions/setup-python@v4
+       #----------------------------------------------
+       #        load pip cache if cache exists
+       #----------------------------------------------
+       - uses: actions/cache@v3
+         with:
+           path: ~/.cache/pip
+           key: ${{ runner.os }}-pip
+           restore-keys: ${{ runner.os }}-pip
+       #----------------------------------------------
+       #          install and run linters
+       #----------------------------------------------
+       - run: python -m pip install black flake8 isort
   #     - run: |
   #         flake8 --output-file=flake8_output.txt --exit-zero metobs_toolkit/
   #         black . --check
