+++ .github/workflows/os_istall_test.yml
@@ -23,10 +23,10 @@ jobs:
        #          install and run linters
        #----------------------------------------------
        - run: python -m pip install black flake8 isort
-  #     - run: |
-  #         flake8 --output-file=flake8_output.txt --exit-zero metobs_toolkit/
-  #         black . --check
-  #         isort .
+       - run: |
+           flake8 --output-file=flake8_output.txt --exit-zero metobs_toolkit/
+           black . --check
+           isort .
   test:
     needs: linting
     strategy:
