+++ .github/workflows/os_istall_test.yml
@@ -33,7 +33,7 @@ jobs:
       fail-fast: true
       matrix:
         os: [ "ubuntu-latest", "macos-latest" ]
-        python-version: [ "3.7", "3.8", "3.9", "3.10", "3.11" ]
+        python-version: [ "3.8", "3.9", "3.10", "3.11" ]
         django-version: ["3", "4" ]
     runs-on: ${{ matrix.os }}
     steps:
