### 2023-08-31 c5e2b5e
fixed issues with extracting tz and passing in trough
