### 2023-08-31 3036766
restore obstypes in metadf of station class
