+++ metobs_toolkit/station.py
@@ -6,7 +6,7 @@ class.
 
 A Station holds all observations of one station.
 """
-
+import pandas as pd
 from metobs_toolkit import dataset
 
 
@@ -40,3 +40,17 @@ class Station(dataset.Dataset):
         self._applied_qc = _applied_qc
 
         self._istype = "Station"
+        self.setup_metadata_dtyes()
+
+    def setup_metadata_dtyes(self):
+        """Make sure the dtypes are not lost when subsetting."""
+        numeric_columns = ['lat', 'lon']
+        timedelta_columns = ['assumed_import_frequency', 'dataset_resolution']
+
+        for col in numeric_columns:
+            if col in self.metadf.columns:
+                self.metadf[col] = pd.to_numeric(self.metadf[col])
+
+        for col in timedelta_columns:
+            if col in self.metadf.columns:
+                self.metadf[col] = pd.to_timedelta(self.metadf[col])
