+++ .github/workflows/os_istall_test.yml
@@ -60,6 +60,7 @@ jobs:
       - name: Run tests
         run: |
           source .venv/bin/activate
+          pip install pytest
           pytest tests/push_test/import_test.py
           coverage report
 
