+++ .github/workflows/os_istall_test.yml
@@ -62,6 +62,5 @@ jobs:
           source .venv/bin/activate
           pip install pytest
           pytest tests/push_test/import_test.py
-          coverage report
 
 
