+++ .github/workflows/os_istall_test.yml
@@ -7,9 +7,8 @@ jobs:
     strategy:
       fail-fast: true
       matrix:
-        os: [ "ubuntu-latest", "macos-latest" ]
+        os: [ "ubuntu-latest", "macos-latest", "windows-latest" ]
         python-version: [ "3.8", "3.9", "3.10", "3.11" ]
-        django-version: ["3", "4" ]
     runs-on: ${{ matrix.os }}
     steps:
       #----------------------------------------------
@@ -50,13 +49,9 @@ jobs:
       #----------------------------------------------
       - name: Install library
         run: poetry install --no-interaction
-      #----------------------------------------------
-      #    add matrix specifics and run test suite
-      #----------------------------------------------
-      - name: Install django ${{ matrix.django-version }}
-        run: |
-          source .venv/bin/activate
-          pip install "Django==${{ matrix.django-version }}"
+      # =============================================================================
+      # run import test.
+      # =============================================================================
       - name: Run tests
         run: |
           source .venv/bin/activate
