+++ .github/workflows/os_istall_test.yml
@@ -8,7 +8,7 @@ jobs:
       fail-fast: true
       matrix:
         os: [ "ubuntu-latest", "macos-latest", "windows-latest" ]
-        python-version: [ "3.8", "3.9", "3.10", "3.11" ]
+        python-version: [ "3.8", "3.9", "3.10"]
     runs-on: ${{ matrix.os }}
     steps:
       #----------------------------------------------
