### 2023-08-31 cb2465b
add testing to os installation action
