+++ .github/workflows/os_istall_test.yml
@@ -60,7 +60,7 @@ jobs:
   run-tests:
     needs: test-os-installations
     build:
-      runs-on: ${{ matrix.os }}
+      # runs-on: ${{ matrix.os }}
       strategy:
         # You can use PyPy versions in python-version.
         # For example, pypy-2.7 and pypy-3.8
@@ -69,7 +69,6 @@ jobs:
           # os: [ubuntu-latest, windows-latest]
           os: [ubuntu-latest]
           node-version: [12.x]
-
       steps:
         - uses: actions/checkout@v3
         - name: Set up Python ${{ matrix.python-version }}
@@ -80,25 +79,6 @@ jobs:
         # You can test your matrix by printing the current Python version
         - name: Display Python version
           run: python -c "import sys; print(sys.version)"
-        - name: Install poetry
-          run: |
-              pip install poetry
-
-        - name: Install dependencies and build
-          run: |
-              poetry update
-              poetry install --all-extras
-              poetry build
-        - name: Run tests
-          run: |
-              cd tests
-              cd push_test
-              poetry run python gap_and_fill_test.py
-              poetry run python IO_test.py
-              poetry run python qc_test.py
-              poetry run python breaking_test.py
-              #poetry run python gui_launch_test.py
-              poetry run python analysis_test.py
 
 
 
