+++ .github/workflows/os_istall_test.yml
@@ -10,7 +10,7 @@ jobs:
         os: [ "ubuntu-latest", "macos-latest", "windows-latest" ]
         python-version: [ "3.8", "3.9", "3.10"]
     runs-on: ${{ matrix.os }}
-    steps:
+    steps-a:
       #----------------------------------------------
       #       check-out repo and set-up python
       #----------------------------------------------
@@ -69,7 +69,7 @@ jobs:
           # os: [ubuntu-latest, windows-latest]
           os: [ubuntu-latest]
           node-version: [12.x]
-      steps:
+      steps-b:
         - uses: actions/checkout@v3
         - name: Set up Python ${{ matrix.python-version }}
           uses: actions/setup-python@v4
