+++ .github/workflows/os_istall_test.yml
@@ -10,7 +10,7 @@ jobs:
         os: [ "ubuntu-latest", "macos-latest", "windows-latest" ]
         python-version: [ "3.8", "3.9", "3.10"]
     runs-on: ${{ matrix.os }}
-    steps-a:
+    steps:
       #----------------------------------------------
       #       check-out repo and set-up python
       #----------------------------------------------
@@ -59,19 +59,19 @@ jobs:
           python tests/push_test/import_test.py
   run-tests:
     needs: test-os-installations
-    build:
-      # runs-on: ${{ matrix.os }}
-      strategy:
-        # You can use PyPy versions in python-version.
-        # For example, pypy-2.7 and pypy-3.8
-        matrix:
-          python-version: ["3.9", "3.10"]
-          # os: [ubuntu-latest, windows-latest]
-          os: [ubuntu-latest]
-          node-version: [12.x]
-      steps-b:
+    # build:
+    #   # runs-on: ${{ matrix.os }}
+    #   strategy:
+    #     # You can use PyPy versions in python-version.
+    #     # For example, pypy-2.7 and pypy-3.8
+    #     matrix:
+    #       python-version: ["3.9", "3.10"]
+    #       # os: [ubuntu-latest, windows-latest]
+    #       os: [ubuntu-latest]
+    #       node-version: [12.x]
+      steps:
         - uses: actions/checkout@v3
-        - name: Set up Python ${{ matrix.python-version }}
+        - name: Set up Python
           uses: actions/setup-python@v4
           with:
             python-version: ${{ matrix.python-version }}
