+++ .github/workflows/os_istall_test.yml
@@ -11,9 +11,9 @@ jobs:
         python-version: [ "3.8", "3.9", "3.10"]
     runs-on: ${{ matrix.os }}
     steps:
-      #----------------------------------------------
+      #---
       #       check-out repo and set-up python
-      #----------------------------------------------
+      #----
       - name: Check out repository
         uses: actions/checkout@v3
       - name: Set up python ${{ matrix.python-version }}
@@ -21,41 +21,29 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
-      #----------------------------------------------
+      #--
       #  -----  install & configure poetry  -----
-      #----------------------------------------------
+      #--
       - name: Install Poetry
         uses: snok/install-poetry@v1
         with:
           virtualenvs-create: true
           virtualenvs-in-project: true
-      #----------------------------------------------
-      #       load cached venv if cache exists
-      #----------------------------------------------
       - name: Load cached venv
-        id: cached-poetry-dependencies
+        id: cached-pip-wheels
         uses: actions/cache@v3
         with:
-          path: .venv
+          path: ~/.cache
           key: venv-${{ runner.os }}-${{ steps.setup-python.outputs.python-version }}-${{ hashFiles('**/poetry.lock') }}
-      #----------------------------------------------
-      # install dependencies if cache does not exist
-      #----------------------------------------------
       - name: Install dependencies
-        if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
         run: poetry install --no-interaction --no-root
-      #----------------------------------------------
-      # install your root project, if required
-      #----------------------------------------------
       - name: Install library
         run: poetry install --no-interaction
-      # =============================================================================
-      # run import test.
-      # =============================================================================
+      - run: |
+          source $VENV
+          pytest --version
       - name: Run tests
         run: |
           source .venv/bin/activate
           pip install pytest
           python tests/push_test/import_test.py
-
-
