+++ .github/workflows/os_istall_test.yml
@@ -7,7 +7,7 @@ jobs:
     strategy:
       fail-fast: true
       matrix:
-        os: [ "ubuntu-latest", "macos-latest", "windows-latest" ]
+        os: [ "windows-latest" ]
         python-version: [ "3.8", "3.9", "3.10"]
     defaults:
       run:
