+++ .github/workflows/os_istall_test.yml
@@ -38,6 +38,8 @@ jobs:
         with:
           path: ~/.cache
           key: venv-${{ runner.os }}-${{ steps.setup-python.outputs.python-version }}-${{ hashFiles('**/poetry.lock') }}
+      - name: Update lock
+        run: poetry update
       - name: Install dependencies
         run: poetry install --no-interaction --no-root
       - name: Install library
