+++ .github/workflows/os_istall_test.yml
@@ -8,7 +8,7 @@ jobs:
       fail-fast: true
       matrix:
         os: [ "windows-latest" ]
-        python-version: [ "3.8", "3.9", "3.10"]
+        python-version: ["3.9", "3.10"]
     defaults:
       run:
         shell: bash
