### 2023-08-31 bcb4edc
reduce version of poetry because of bug
