+++ .github/workflows/os_istall_test.yml
@@ -30,6 +30,7 @@ jobs:
       - name: Install Poetry
         uses: snok/install-poetry@v1
         with:
+          version: 1.3.2
           virtualenvs-create: true
           virtualenvs-in-project: true
       - name: Load cached venv
