### 2023-08-31 dd62768
framework for a python buddy check equivalent of TITAN
