+++ metobs_toolkit/settings_files/default_formats_settings.py
@@ -97,6 +97,7 @@ plot_settings["color_mapper"] = {
     "repetitions": "#056ff0",
     "step": "#05d4f0",
     "window_variation": "#05f0c9",
+    'toolkit_buddy_check': '#8300c4',
     "titan_buddy_check": '#8300c4',
     "titan_sct_resistant_check": '#c17fe1',
     # missing and gap
