### 2023-09-01 592b063
write test for buddy check
