### 2023-09-01 017e7b3
added haversine approximation for global usage
