### 2023-09-04 c47406f
remove invalid prefixes of DOI's
