+++ docs/paper/paper.bib
@@ -45,7 +45,7 @@ volume = {19},
 pages = {170-191},
 year = {2017},
 issn = {2212-0955},
-doi = {https://doi.org/10.1016/j.uclim.2017.01.006},
+doi = {10.1016/j.uclim.2017.01.006},
 url = {https://www.sciencedirect.com/science/article/pii/S2212095517300068},
 author = {Fred Meier and Daniel Fenner and Tom Grassmann and Marco Otto and Dieter Scherer},
 keywords = {Urban climate observations, Crowdsourcing air temperature, Data quality assessment, Urban heat island, Netatmo weather station, Berlin},
@@ -60,7 +60,7 @@ volume = {35},
 number = {11},
 pages = {3185-3203},
 keywords = {Internet of things, Big data, Citizen science, Sensors, Amateur, Applications},
-doi = {https://doi.org/10.1002/joc.4210},
+doi = {10.1002/joc.4210},
 url = {https://rmets.onlinelibrary.wiley.com/doi/abs/10.1002/joc.4210},
 eprint = {https://rmets.onlinelibrary.wiley.com/doi/pdf/10.1002/joc.4210},
 abstract = {ABSTRACT Crowdsourcing is traditionally defined as obtaining data or information by enlisting the services of a (potentially large) number of people. However, due to recent innovations, this definition can now be expanded to include ‘and/or from a range of public sensors, typically connected via the Internet.’ A large and increasing amount of data is now being obtained from a huge variety of non-traditional sources – from smart phone sensors to amateur weather stations to canvassing members of the public. Some disciplines (e.g. astrophysics, ecology) are already utilizing crowdsourcing techniques (e.g. citizen science initiatives, web 2.0 technology, low-cost sensors), and while its value within the climate and atmospheric science disciplines is still relatively unexplored, it is beginning to show promise. However, important questions remain; this paper introduces and explores the wide-range of current and prospective methods to crowdsource atmospheric data, investigates the quality of such data and examines its potential applications in the context of weather, climate and society. It is clear that crowdsourcing is already a valuable tool for engaging the public, and if appropriate validation and quality control procedures are adopted and implemented, it has much potential to provide a valuable source of high temporal and spatial resolution, real-time data, especially in regions where few observations currently exist, thereby adding value to science, technology and society.},
@@ -75,7 +75,7 @@ volume = {146},
 number = {730},
 pages = {1999-2049},
 keywords = {climate reanalysis, Copernicus Climate Change Service, data assimilation, ERA5, historical observations},
-doi = {https://doi.org/10.1002/qj.3803},
+doi = {10.1002/qj.3803},
 url = {https://rmets.onlinelibrary.wiley.com/doi/abs/10.1002/qj.3803},
 eprint = {https://rmets.onlinelibrary.wiley.com/doi/pdf/10.1002/qj.3803},
[…]
