+++ docs/paper/paper.bib
@@ -20,6 +20,7 @@ language     = {eng},
 pages        = {19},
 title        = {The urban climate of Ghent, Belgium : a case study combining a high-accuracy monitoring network with numerical simulations},
 url          = {http://dx.doi.org/10.1016/j.uclim.2019.100565},
+doi          = {10.1016/j.uclim.2019.100565},
 volume       = {31},
 year         = {2020}
 }
