### 2023-09-04 0f1f5b3
code analysis on geometry functions
