### 2023-09-04 80d6150
code analysis of missing observations
