+++ metobs_toolkit/missingobs.py
@@ -10,13 +10,13 @@ fill them.
 
 import pandas as pd
 import numpy as np
-from datetime import datetime, timedelta
+from datetime import timedelta
 import logging
 
 from metobs_toolkit.df_helpers import (
     _find_closes_occuring_date,
     xs_save,
-    )
+)
 
 logger = logging.getLogger(__name__)
 
@@ -28,14 +28,14 @@ logger = logging.getLogger(__name__)
 
 
 class Missingob_collection:
-    """ Class object handling a set of missing observations. """
-    def __init__(self, missing_obs_series):
+    """Class object handling a set of missing observations."""
 
+    def __init__(self, missing_obs_series):
+        """Init missing observations."""
         missing_obs_series.name = 'datetime'
         missing_obs_series.index.name = 'name'
 
-
-        missing_obs_df = missing_obs_series.reset_index() #needed to find duplicates
+        missing_obs_df = missing_obs_series.reset_index()  # needed to find duplicates
         missing_obs_df = missing_obs_df.drop_duplicates()
         missing_obs_series = missing_obs_df.set_index('name')['datetime']
         missing_obs_series = missing_obs_series.sort_index()
@@ -51,10 +51,11 @@ class Missingob_collection:
         self.fill_technique = None
 
     def __add__(self, other):
+        """Append two collections of missing observations."""
         comb_series = pd.concat([self.series, other.series])
 
         # drop duplicates and sort
-        comb_df = comb_series.reset_index() #needed to find duplicates
+        comb_df = comb_series.reset_index()  # needed to find duplicates
         comb_df = comb_df.drop_duplicates()
         comb_series = comb_df.set_index('name')['datetime']
         comb_series = comb_series.sort_index()
@@ -64,29 +65,31 @@ class Missingob_collection:
         comb_idx = comb_idx.set_index(['name', 'datetime'])
         self.idx = comb_idx.index
         return self
+
     def __len__(self):
+        """Get number of missing observations."""
         return self.series.shape[0]
 
     def __str__(self):
+        """Print overview info of missing observations."""
         if self.series.empty:
-            return f'Empty missing observations.'
+            return 'Empty missing observations.'
 
         if not self.fill_df.empty:
             return f'Missing observations with filled ({self.fill_technique}) \
                 values: \n {self.fill_df} \n Original missing observations on import: \n {self.idx}'
 
         return f'Missing observations: \n {self.series}'
+
     def __repr__(self):
+        """Print overview info of missing observations."""
         return self.__str__()
 
-
     def get_info(self, max_disp_list=7):
-        """
-        Print out detailed information on the missing observations.
+        """Print out detailed information on the missing observations.
 
         Parameters
         ----------
-
         max_disp_list : int, optional
             Max size of lists to print out. If listsize is larger, the length of
             the list is printed. The default is 7.
@@ -96,10 +99,9 @@ class Missingob_collection:
         None.
 
         """
-
-        print(f'\n -------- Missing observations info -------- \n')
+        print('\n -------- Missing observations info -------- \n')
         if self.series.empty:
-            print(f'Empty missing observations.')
+            print('Empty missing observations.')
             return
         print('(Note: missing observations are defined on the frequency estimation of the native dataset.)')
 
@@ -116,29 +118,24 @@ class Missingob_collection:
             print(f'  * For {len(stations)} stations')
 
         if self.fill_df.empty:
-            print(f'  * The missing observations are not filled.')
+            print('  * The missing observations are not filled.')
         else:
             filled_obstypes = [obs for obs in self.fill_df.columns if not obs.endswith('_final_label')]
-            print(f'  * Missing observations are filled with {{self.fill_technique}} for: ')
+            print(f'  * Missing observations are filled with {self.fill_technique} for: ')
             for obstype in filled_obstypes:
[…]
