+++ metobs_toolkit/modeldata.py
@@ -10,13 +10,19 @@ import pickle
 import pandas as pd
 import logging
 
-from metobs_toolkit.df_helpers import init_multiindexdf, conv_tz_multiidxdf, xs_save, multiindexdf_datetime_subsetting
+from metobs_toolkit.df_helpers import (init_multiindexdf,
+                                       conv_tz_multiidxdf,
+                                       xs_save,
+                                       multiindexdf_datetime_subsetting)
 
-from metobs_toolkit.landcover_functions import connect_to_gee, gee_extract_timeseries
+from metobs_toolkit.landcover_functions import (connect_to_gee,
+                                                gee_extract_timeseries)
 
-from metobs_toolkit.plotting_functions import model_timeseries_plot, timeseries_plot
+from metobs_toolkit.plotting_functions import (model_timeseries_plot,
+                                               timeseries_plot)
 
-from metobs_toolkit.convertors import convert_to_toolkit_units, standard_tlk_units
+from metobs_toolkit.convertors import (convert_to_toolkit_units,
+                                       standard_tlk_units)
 
 from metobs_toolkit.settings import Settings
 
@@ -28,7 +34,10 @@ logger = logging.getLogger(__name__)
 
 
 class Modeldata:
+    """Class holding data and methods for a modeldata-timeseries."""
+
     def __init__(self, modelname):
+        """Initialize modeldata."""
         self.df = init_multiindexdf()
         self.modelname = modelname
 
@@ -36,15 +45,15 @@ class Modeldata:
         self.mapinfo = self._settings.gee["gee_dataset_info"]
         self.mapinfo.update(self._settings.alaro['info'])
 
-        self._df_units = {} #the units of the data stored in the df
-        self.df_tz = 'UTC'# the timezone of the datetimes stored in the df
+        self._df_units = {}  # the units of the data stored in the df
+        self.df_tz = 'UTC'  # the timezone of the datetimes stored in the df
 
-        self._is_alaro25=False
-
+        self._is_alaro25 = False
 
     def __str__(self):
+        """Print overview information of the modeldata."""
         if self.df.empty:
-            return f'Empty Modeldata instance.'
+            return 'Empty Modeldata instance.'
         n_stations = self.df.index.get_level_values('name').unique().shape[0]
         obstypes = self.df.columns.to_list()
         startdt = self.df.index.get_level_values('datetime').min()
@@ -58,14 +67,12 @@ class Modeldata:
     * From {startdt} --> {enddt} (with tz={self.df_tz}) \n \n (Data is stored in the .df attribute)")
 
     def __repr__(self):
+        """Print overview information of the modeldata."""
         return self.__str__()
-
 
     def add_gee_dataset(self, mapname, gee_location, obstype, bandname, units,
                         scale, time_res='1H', is_image=False, is_numeric=True, credentials=''):
-        """
-        Method to add a new gee dataset to the available gee datasets.
-
+        """Add a new gee dataset to the available gee datasets.
 
         Parameters
         ----------
@@ -112,37 +119,28 @@ class Modeldata:
             logger.warning(f'{mapname} is found in the list of known gee datasets: {list(self.mapinfo.keys())}, choose a different mapname.')
             return
 
-
-
         if is_numeric:
-            val_typ='numeric'
+            val_typ = 'numeric'
         else:
             val_typ = 'categorical'
-
-        if is_image:
-            im_bool=True
-        else:
-            im_bool=False
-
-
 
         new_info = {
             mapname: {
                 'location': f'{gee_location}',
                 'usage': 'user defined addition',
-                'band_of_use' :
+                'band_of_use':
                     {f'{obstype}':
                          {'name': f'{bandname}',
                           'units': f'{units}'}
-                         },
-                        'value_type': val_typ,
-                        'dynamical': not bool(is_image),
[…]
