### 2023-09-04 05a99af
code analysis for the plotting functions
