+++ metobs_toolkit/plotting_functions.py
@@ -6,7 +6,7 @@ Created on Fri Oct 21 11:26:52 2022
 @author: thoverga
 """
 
-import warnings
+
 import pandas as pd
 import math
 import numpy as np
@@ -16,9 +16,7 @@ import logging
 
 import matplotlib
 import matplotlib.pyplot as plt
-from matplotlib.colors import Normalize
 from matplotlib.lines import Line2D
-import matplotlib.gridspec as gridspec
 import matplotlib.dates as mdates
 from matplotlib.collections import LineCollection
 
@@ -33,9 +31,11 @@ from metobs_toolkit.df_helpers import xs_save
 
 logger = logging.getLogger(__name__)
 
+
 def folium_plot(mapinfo, band, vis_params, labelnames, layername,
                 basemap='SATELLITE', legendname=None,
                 legendpos='bottomleft'):
+    """Make an interactive folium plot of an Image."""
     # get the ee.Image
     im = get_ee_obj(mapinfo, band)
 
@@ -45,20 +45,22 @@ def folium_plot(mapinfo, band, vis_params, labelnames, layername,
         MAP.add_basemap(basemap)
     MAP.add_layer(im, vis_params, layername)
     if legendname:
-        MAP.add_legend(title=legendname,labels=labelnames, colors=vis_params.get('palette'), position=legendpos)
+        MAP.add_legend(title=legendname,
+                       labels=labelnames,
+                       colors=vis_params.get('palette'),
+                       position=legendpos)
 
     return MAP
 
 
-
 def add_stations_to_folium_map(Map, metadf):
-
+    """Add stations as markers to the folium map."""
     points = metadf['geometry'].to_crs("epsg:4326")
     for station, point in points.items():
-        folium.Marker(location=[ point.y, point.x ],
+        folium.Marker(location=[point.y, point.x],
                       fill_color='#43d9de',
                       popup=station,
-                      radius=8 ).add_to( Map )
+                      radius=8).add_to(Map)
 
     return Map
 
@@ -67,15 +69,14 @@ def add_stations_to_folium_map(Map, metadf):
 # Helpers
 # =============================================================================
 
+
 def map_obstype(obstype, template):
+    """Convert default obstype to the user-specific obstype."""
     return template[obstype].to_dict()
 
 
-
-
 def make_cat_colormapper(catlist, cmapname):
-    """
-    Create a dictionary {cat : color} for a list of categorical values.
+    """Create a dictionary {cat : color} for a list of categorical values.
 
     If the colormap has more colors than the catlist, optimal color distance is
     done. If a colormap has less colors than unique categories, the categories are grourped.
@@ -93,8 +94,7 @@ def make_cat_colormapper(catlist, cmapname):
         {cat: color} where the color is a RGBalpha tuple.
 
     """
-
-    catlist = list(set(catlist)) #get unique categories
+    catlist = list(set(catlist))  # get unique categories
 
     cmap = matplotlib.colormaps[cmapname]
 
@@ -108,7 +108,7 @@ def make_cat_colormapper(catlist, cmapname):
         col_idx = -1
         _cat_index = 0
         for cat in catlist:
-            if _cat_index%same_col_n_groups == 0:
+            if _cat_index % same_col_n_groups == 0:
                 col_idx += 1
             colordict[cat] = cmap(int(col_idx))
             _cat_index += 1
@@ -125,27 +125,58 @@ def make_cat_colormapper(catlist, cmapname):
     return colordict
 
 
-
 # =============================================================================
 # Plotters
 # =============================================================================
 
-def geospatial_plot(
-    plotdf,
-    variable,
-    timeinstance,
-    title,
-    legend,
-    vmin,
[…]
