### 2023-09-04 1c17a5e
proper renaming of diurnal plot to more general cycleplot
