### 2023-09-04 a49dd7a
remove unused argument
