+++ metobs_toolkit/qc_checks.py
@@ -12,11 +12,9 @@ import numpy as np
 import logging
 
 
-
 from metobs_toolkit.df_helpers import (
     init_multiindex,
     init_multiindexdf,
-    init_triple_multiindexdf,
     xs_save
 )
 
@@ -34,11 +32,10 @@ except ModuleNotFoundError:
 # =============================================================================
 
 
-def make_outlier_df_for_check(
-    station_dt_list, obsdf, obstype, flag, stationname=None, datetimelist=None
-):
-    """
-    V5
+def make_outlier_df_for_check(station_dt_list, obsdf, obstype, flag,
+                              stationname=None, datetimelist=None):
+    """Construct obsdf and outliersdf from a list of outlier timestamps.
+
     Helper function to create an outlier dataframe for the given station(s) and datetimes. This will be returned by
     a quality control check and later added to the dastes.outlierdf.
 
@@ -47,27 +44,27 @@ def make_outlier_df_for_check(
     A multiindex dataframe with the relevant observationtypes i.e. the values_in_dict and a specific quality flag column (i.g. the labels) is returned.
 
     Parameters
-
+    ------------
     station_dt_list : MultiIndex or list of tuples: (name, datetime)
         The stations with corresponding datetimes that are labeled as outliers.
-    values_in_dict : Dictionary
-        A Dictionary {columnname: pd.Series()} with the values on which this check is applied.
-    flagcolumnname : String
-        Name of the labels column
+    obsdf : pandas.DataFrame
+        The observations dataframe to update.
+    obstype : str
+        The observation type of the outliers.
     flag : String
-        The label for all the outliers.
+        The label for the outliers.
     stationname : String, optional
         It is possible to give the name of one station. The default is None.
     datetimelist : DatetimeIndex or List, optional
         The outlier timestamps for the stationname. The default is None.
 
     Returns
-
-    check_outliers : pd.Dataframe
-        A multiindex (name -- datetime) datatframe with one column (columname) and all
-        values are the flag.
+    ----------
+    obsdf : pandas.DataFrame
+        The updated observations dataframe.
+     outliersdf : pandas.DataFrame
+         The updated outliers dataframe.
     """
-
     if isinstance(station_dt_list, pd.MultiIndex):
         multi_idx = station_dt_list
 
@@ -112,13 +109,30 @@ def make_outlier_df_for_check(
 
 
 def invalid_input_check(df, checks_info):
+    """Test if values are numeric and not Nan.
+
+    Parameters
+    ----------
+    df : pandas.DataFrame
+        The observations to check the values for. Must contain a column 'name'.
+    checks_info : dict
+        Specific settings for the invalid check test.
+
+    Returns
+    -------
+    df : pandas.DataFrame
+        The observations with NaN values at the location of invalid input.
+    outl_df : pandas.DataFrame
+        The updated outliersdf.
+
+    """
     checkname = "invalid_input"
 
     # fast scan wich stations and obstypes have nan outliers
     groups = (
         df.reset_index()
         .groupby("name")
-        .apply(lambda x: (np.isnan(x).any()) & (np.isnan(x).all() == False))
+        .apply(lambda x: (np.isnan(x).any()) & (np.isnan(x).all() is False))
     )
 
     # extract all obstype that have outliers
@@ -160,22 +174,26 @@ def invalid_input_check(df, checks_info):
 
 
 def duplicate_timestamp_check(df, checks_info, checks_settings):
-    """
-    V3
+    """Test for duplicate timestamps in the observations.
+
     Looking for duplcate timestaps per station. Duplicated records are removed by the method specified in the qc_settings.
 
     Parameters
-
+    ------------
     df : pandas.DataFrame
[…]
