### 2023-09-04 19aba95
code analysis of the setting module
