+++ metobs_toolkit/settings.py
@@ -17,10 +17,13 @@ logger = logging.getLogger(__name__)
 
 
 class Settings:
+    """Class defenition to store all settings."""
+
     # make settingsfiles path
     _settings_files_path = os.path.join(str(Path(__file__).parent), "settings_files")
 
     def __init__(self):
+        """Initiate the settings."""
         logger.info("Initialising settings")
 
         # define thematics in settings. Corresponds to settings files.
@@ -87,8 +90,9 @@ class Settings:
         """
         Update settings on time resolutions of self using the default settings templates.
 
-        :return: No return
-        :rtype: No return
+        Returns
+        -------
+        None.
         """
         logger.debug("Updating time resolution settings.")
         f = open(
@@ -109,12 +113,13 @@ class Settings:
         self.time_settings['freq_estimation_simplify'] = bool(res_settings["freq_estimation_simplify"])
         self.time_settings['freq_estimation_simplify_error'] = res_settings["freq_estimation_simplify_error"]
 
-
     def _update_app_settings(self):
         """
         Update prefered display, print, plot and staticinfo settings of self using the default settings templates.
-        :return: No return
-        :rtype: No return
+
+        Returns
+        -------
+        None.
         """
         logger.debug("Updating app settings.")
         from .settings_files.default_formats_settings import (
@@ -155,8 +160,10 @@ class Settings:
     def _update_qc_settings(self):
         """
         Update quality control settings of self using the default settings templates.
-        :return: No return
-        :rtype: No return
+
+        Returns
+        -------
+        None.
         """
         logger.debug("Updating QC settings.")
         from .settings_files.qc_settings import (check_settings, checks_info,
@@ -171,8 +178,10 @@ class Settings:
     def _update_gap_settings(self):
         """
         Update gap defenition and fill settings of self using the default settings templates.
-        :return: No return
-        :rtype: No return
+
+        Returns
+        -------
+        None.
         """
         logger.debug("Updating gap settings.")
         from .settings_files.gaps_and_missing_settings import (
@@ -194,10 +203,11 @@ class Settings:
 
     def _update_templates(self):
         """
-        Import the default mapper-template, and assign it to be used on the
-        observations and metadata.
-        :return: No return
-        :rtype: No return
+        Import the default mapper-template, and used it on the observations and metadata.
+
+        Returns
+        -------
+        None.
 
         """
         logger.debug("Updating data templates settings.")
@@ -206,12 +216,13 @@ class Settings:
         # Set default templates
         self.templates["template_file"] = default_template_file
 
-
     def _update_gee_settings(self):
         """
         Update the google earth enginge settings using the default settings templates.
-        :return: No return
-        :rtype: No return
+
+        Returns
+        -------
+        None.
         """
         logger.debug("Updating gee settings.")
         from .settings_files.gee_settings import gee_datasets
@@ -221,25 +232,29 @@ class Settings:
     def _update_alaro_settings(self):
         """
         Update the Alaro settings using the default settings templates.
-        :return: No return
-        :rtype: No return
+
+        Returns
+        -------
+        None.
         """
         logger.debug("Updating gee settings.")
         from .settings_files.alaro_25_settings import al25_mapinfo
         self.alaro["info"] = al25_mapinfo
 
-
-
     def update_timezone(self, timezonestr):
         """
         Change the timezone of the input data.
 
-        :param timezonestr: Timezone string of the input observations.
-        :type timezonestr: String
-        :return: None
-        :rtype: None
+        Parameters
+        ------------
+        timezonestr : str
+            Timezone string of the input observations.
+
+        Returns
+        -------
+        None.
         """
-        if not timezonestr in all_timezones:
[…]
