### 2023-09-04 9e37979
conda analysis of the writing files module
