### 2023-09-05 5a9700b
bug fixes in the qc checks after code analysis (== bool does not do the same as is bool)
