### 2023-09-05 ddae925
setup a os matrix for pip install testing of the metobs toolkit
