+++ .github/workflows/os_pip_install_test.yml
@@ -1,0 +1,31 @@
+name: PyPi Installation testing
+
+on:
+  push:
+    branches:
+      - master
+      - dev
+      - pip_on_os_matrix_test
+
+
+jobs:
+  test-os-installations:
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ "ubuntu-latest", "macos-latest", "windows-latest" ]
+        python-version: ["3.9", "3.10"]
+    defaults:
+      run:
+        shell: bash
+    runs-on: ${{ matrix.os }}
+    steps:
+      - name: Check out repository
+        uses: actions/checkout@v3
+      - name: Set up python ${{ matrix.python-version }}
+        id: setup-python
+        uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install from pip
+        run: pip install metobs-toolkit
