+++ README.md
@@ -2,7 +2,7 @@
 [![Installation testing](https://github.com/vergauwenthomas/MetObs_toolkit/actions/workflows/os_istall_test.yml/badge.svg?branch=dev)](https://github.com/vergauwenthomas/MetObs_toolkit/actions/workflows/os_istall_test.yml)
 [![Test pushed version](https://github.com/vergauwenthomas/MetObs_toolkit/actions/workflows/master_test.yml/badge.svg?branch=dev)](https://github.com/vergauwenthomas/MetObs_toolkit/actions/workflows/master_test.yml)
 [![Build and deploy documentation](https://github.com/vergauwenthomas/MetObs_toolkit/actions/workflows/build_and_deploy_doc_dev.yml/badge.svg?branch=dev)](https://github.com/vergauwenthomas/MetObs_toolkit/actions/workflows/build_and_deploy_doc_dev.yml)
-
+[![PyPi Installation testing](https://github.com/vergauwenthomas/MetObs_toolkit/actions/workflows/os_pip_install_test.yml/badge.svg?branch=dev)](https://github.com/vergauwenthomas/MetObs_toolkit/actions/workflows/os_pip_install_test.yml)
 <img src="https://raw.githubusercontent.com/vergauwenthomas/MetObs_toolkit/dev/docs/logo_small.jpeg" alt="drawing" style="width:200px;"/>
 
 # MetObs-toolkit
