+++ .github/workflows/os_pip_install_test.yml
@@ -5,7 +5,6 @@ on:
     branches:
       - master
       - dev
-      - pip_on_os_matrix_test
 
 
 jobs:
