### 2023-09-05 425614c
setup a os matrix for pip install testing of the metobs toolkit (#376)
* setup a os matrix for pip install testing of the metobs toolkit

* remove trigger of branch soon to be deleted
