+++ metobs_toolkit/df_helpers.py
@@ -20,9 +20,9 @@ import logging
 logger = logging.getLogger(__name__)
 
 
-
 def fmt_datetime_argument(dt, target_tz_str):
-    """
+    """Convert naive datetime to tz-aware.
+
     Helper function to format the datetime, a user enters as argument, to the
     correct timezone.
 
@@ -48,18 +48,17 @@ def fmt_datetime_argument(dt, target_tz_str):
 
     # check if datime is timezone aware
     if (dt.tzinfo is not None and dt.tzinfo.utcoffset(dt) is not None):
-        #timezone aware
+        # timezone aware
         dt = dt.astimezone(pytz.timezone(target_tz_str))
 
-    else: #timezon unaware
+    else:  # timezon unaware
         # assume timezone is the timezone of the data!
         dt = pytz.timezone(target_tz_str).localize(dt)
     return dt
 
 
-
 def xs_save(df, key, level, drop_level=True):
-    """ wrapper on pandas xs, but returns an empty df when key is not found """
+    """Similar as pandas xs, but returns an empty df when key is not found."""
     try:
         return df.xs(key, level=level, drop_level=drop_level)
     except KeyError:
@@ -71,29 +70,28 @@ def xs_save(df, key, level, drop_level=True):
 
         levels = [[name] for name in names]
         codes = [[] for name in names]
-        idx = pd.MultiIndex(
-                            levels=levels,
+        idx = pd.MultiIndex(levels=levels,
                             codes=codes,
                             names=names,
                             )
 
-
     return pd.DataFrame(index=idx, columns=columns)
 
 
-
-
 def init_multiindex():
+    """Construct a name-datetime pandas multiindex."""
     return pd.MultiIndex(
         levels=[["name"], ["datetime"]], codes=[[], []], names=["name", "datetime"]
     )
 
 
 def init_multiindexdf():
+    """Construct a name-datetime pandas multiindexdataframe."""
     return pd.DataFrame(index=init_multiindex())
 
 
 def init_triple_multiindex():
+    """Construct a name-datetime-obstype pandas multiindex."""
     my_index = pd.MultiIndex(
         levels=[["name"], ["datetime"], ["obstype"]],
         codes=[[], [], []],
@@ -103,12 +101,12 @@ def init_triple_multiindex():
 
 
 def init_triple_multiindexdf():
+    """Construct a name-datetime-obstype pandas multiindexdataframe."""
     return pd.DataFrame(index=init_triple_multiindex())
 
 
 def format_outliersdf_to_doubleidx(outliersdf):
-    """
-    Convert outliersdf to multiindex dataframe if needed.
+    """Convert outliersdf to multiindex dataframe if needed.
 
     This is applied when the obstype level in the index is not relevant.
 
@@ -124,14 +122,16 @@ def format_outliersdf_to_doubleidx(outliersdf):
         The outliersdfdataframe where the 'obstype' level is dropped, if it was present.
 
     """
-
     if "obstype" in outliersdf.index.names:
         return outliersdf.droplevel("obstype")
     else:
         return outliersdf
 
-def value_labeled_doubleidxdf_to_triple_idxdf(df,value_col_name='value', label_col_name = 'label'):
-    """
+
+def value_labeled_doubleidxdf_to_triple_idxdf(df ,value_col_name='value',
+                                              label_col_name='label'):
+    """Convert double to triple index based on obstype column.
+
     This function converts a double index dataframe with an 'obstype' column,
     and a 'obstype_final_label' column to a triple index dataframe where the
     obstype values are added to the index.
@@ -164,21 +164,21 @@ def value_labeled_doubleidxdf_to_triple_idxdf(df,value_col_name='value', label_c
[…]
