+++ metobs_toolkit/printing.py
@@ -1,17 +1,31 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-Created on Thu Mar  2 14:56:26 2023
+Printing Functions
 
 @author: thoverga
 """
 
 
+def print_dataset_info(dataset, show_all_settings=False, max_disp_n_gaps=5):
+    """Print out settings of a dataset.
 
+    Parameters
+    ----------
+    dataset : metobs_toolkit.Dataset
+        The dataset to print the settings of.
+    show_all_settings : bool, optional
+        If True all settings are printed else a selection of the settings is
+        printed. The default is False.
+    max_disp_n_gaps : int, optional
+        The maximum number of gaps to print detailed information of. The
+        default is 5.
 
+    Returns
+    -------
+    None.
 
-def print_dataset_info(dataset, show_all_settings=False, max_disp_n_gaps = 5):
-
+    """
     print("\n", "--------  General ---------", "\n")
     print(dataset)
 
@@ -41,12 +55,12 @@ def print_dataset_info(dataset, show_all_settings=False, max_disp_n_gaps = 5):
         print(f'{dataset.metadf[relev_columns].head()}')
 
     # "--------  Missing observations ---------")
-    if not dataset.missing_obs is None:
+    if dataset.missing_obs is not None:
         print(dataset.get_missing_obs_info())
 
-    if not dataset.gaps is None:
+    if dataset.gaps is not None:
         print("\n", "--------  Gaps ---------", "\n")
-        if len(dataset.gaps) <=  max_disp_n_gaps:
+        if len(dataset.gaps) <= max_disp_n_gaps:
             print(dataset.get_gaps_info())
         else:
             print(f'The info on {len(dataset.gaps)} is to long to print. Use the .get_gaps_info() to print out the details of all gaps.')
