### 2023-09-06 59f34c3
start of the obervation type module
