+++ metobs_toolkit/obstypes.py
@@ -1,0 +1,247 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Wed Sep  6 16:08:24 2023
+
+@author: thoverga
+"""
+
+import sys
+import logging
+from collections.abc import Iterable
+
+import numpy as np
+logger = logging.getLogger(__name__)
+
+
+
+# =============================================================================
+# Hardcode obstypes, conversion tables and aliases
+# =============================================================================
+
+tlk_std_units = {
+    "temp": 'Celsius',
+    "radiation_temp": 'Celcius',
+    "humidity": '%',
+    "precip": 'mm/m² per hour',
+    "precip_sum": 'mm/m² from midnight',
+    "wind_speed": 'm/s',
+    "wind_gust": 'm/s',
+    "wind_direction": '° from north (CW)',
+    "pressure": 'pa',
+    "pressure_at_sea_level": 'pa'}
+
+
+
+
+# conversion between standard-NAMES and aliases
+all_units_aliases = {
+    "temp": {
+        'Celcius': ['celcius', '°C', '°c'],
+        'Kelvin': ['K', 'kelvin'],
+        'Farenheit': []
+        },
+    "radiation_temp": {
+        'Celcius': ['celcius', '°C', '°c'],
+        'Kelvin': ['K', 'kelvin'],
+        'Farenheit': []
+        },
+    "humidity": {
+        "%": ['percent', 'percentage']
+        },
+    'pressure': {
+        'pa': ['Pascal', 'pascal', 'Pa'],
+        'hpa': ['hecto pascal', 'hPa'],
+        'psi': ['Psi'],
+        'bar': ['Bar']
+        }
+    }
+
+
+
+all_conversion_table = {
+    'temp': {
+        'Kelvin': ["x - 273.15"], #result is in tlk_std_units
+        'Farenheit' : ["x-32.0", "x/1.8"]}, # -->execute from left to write  = (x-32)/1.8
+    'radiation_temp': {
+        'Kelvin': ["x - 273.15"], #result is in tlk_std_units
+        'Farenheit' : ["x-32.0", "x/1.8"]},
+    'humidity': {},
+    'pressure': {
+        'hpa': ["x * 100"],
+        'psi': ['x * 6894.7573'],
+        'bar': ['x * 100000.']
+    }
+
+}
+
+
+
+
+
+
+
+# =============================================================================
+# Observation type class
+# =============================================================================
+
+class Obstype:
+    """Object with all info and methods for a specific observation type."""
+
+    def __init__(self, obsname, std_unit, description=None):
+
+        self.name = str(obsname)
+        self.std_unit = str(std_unit)
+        self.description = str(description)
+
+        self.units_aliases = all_units_aliases[obsname]
+        self.conv_table = all_conversion_table[obsname]
+
+    def __repr__(self):
+        return f"Obstype instance of {self.name}"
+
+    def __str__(self):
+        info_str = f"{self.name} observation with: \n \
+    * standard unit: {self.std_unit} \n \
+    * description: {self.description} \n \
+    * conversions to known units: {self.conv_table} \n"
+        return info_str
+
+    def get_standard_unit(self):
+        print(self.std_unit)
+        return self.std_unit
+
+    def add_unit(self, unit_name, conversion=["x"]):
+        # check if unit name is already known
+        known = self._test_if_unit_is_known(unit_name)
+        if known:
+            return
+
+        # convert expression to list if it is a string
+        if isinstance(conversion, str):
+            conversion = [conversion]
+
+        # add converstion to the table
+        self.conv_table[str(unit_name)] = conversion
+
+        # add to alias table (without aliasses)
[…]
