+++ metobs_toolkit/obstypes.py
@@ -16,14 +16,14 @@ logger = logging.getLogger(__name__)
 
 
 # =============================================================================
-# Hardcode obstypes, conversion tables and aliases
+# Standard toolkit units for each observation type
 # =============================================================================
 
 tlk_std_units = {
     "temp": 'Celsius',
-    "radiation_temp": 'Celcius',
+    "radiation_temp": 'Celsius',
     "humidity": '%',
-    "precip": 'mm/m² per hour',
+    "precip": 'mm/m²',
     "precip_sum": 'mm/m² from midnight',
     "wind_speed": 'm/s',
     "wind_gust": 'm/s',
@@ -31,33 +31,45 @@ tlk_std_units = {
     "pressure": 'pa',
     "pressure_at_sea_level": 'pa'}
 
+# =============================================================================
+# Aliases for units
+# =============================================================================
 
+temp_aliases = {'Celsius': ['celsius', '°C', '°c'],
+                'Kelvin': ['K', 'kelvin'],
+                'Farenheit': []}
+pressure_aliases = {'pa': ['Pascal', 'pascal', 'Pa'],
+                    'hpa': ['hecto pascal', 'hPa'],
+                    'psi': ['Psi'],
+                    'bar': ['Bar']}
+
+precip_aliases = {'mm/m²': ['mm', 'liter', 'liters', 'l/m²', "milimeter"]}
+
+wind_aliases = {'m/s': ['meters/second', 'm/sec'],
+                'km/h': ['kilometers/hour', 'kph'],
+                'mph': ['miles/hour']}
+direction_aliases = {'° from north (CW)': ['°', 'degrees']}
 
 
 # conversion between standard-NAMES and aliases
 all_units_aliases = {
-    "temp": {
-        'Celcius': ['celcius', '°C', '°c'],
-        'Kelvin': ['K', 'kelvin'],
-        'Farenheit': []
-        },
-    "radiation_temp": {
-        'Celcius': ['celcius', '°C', '°c'],
-        'Kelvin': ['K', 'kelvin'],
-        'Farenheit': []
-        },
+    "temp": temp_aliases,
+    "radiation_temp": temp_aliases,
     "humidity": {
         "%": ['percent', 'percentage']
         },
-    'pressure': {
-        'pa': ['Pascal', 'pascal', 'Pa'],
-        'hpa': ['hecto pascal', 'hPa'],
-        'psi': ['Psi'],
-        'bar': ['Bar']
-        }
+    'pressure': pressure_aliases,
+    'pressure_at_sea_level': pressure_aliases,
+    'precip': precip_aliases,
+    'precip_sum': precip_aliases,
+    'wind_speed': wind_aliases,
+    'wind_gust': wind_aliases,
+    'wind_direction' : direction_aliases,
     }
 
-
+# =============================================================================
+# Unit conversion expressions
+# =============================================================================
 
 all_conversion_table = {
     'temp': {
@@ -70,13 +82,22 @@ all_conversion_table = {
     'pressure': {
         'hpa': ["x * 100"],
         'psi': ['x * 6894.7573'],
-        'bar': ['x * 100000.']
-    }
+        'bar': ['x * 100000.']},
+    'pressure_at_sea_level' : {
+        'hpa': ["x * 100"],
+        'psi': ['x * 6894.7573'],
+        'bar': ['x * 100000.']},
+    'precip': {},
+    'precip_sum': {},
+    'wind_speed': {
+        "km/h": ["x / 3.6"],
+        "mph": ["x * 0.44704"]},
+    'wind_gust' : {
+        "km/h": ["x / 3.6"],
+        "mph": ["x * 0.44704"]},
+    'wind_direction' : {}
 
 }
-
-
-
 
 
 
[…]
