### 2023-09-07 590632d
fixed implementation of obstype class and test are succesfull
