### 2023-09-08 #5783 @kthyng
@vergauwenthomas Actually those are all editors here at JOSS. Please take a look for reviewers here: https://reviewers.joss.theoj.org/lookup (some editors are probably listed there too).
