### 2023-09-13 32c807a
add test for new observation and new units
