### 2023-09-14 9a4a83c
obstype framework now extended to modeldata
