+++ metobs_toolkit/modeldata.py
@@ -22,10 +22,10 @@ from metobs_toolkit.landcover_functions import (connect_to_gee,
 from metobs_toolkit.plotting_functions import (model_timeseries_plot,
                                                timeseries_plot)
 
-# from metobs_toolkit.convertors import (convert_to_toolkit_units,
-#                                        standard_tlk_units)
-from metobs_toolkit.obstypes import tlk_obstypes
+# from metobs_toolkit.obstypes import tlk_obstypes
 from metobs_toolkit.obstypes import Obstype as Obstype_class
+from metobs_toolkit.obstype_modeldata import model_obstypes, create_model_obstype
+
 from metobs_toolkit.settings import Settings
 
 logger = logging.getLogger(__name__)
@@ -50,7 +50,7 @@ class Modeldata:
         self._df_units = {}  # the units of the data stored in the df
         self.df_tz = 'UTC'  # the timezone of the datetimes stored in the df
 
-        self.obstypes = tlk_obstypes  # Dict name: Obstype-instance
+        self.obstypes = model_obstypes  # Dict name: Obstype-instance
 
         self._is_alaro25 = False
 
@@ -74,30 +74,25 @@ class Modeldata:
         """Print overview information of the modeldata."""
         return self.__str__()
 
-    def append_obstypes(self, other_obstypes):
-        """Update the known observation types.
 
-        This method updates the obstypes stored as attribute in the Modeldata,
-        with other_obstypes.
+    def get_info(self):
 
-        Parameters
-        ----------
-        other_obstypes : {obstype_name, metobs_toolkit.Obstype}
-            Name and Obstype (class instance) pairs to update the current
-            obstypes with.
+        print(str(self))
 
-        Returns
-        -------
-        None.
+        print("\n ------ Known gee datasets -----------")
+        self.list_gee_datasets()
 
-        """
-        self.obstypes.update(other_obstypes)
-
-    def add_obstype_definition(self, Obstype):
-        """Add a new Observation type to the known observation types.
+    def add_obstype(self, bandname, band_units, Obstype=None, band_description=None,
+                               obsname=None, standard_units=None, obstype_description=None,
+                               unit_alias_dict={}, unit_conv_dict={}):
+        """Add a new Observation type to the known observation types for the
+        current dataset.
 
         Parameters
         ----------
+        bandname : str
+            The name of the band that represents the obstype.
+        band_units:
         Obstype : metobs_toolkit.obstype.Obstype
             The new Obstype to add.
 
@@ -106,21 +101,42 @@ class Modeldata:
         None.
 
         """
-        # check if Obstype is of the proper type
-        if not isinstance(Obstype, Obstype_class):
-            sys.exit(f"{Obstype} is not an instance of metobs_toolkit.obstypes.Obstype.")
 
-        # check if the Obstype is already present
-        if Obstype.name in self.obstypes.keys():
-            logger.warning(f"The {Obstype} is already present in the known obstypes.")
-            return
+        # Get a regular obstype
+        if Obstype is None:
+            obs = Obstype_class(obsname=obsname,
+                                    std_unit=standard_units,
+                                    description=obstype_description,
+                                    unit_aliases=unit_alias_dict,
+                                    unit_conversions=unit_conv_dict)
+
+
+
+        else:
+            if not isinstance(Obstype, Obstype_class):
+                sys.exit(f"{Obstype} is not an instance of metobs_toolkit.obstypes.Obstype.")
+
+            obs = Obstype
+
+        # Test if the band unit is a knonw unit
+        if not obs.test_if_unit_is_known(band_units):
+            sys.exit(f"The {bandname} unit: {band_units} is not a knonw unit for {obs.name}")
+
+
+        # Make the modeldata extension
+        equiv_dict = {self.modelname: {'name': str(bandname),
[…]
