### 2023-09-14 #5783 @vergauwenthomas
> @vergauwenthomas Actually those are all editors here at JOSS. Please take a look for reviewers here: https://reviewers.joss.theoj.org/lookup (some editors are probably listed there too).

Here I list 5 potential reviewers, based on topic areas and code language:

- Kanishkan91
- Zeitsperre
- malmans2
- nehakn
- eviatarbach

If the editor (@hugoledoux ) wants to assign other reviewers he is free to do so.
