### 2023-09-15 c49c254
documentation and addition of modelobstype and modelobstype_vectorfields + multi bandname extraction of gee
