+++ metobs_toolkit/obstypes.py
@@ -14,7 +14,6 @@ import numpy as np
 logger = logging.getLogger(__name__)
 
 
-
 # =============================================================================
 # Standard toolkit units for each observation type
 # =============================================================================
@@ -30,20 +29,6 @@ tlk_std_units = {
     "wind_direction": '° from north (CW)',
     "pressure": 'pa',
     "pressure_at_sea_level": 'pa'}
-
-# =============================================================================
-# Standard modeldata equivalents
-# =============================================================================
-tlk_std_modeldata_obstypes = {
-    'temp': { "ERA5_hourly": {'name': 'temperature_2m', 'units': 'Kelvin',
-                             'band_desc': "Temperature of air at 2m above the surface of land, sea or in-land waters. 2m temperature is calculated by interpolating between the lowest model level and the Earth's surface, taking account of the atmospheric conditions."}},
-    'dewpoint_temp': { "ERA5_hourly":   {'name': 'dewpoint_temperature_2m', "units": 'K',
-                                        'band_desc': "Temperature to which the air, at 2 meters above the surface of the Earth, would have to be cooled for saturation to occur. It is a measure of the humidity of the air. Combined with temperature and pressure, it can be used to calculate the relative humidity. 2m dew point temperature is calculated by interpolating between the lowest model level and the Earth's surface, taking account of the atmospheric conditions."}},
-    }
-
-
-
-
 
 
 # =============================================================================
@@ -72,15 +57,15 @@ all_units_aliases = {
     "radiation_temp": temp_aliases,
     "humidity": {
         "%": ['percent', 'percentage']
-        },
+    },
     'pressure': pressure_aliases,
     'pressure_at_sea_level': pressure_aliases,
     'precip': precip_aliases,
     'precip_sum': precip_aliases,
     'wind_speed': wind_aliases,
     'wind_gust': wind_aliases,
-    'wind_direction' : direction_aliases,
-    }
+    'wind_direction': direction_aliases,
+}
 
 # =============================================================================
 # Unit conversion expressions
@@ -88,17 +73,17 @@ all_units_aliases = {
 
 all_conversion_table = {
     'temp': {
-        'Kelvin': ["x - 273.15"], #result is in tlk_std_units
-        'Farenheit' : ["x-32.0", "x/1.8"]}, # -->execute from left to write  = (x-32)/1.8
+        'Kelvin': ["x - 273.15"],  # result is in tlk_std_units
+        'Farenheit': ["x-32.0", "x/1.8"]},  # -->execute from left to write  = (x-32)/1.8
     'radiation_temp': {
-        'Kelvin': ["x - 273.15"], #result is in tlk_std_units
-        'Farenheit' : ["x-32.0", "x/1.8"]},
+        'Kelvin': ["x - 273.15"],  # result is in tlk_std_units
+        'Farenheit': ["x-32.0", "x/1.8"]},
     'humidity': {},
     'pressure': {
         'hpa': ["x * 100"],
         'psi': ['x * 6894.7573'],
         'bar': ['x * 100000.']},
-    'pressure_at_sea_level' : {
+    'pressure_at_sea_level': {
         'hpa': ["x * 100"],
         'psi': ['x * 6894.7573'],
         'bar': ['x * 100000.']},
@@ -107,26 +92,24 @@ all_conversion_table = {
     'wind_speed': {
         "km/h": ["x / 3.6"],
         "mph": ["x * 0.44704"]},
-    'wind_gust' : {
+    'wind_gust': {
         "km/h": ["x / 3.6"],
         "mph": ["x * 0.44704"]},
-    'wind_direction' : {}
+    'wind_direction': {}
 
 }
-
-
-
 
 # =============================================================================
 # Observation type class
 # =============================================================================
+
 
[…]
