+++ metobs_toolkit/landcover_functions.py
@@ -227,10 +227,10 @@ def coordinates_available(metadf, latcol="lat", loncol="lon"):
     return True
 
 
-def _estimate_data_size(metadf, startdt, enddt, mapinfo):
-    datatimerange = pd.date_range(start=startdt, end=enddt, freq=mapinfo["time_res"])
+def _estimate_data_size(metadf, startdt, enddt, time_res, n_bands=1):
+    datatimerange = pd.date_range(start=startdt, end=enddt, freq=time_res)
 
-    return metadf.shape[0] * len(datatimerange)
+    return metadf.shape[0] * len(datatimerange) * n_bands
 
 
 # =============================================================================
@@ -411,7 +411,7 @@ def extract_buffer_frequencies(metadf, mapinfo, bufferradius):
 
 
 def gee_extract_timeseries(
-    metadf, bandname, mapinfo, startdt, enddt, obstype="temp", latcolname="lat", loncolname="lon"
+    metadf, band_mapper, mapinfo, startdt, enddt, latcolname="lat", loncolname="lon"
 ):
     """Extract timeseries data at the stations location from a GEE dataset.
 
@@ -427,16 +427,15 @@ def gee_extract_timeseries(
     ----------
     metadf : pd.DataFrame
         dataframe containing coordinates and a column "name", representing the name for each location.
-    bandname : str
-        the name of the band to extract data from.
+    band_mapper : dict
+        the name of the band to extract data from as keys, the default name of
+        the corresponding obstype as values.
     mapinfo : Dict
         The information about the GEE dataset.
     startdt : datetime obj
         Start datetime for timeseries (included).
     enddt : datetime obj
         End datetime for timeseries (excluded).
-    obstype : String, optional
-        toolkit observation type. The default is 'temp'.
     latcolname : String, optional
         Columnname of latitude values. The default is 'lat'.
     loncolname : String, optional
@@ -446,18 +445,23 @@ def gee_extract_timeseries(
     -------
     pd.DataFrame
         A dataframe with name - datetime multiindex, all columns from the metadf + extracted timeseries
-        column with the same name as the obstype.
+        column with the same name as the obstypes.
 
     """
     scale = mapinfo["scale"]
-    # bandname = mapinfo["band_of_use"][obstype]["name"]
+    bandnames = list(band_mapper.keys())
 
     # test if coordiantes are available
     if not coordinates_available(metadf, latcolname, loncolname):
         return pd.DataFrame()
 
     use_drive = False
-    _est_data_size = _estimate_data_size(metadf, startdt, enddt, mapinfo)
+    _est_data_size = _estimate_data_size(metadf=metadf,
+                                         startdt=startdt,
+                                         enddt=enddt,
+                                         time_res = mapinfo["time_res"],
+                                         n_bands=len(bandnames)
+                                         )
     if _est_data_size > 4000:
         print(
             "THE DATA AMOUT IS TO LAREGE FOR INTERACTIVE SESSION, THE DATA WILL BE EXPORTED TO YOUR GOOGLE DRIVE!"
@@ -488,7 +492,7 @@ def gee_extract_timeseries(
     enddt = enddt + pd.Timedelta(mapinfo['time_res'])
 
 
-    raster = get_ee_obj(mapinfo, bandname)  # dataset
+    raster = get_ee_obj(mapinfo, bandnames)  # dataset
     results = (
         raster.filter(
             ee.Filter.date(
@@ -500,7 +504,7 @@ def gee_extract_timeseries(
         .flatten()
     )
 
-    def format_df(df, obstype, bandname):
+    def format_df(df, band_mapper):
         # format datetime
         df["datetime"] = pd.to_datetime(df["datetime"], format="%Y%m%d%H%M%S")
[…]

[…]
