### 2023-09-22 ca6d2f3
minor code review fixes
