+++ metobs_toolkit/modeldata.py
@@ -81,9 +81,7 @@ class Modeldata:
         print("\n ------ Known gee datasets -----------")
         self.list_gee_datasets()
 
-    def add_obstype(self, bandname, band_units, Obstype=None, band_description=None,
-                    obsname=None, standard_units=None, obstype_description=None,
-                    unit_alias_dict={}, unit_conv_dict={}):
+    def add_obstype(self, Obstype, bandname, band_units, band_description=None):
         """Add a new Observation type for the current Modeldata.
 
         Parameters
@@ -99,19 +97,10 @@ class Modeldata:
         None.
 
         """
-        # Get a regular obstype
-        if Obstype is None:
-            obs = Obstype_class(obsname=obsname,
-                                std_unit=standard_units,
-                                description=obstype_description,
-                                unit_aliases=unit_alias_dict,
-                                unit_conversions=unit_conv_dict)
+        if not isinstance(Obstype, Obstype_class):
+            sys.exit(f"{Obstype} is not an instance of metobs_toolkit.obstypes.Obstype.")
 
-        else:
-            if not isinstance(Obstype, Obstype_class):
-                sys.exit(f"{Obstype} is not an instance of metobs_toolkit.obstypes.Obstype.")
-
-            obs = Obstype
+        obs = Obstype
 
         # Test if the band unit is a knonw unit
         if not obs.test_if_unit_is_known(band_units):
@@ -140,7 +129,7 @@ class Modeldata:
         gee_location : str
             Location of the gee dataset (like "ECMWF/ERA5_LAND/HOURLY" for ERA5).
         obstype : str
-            The observation type the band corresponds to.
+            The observation type name the band corresponds to.
         bandname : str
             Name of the dataset band as stored on the GEE.
         units : str
@@ -190,11 +179,6 @@ class Modeldata:
             mapname: {
                 'location': f'{gee_location}',
                 'usage': 'user defined addition',
-                # 'band_of_use':
-                #     {f'{obstype}':
-                #           {'name': f'{bandname}',
-                #           'units': f'{units}'}
-                #       },
                 'value_type': val_typ,
                 'dynamical': not bool(is_image),
                 'scale': int(scale),
@@ -278,19 +262,10 @@ class Modeldata:
         ----------
         obstype : str
             Observation type to convert to standard units.
-        target_unit_name : str, optional
-            Target unit name to convert to. The default is 'Celsius'.
-        conv_expr : str, optional
-            If the target_unit_name is not a default, you can add the
-            conversion expression here (i.g. "x - 273.15"). The default is None.
 
         Returns
         -------
         None.
-
-        Note
-        -------
-        If the observationtype is not a default, you must first make this metobs_toolkit.Obstype and add it to the
 
         """
         # chech if data is available
@@ -312,9 +287,26 @@ class Modeldata:
         self.df[obstype] = converted_data
         logger.info(f'{obstype} are converted from {cur_unit} --> {self.obstypes[obstype].get_standard_unit()}.')
 
-    def expoid_vector_field(self, obstype):
+    def exploid_2d_vector_field(self, obstype):
+        """Compute amplitude and direction of 2D vector field components.
+
+        The amplitude and directions are added to the data attribute, and their
+        equivalent observationtypes are added to the known ModelObstypes.
+
+        (The vector components are not saved.)
+        Parameters
+        ----------
+        obstype : str
+            The name of the observationtype that is a ModelObstype_Vectorfield.
+
+        Returns
[…]
