+++ metobs_toolkit/data_templates/template_build_prompt.py
@@ -10,11 +10,83 @@ import os
 import sys
 import pandas as pd
 import numpy as np
+import copy
 from datetime import datetime
 import pytz
 
-from metobs_toolkit.obstypes import tlk_obstypes
+from metobs_toolkit.obstypes import Obstype, tlk_obstypes
 from metobs_toolkit.data_import import _read_csv_to_df
+
+
+def add_new_obstype():
+
+    print('\n --- Adding a new observation type --- \n')
+
+    # get obsname
+    name_ok = False
+    while not name_ok:
+        obsname = str(input("Give the name of your observation type: "))
+        if obsname in tlk_obstypes.keys():
+            print(f"!! {obsname} is already a knonw observation type. This cannot be added.")
+        else:
+            name_ok = True
+
+    # get std unit
+    std_unit = str(input('Give the standard unit (how the toolkit should store/present the data): '))
+
+    # Get input data unit
+    is_std_unit = yes_no_ques(f' Are the {obsname} values in your data in {std_unit}')
+    if is_std_unit:
+        cur_unit = std_unit
+        unit_conv={}
+    else:
+        cur_unit = str(input('Give the unit your data is in: '))
+        print(f'Give the expression on how to convert {cur_unit} values to {std_unit}. ')
+        print('  * Example: Kelvin (= new unit) to °C :  x - 273.15 ')
+        print('  * Example: Farenheit to °C : x-32.0; x/1.8    (executed left to right)')
+
+        conv_str = str(input(' : '))
+        # cleanup and make list if needend
+        conv_str = list(conv_str.replace(' ', '').split(';'))
+
+        unit_conv = {cur_unit: conv_str}
+    # Description
+    description = str(input(f'Give a detailed description of the {obsname} type (optional): '))
+
+    #Aliases and coversions
+
+    # Do not add this in the prompt, the prompt should not check the more advanced
+    # settigns. If the prompt could cover 95% of all user needs, that would be great.
+    # The others should help themself with the documentation to create aliases
+    #and conversions
+
+    unit_aliases = {}
+
+    # create obstype:
+    new_obstype = Obstype(obsname=obsname, std_unit=std_unit,
+                          description=description, unit_aliases=unit_aliases,
+                          unit_conversions = unit_conv)
+    return new_obstype, cur_unit
+
+def get_unit(obstype):
+
+    available_units = obstype.get_all_units()
+    available_units.append('ADD A NEW UNIT')
+
+    print(f'\n Select the unit your {obstype.name} data is in:  \n')
+    conv_str = None
+    unit = col_option_input(available_units)
+    if unit == 'ADD A NEW UNIT':
+        unit = str(input('Give the unit your data is in: '))
+        print(f'Give the expression on how to convert {unit} values to {obstype.get_standard_unit()}. ')
+        print('  * Example: Kelvin (= new unit) to °C :  x - 273.15 ')
+        print('  * Example: Farenheit to °C : x-32.0; x/1.8    (executed left to right)')
+
+        conv_str = str(input(' : '))
+        # cleanup and make list if needend
+        conv_str = list(conv_str.replace(' ', '').split(';'))
+
+    return unit, conv_str
 
 
 def col_option_input(columns):
@@ -107,6 +179,11 @@ def build_template_prompt(debug=False):
     """Launch the prompt to help make a template."""
     template_dict = {}
     options_dict = {}
+
+    known_obstypes =  copy.copy(tlk_obstypes)
+    new_units = {}
+
+
     print('This prompt will help to build a template for your data and metadata. Answer the prompt and hit Enter. \n \n')
 
     print(' *******      File locations   *********** \n')
@@ -183,18 +260,23 @@ def build_template_prompt(debug=False):
[…]
