### 2023-09-22 c959597
bug fixes in template prompt helper
