### 2023-09-25 f5986b8
better naming of variable in the documentation examples
