### 2023-09-25 26a44b3
fixed current functions for vector observation types
