### 2023-09-25 dfd2a54
extend the modelobstype classes with required methods
