+++ metobs_toolkit/obstype_modeldata.py
@@ -6,6 +6,7 @@ Created on Wed Sep 13 13:43:15 2023
 @author: thoverga
 """
 import sys
+import copy
 import math
 import numpy as np
 import logging
@@ -87,6 +88,14 @@ class ModelObstype(Obstype):
 
         self.modl_equi_dict = model_equivalent_dict
         self._is_valid()
+
+    def __repr__(self):
+        """Instance representation."""
+        return f"ModelObstype instance of {self.name}"
+
+    def __str__(self):
+        """Text representation."""
+        return f"ModelObstype instance of {self.name}"
 
     def get_info(self):
         """Print out detailed information of the observation type.
@@ -199,6 +208,14 @@ class ModelObstype_Vectorfield(Obstype):
         self.modl_comp_dict = mod_comp_dict
         self._is_valid()
 
+    def __repr__(self):
+        """Instance representation."""
+        return f"ModelObstype_Vectorfield instance of {self.name}"
+
+    def __str__(self):
+        """Text representation."""
+        return f"ModelObstype_Vectorfield instance of {self.name}"
+
     def get_info(self):
         """Print out detailed information of the observation type.
 
@@ -252,92 +269,6 @@ class ModelObstype_Vectorfield(Obstype):
         return f'u_comp_{self.name}'
     def get_v_column(self):
         return f'v_comp_{self.name}'
-
-    def compute_amplitude(self, df):
-        # Compute the data
-        data =((df[self.get_u_column()].pow(2)) + (df[self.get_v_column()].pow(2))).pow(1./2)
-        # Create a new obstype for the amplitude
-        amplitude_obstype = Obstype(obsname = f'{self.name}_amplitude',
-                                    std_unit = self.std_unit,
-                                    description=f'2D-vector amplitde of {self.name} components.',
-                                    unit_aliases = self.units_aliases,
-                                    unit_conversions = self.conv_table)
-        # convert to model obstype
-        mod_equi = {}
-        for key, val in self.modl_comp_dict.items():
-            mod_equi[key] = val['u_comp']
-            mod_equi[key]['name'] = f"{val['u_comp']['name']} and {val['v_comp']['name']}"
-
-        amplitude_obstype = ModelObstype(amplitude_obstype,
-                                         model_equivalent_dict=mod_equi
-                                         )
-        # amplitude_obstype.plotlabel=f"Amplitude of 2D-{self.name} field ({self.std_unit})"
-        return data, amplitude_obstype
-
-    def compute_angle(self, df):
-
-        def unit_vector(vector):
-            """ Returns the unit vector of the vector.  """
-            return vector / np.linalg.norm(vector)
-
-        def angle_between(u_comp, v_comp):
-            """ Returns the angle in radians between vectors 'v1' and 'v2'::
-
-                    >>> angle_between((1, 0, 0), (0, 1, 0))
-                    1.5707963267948966
-                    >>> angle_between((1, 0, 0), (1, 0, 0))
-                    0.0
-                    >>> angle_between((1, 0, 0), (-1, 0, 0))
-                    3.141592653589793
-            """
-
-            v2 = (u_comp, v_comp)
-            v1_u = unit_vector((0,1)) #North unit arrow
-            v2_u = unit_vector(v2)
-
-            angle_rad = np.arccos(np.clip(np.dot(v1_u, v2_u), -1.0, 1.0))
-            angle_degrees = angle_rad * ((180.0/math.pi))
-            # return angle_degrees
-            # fix the quadrants
-            if (v2[0] >= 0) & (v2[1] >= 0):
-                # N-E quadrant
-                return angle_degrees
-            if (v2[0] >= 0) & (v2[1] < 0):
-                # S-E quadrant
-                return angle_degrees
-            if (v2[0] < 0) & (v2[1] < 0):
[…]
