### 2023-09-25 56e66ed
Update test with new observation types etc
