+++ metobs_toolkit/__init__.py
@@ -70,5 +70,5 @@ from metobs_toolkit.dataset_settings_updater import Dataset
 # =============================================================================
 
 # DO not change this manually!
-__version__ = "0.1.1a2"
+__version__ = "0.1.3a2"
 
