### 2023-09-26 553637b
fix interpolation of and import from csv for multi obs for modeldata
