+++ requirements.txt
@@ -256,12 +256,12 @@ geocoder==1.38.1 ; python_version >= "3.9" and python_version < "4.0" \
 geopandas==0.9.0 ; python_version >= "3.9" and python_version < "4.0" \
     --hash=sha256:63972ab4dc44c4029f340600dcb83264eb8132dd22b104da0b654bef7f42630a \
     --hash=sha256:79f6e557ba0dba76eec44f8351b1c6b42a17c38f5f08fef347e98fe4dae563c7
-google-api-core==2.11.1 ; python_version >= "3.9" and python_version < "4.0" \
-    --hash=sha256:25d29e05a0058ed5f19c61c0a78b1b53adea4d9364b464d014fbda941f6d1c9a \
-    --hash=sha256:d92a5a92dc36dd4f4b9ee4e55528a90e432b059f93aee6ad857f9de8cc7ae94a
-google-api-python-client==2.100.0 ; python_version >= "3.9" and python_version < "4.0" \
-    --hash=sha256:226ca35355993d6182506c51745ab5149405cdf6a92975b2725ab3e0d757dbe9 \
-    --hash=sha256:eaed50efc2f8a4027dcca8fd0037f4b1b03b8093efc84ce3cb6c75bfc79a7e31
+google-api-core==2.12.0 ; python_version >= "3.9" and python_version < "4.0" \
+    --hash=sha256:c22e01b1e3c4dcd90998494879612c38d0a3411d1f7b679eb89e2abe3ce1f553 \
+    --hash=sha256:ec6054f7d64ad13b41e43d96f735acbd763b0f3b695dabaa2d579673f6a6e160
+google-api-python-client==2.101.0 ; python_version >= "3.9" and python_version < "4.0" \
+    --hash=sha256:71760dcf11d191b65520d1c13757a776f4f43cf87f302097a0d8e491c2ef87b0 \
+    --hash=sha256:e9620a809251174818e1fce16604006f10a9c2ac0d3d94a139cdddcd4dbea2d8
 google-auth-httplib2==0.1.1 ; python_version >= "3.9" and python_version < "4.0" \
     --hash=sha256:42c50900b8e4dcdf8222364d1f0efe32b8421fb6ed72f2613f12f75cc933478c \
     --hash=sha256:c64bc555fdc6dd788ea62ecf7bccffcf497bf77244887a3f3d7a5a02f8e3fc29
