### 2023-09-26 b3d6b29
start of the obervation type module (#377)
* start of the obervation type module

* framework for the obstypeclass in the Dataset

* fixed implementation of obstype class and test are succesfull

* add test for new observation and new units

* obstype framework now extended to modeldata

* documentation and addition of modelobstype and modelobstype_vectorfields + multi bandname extraction of gee

* minor code review fixes

* fixes of codereview

* update the template prompter to create new obstypes/units

* bug fixes in template prompt helper

* new build

* better naming of variable in the do
[…]
