+++ tests/push_test/modeldata_test.py
@@ -35,27 +35,65 @@ dataset.coarsen_time_resolution()
 
 #%% test adding gee information
 model_data = metobs_toolkit.Modeldata("ERA5_hourly")
-model_data.add_band_to_gee_dataset(bandname='surface_pressure',
-                                   obstype='pressure',
-                                   units='pa')
 
-model_data.add_gee_dataset(mapname='new dataset name',
-                            gee_location='location/loc/dfmijfe',
-                            obstype='temp',
-                            bandname='temp 2m passive field',
-                            units ='C',
-                            scale = 100,
-                            time_res='1H',
-                            is_image=False,
-                            is_numeric=True,
-                            credentials='bladiblie')
+# Define a regular obstype
+new_obstype = metobs_toolkit.Obstype(obsname='special_pressure',
+                                 std_unit='pa',
+                                 description='just for testing',
+                                 unit_aliases={'pa': ['Pascal', 'Pa', 'N/m²'],
+                                               },
+                                 unit_conversions={'hpa': ["x * 100"]},
+                                 )
+
+# add new obstype to model_data
+model_data.add_obstype(Obstype=new_obstype,
+                                  bandname='surface_pressure',
+                                  band_units='hpa',
+                                  )
+
+
+model_data.get_info()
+from datetime import datetime
+tstart = datetime(2022, 10, 3,23)
+tend = datetime(2022, 10,4, 4)
+model_data = dataset.get_modeldata(modeldata = model_data, obstype = 'special_pressure',startdt = tstart, enddt = tend)
+
+assert model_data.df.shape[0] == 168, 'No modeldata extracted from gee for new unit and obstype!'
+assert model_data.df.columns.to_list() ==  ['special_pressure'], 'Something is wrong with column names'
+
+model_data.make_plot(obstype_model='special_pressure')
+#%% Test 2D vector fields
+
+model_data = dataset.get_modeldata(modeldata = model_data,
+                                   obstype = 'wind',
+                                   startdt = tstart,
+                                   enddt = tend)
+
+print(model_data)
+
+assert model_data.df.columns.to_list() ==  ['wind_amplitude', 'wind_direction'], 'Something is wrong with column names'
+
+
+#%% Testing multiple field extraction
+model_data.get_gee_dataset_data(mapname = model_data.modelname,
+                                metadf = dataset.metadf,
+                                   obstype = ['temp', 'wind'],
+                                   startdt_utc = tstart,
+                                   enddt_utc = tend)
+
+assert model_data.df.columns.to_list() ==  ['temp', 'wind_amplitude', 'wind_direction'], 'Something is wrong with column names'
+
+
+
 #%% Import modeldata
 model_data = metobs_toolkit.Modeldata("ERA5_hourly")
-
+ #mutliple observations and vector components
 csv_file = os.path.join(lib_folder, 'tests', 'test_data', 'era5_modeldata_test.csv')
 
 model_data.set_model_from_csv(csv_file)
 
+assert model_data.df.columns.to_list() == ['temp', 'wind_amplitude', 'wind_direction'], 'something wrong with reading modeldata from csv (drive).'
+model_data.make_plot(obstype_model='wind_amplitude')
 #%% Test repr
 
 print(model_data)
@@ -77,9 +115,28 @@ if os.path.exists(fullpath):
 
 
 #%% test interpolation
+
+dataset = metobs_toolkit.Dataset()
+dataset.update_settings(input_data_file=metobs_toolkit.demo_datafile,
[…]
