+++ requirements.txt
@@ -169,14 +169,14 @@ contourpy==1.1.1 ; python_version >= "3.9" and python_version < "4.0" \
     --hash=sha256:f44d78b61740e4e8c71db1cf1fd56d9050a4747681c59ec1094750a658ceb970 \
     --hash=sha256:f6aec19457617ef468ff091669cca01fa7ea557b12b59a7908b9474bb9674cf0 \
     --hash=sha256:f9dc7f933975367251c1b34da882c4f0e0b2e24bb35dc906d2f598a40b72bfc7
-cycler==0.11.0 ; python_version >= "3.9" and python_version < "4.0" \
-    --hash=sha256:3a27e95f763a428a739d2add979fa7494c912a32c17c4c38c4d5f082cad165a3 \
-    --hash=sha256:9c87405839a19696e837b3b818fed3f5f69f16f1eec1a1ad77e043dcea9c772f
+cycler==0.12.0 ; python_version >= "3.9" and python_version < "4.0" \
+    --hash=sha256:7896994252d006771357777d0251f3e34d266f4fa5f2c572247a80ab01440947 \
+    --hash=sha256:8cc3a7b4861f91b1095157f9916f748549a617046e67eb7619abed9b34d2c94a
 decorator==5.1.1 ; python_version >= "3.9" and python_version < "4.0" \
     --hash=sha256:637996211036b6385ef91435e4fae22989472f9d571faba8927ba8253acbc330 \
     --hash=sha256:b8c3f85900b9dc423225913c5aace94729fe1fa9763b38939a95226f02d37186
-earthengine-api==0.1.370 ; python_version >= "3.9" and python_version < "4.0" \
-    --hash=sha256:ca35f9c9f562d5509c19a94a172c577cf5f123ab63272b5a0929021dbb2dd5b3
+earthengine-api==0.1.371 ; python_version >= "3.9" and python_version < "4.0" \
+    --hash=sha256:06bc9f302a5862dfe1fec6e65b58ca633981259a2180a63f6b104a42c2f08333
 eerepr==0.0.4 ; python_version >= "3.9" and python_version < "4.0" \
     --hash=sha256:79595fd643e99d91a68041cdba3f5347d5559e6744a34983291a46cada6edbd4 \
     --hash=sha256:920cccbe498b2880f42ed39a2db8945f705b1497523c48b8d392c5380d511926
@@ -210,41 +210,49 @@ fiona==1.9.4.post1 ; python_version >= "3.9" and python_version < "4.0" \
 folium==0.14.0 ; python_version >= "3.9" and python_version < "4.0" \
     --hash=sha256:8ec44697d18f5932e0fdaee8b19da98625de4d0e72cb30ef56f9479f18e11b9f \
     --hash=sha256:c894e2c029a8ca40e043a311978a895cefe32d746a94263f807dd7b6b2e9c679
-fonttools==4.42.1 ; python_version >= "3.9" and python_version < "4.0" \
-    --hash=sha256:0eb79a2da5eb6457a6f8ab904838454accc7d4cccdaff1fd2bd3a0679ea33d64 \
-    --hash=sha256:113337c2d29665839b7d90b39f99b3cac731f72a0eda9306165a305c7c31d341 \
-    --hash=sha256:12a7c247d1b946829bfa2f331107a629ea77dc5391dfd34fdcd78efa61f354ca \
-    --hash=sha256:179737095eb98332a2744e8f12037b2977f22948cf23ff96656928923ddf560a \
[…]
