### 2023-09-29 a9fc010
add interactive and time evolution folium plot (#381)
* add interactive and time evolution folium plot

* cleanup
