### 2023-10-02 #5916 @editorialbot
[REVIEW]: MetObs - a Python toolkit for using non-traditional meteorological observations
**Submitting author:** @vergauwenthomas (<a href="http://orcid.org/0000-0003-2899-9218">Thomas Vergauwen</a>)
**Repository:** https://github.com/vergauwenthomas/MetObs_toolkit
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.2.0
