### 2023-10-07 vergauwenthomas/MetObs_toolkit#382 @ashwinvis
Copyright holder possibly incorrect
The text is a valid open source license, but the holder... it must be an error, isn't it?

> Copyright (c) 2018 The Python Packaging Authority

cf. https://github.com/openjournals/joss-reviews/issues/5916
