### 2023-10-09 287d974
Logo in documentation + small doc fixes
