+++ metobs_toolkit/qc_checks.py
@@ -804,7 +804,7 @@ def toolkit_buddy_check(obsdf, metadf, obstype, buddy_radius, min_sample_size, m
     metadf: Pandas.DataFrame
         The dataframe containing the metadata (e.g. latitude, longitude...)
      obstype: String, optional
-         The observation type that has to be checked. The default is 'temp'
+        The observation type that has to be checked. The default is 'temp'
     buddy_radius : numeric
         The radius to define neighbours in meters.
     min_sample_size : int
