### 2023-10-10 380f946
bugfix when creating a new obstype
