### 2023-10-10 4de9618
fix Copyright holder (see #382)
