### 2023-10-11 f5abf98
fix Copyright holder (see #382) (#383)
