### 2023-10-11 vergauwenthomas/MetObs_toolkit#382 @vergauwenthomas
> The text is a valid open source license, but the holder... it must be an error, isn't it?
>
> > Copyright (c) 2018 The Python Packaging Authority
>
> cf. [openjournals/joss-reviews#5916](https://github.com/openjournals/joss-reviews/issues/5916)

@ashwinvis
Yes that was indeed an error, thank you for pointing it out!

(I made a pull request and merged it into the master. If the pull request does not affect the installation/package itself, then i will add them to the master without specifying a new release. Do you think this is a good strategy? I am pretty new to DevOps so i am unfamiliar with the deployment conversions.)
