### 2023-10-14 vergauwenthomas/MetObs_toolkit#382 @ashwinvis
Great! Yes, no release is necessary when it comes to such things. However for JOSS the code will be archived towards the end, and if a release is indeed needed, because of bugfixes or changes to the documentation,  it can wait until the end of the review.
