### 2023-10-28 vergauwenthomas/MetObs_toolkit#384 @ashwinvis
Unclear instructions in Examples page in the documentation
I would like to get started, but I don't see where I should obtain the data from. Also the code snippet is truncated in the beginning. So some variables are undefined.

https://vergauwenthomas.github.io/MetObs_toolkit/examples.html

cf: https://github.com/openjournals/joss-reviews/issues/5916
