### 2023-10-28 vergauwenthomas/MetObs_toolkit#385 @ashwinvis
Elaborate contributing guidelines
https://vergauwenthomas.github.io/MetObs_toolkit/MetObs_documentation_full.html

I think this is very short. According to JOSS and I agree

> There should be clear guidelines for third-parties wishing to:
>
>     Contribute to the software
>
>     Report issues or problems with the software
>
>     Seek support
>

See here for some general guidelines

https://www.pyopensci.org/python-package-guide/documentation/repository-files/contributing-file.html

cf: https://github.com/openjournals/joss-reviews/issues/5916
