### 2023-11-03 vergauwenthomas/MetObs_toolkit#385 @Zeitsperre
Hi @vergauwenthomas, I'm going to also jump in and agree that the contribution guidelines aren't clear. These should be found in the top-level of the project as well as be somewhere in the documentation (maybe near the developer documentation).

This is an example from one of the projects I develop/maintain: https://github.com/CSHS-CWRA/RavenPy/blob/master/CONTRIBUTING.rst (No need to follow this to the letter, but more for inspiration)
