### 2023-11-03 vergauwenthomas/MetObs_toolkit#386 @Zeitsperre
GitHub-related templates
This isn't a review-blocking issue, but this repo could benefit from some general templates:
 - [Issue Templates](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository)
 - [Pull Request Template](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/creating-a-pull-request-template-for-your-repository)

This can help with supporting both users and developers who wish to contribute to the project

Context: https://github.com/openjournals/joss-reviews/issues/5916
