### 2023-11-03 vergauwenthomas/MetObs_toolkit#387 @Zeitsperre
[JOSS] Paper Commentary
Context: https://github.com/openjournals/joss-reviews/issues/5916

This is a list of some recommendations that could help with the paper submission:
- [x] The Statement of Need could benefit from some/most of the information found in the first paragraph of the summary. Do the software outputs adhere to/enforce existing standards for weather station data (WMO guidelines, ISOs, CF-Conventions)? Expand on the frameworks/standards/approaches you are building upon?
- [x] "Quality control is performed in the form of a series of checks" - Elaborate? What are you checking for exactly at the time of writing (monotonic timestamps, data gaps, extremes, etc.)?
- [x] "It is to be expected that some novel functionalities will be added in the coming years and the documentation will be updated accordingly." - Unless the features are already planned for and can be described, I'd remove this line. The paper should showcase that the software is feature-complete (even if there are new features being planned; the paper doesn't prevent you from expanding the software in future versions). This statement can be easily rolled into the first line of the preceding paragraph if needed.
- [x] The citation for the Pandas team needs to be fixed: `"team, T. pandas development"`

There are a handful of phrases that could be tightened as well (if the authors like, I can suggest a few changes in a PR). I might also have some small additional comments after more reflection.
