### 2023-11-03 vergauwenthomas/MetObs_toolkit#388 @Zeitsperre
[JOSS] Documentation Suggestions
Context: https://github.com/openjournals/joss-reviews/issues/5916

These are a few suggestions for the documentation:
- [x] For the examples, consider using Jupyter Notebooks and [nbsphinx](https://nbsphinx.readthedocs.io/en) to render them. (Addressed in #390)
  - This allows for the use of both Markdown and Python code cells within the documentation.
  - Can double as another way of testing the functionality of the library (`nbsphinx_execute = True`)
  - Much easier to maintain (`literalinclude` with `:linenos:` means that anytime the codebase changes, the docs need to be tweaked; not a great use of your effort).
- [x] The user API and developer documentation could benefit immensely from call signature typing and more elaborate NumPy-docstrings (`Parameters`/`Attributes`, `Returns`, `Notes`, `Examples`, etc.). It isn't clear to me what those documented functions/class methods accept, when being called when reading the documentation, nor when playing around with them in my IDE. This isn't the case for some functions; it just needs to be more consistently applied.
- [x] For users, I'd like to know specifically which QC checks are being performed in the treatment process (maybe listed based on currently implemented features?). (Addressed in #390)
- [x] #498
- [x] For developers, I'd like to see more of the logic under the hood explained in the developer documentation. Perhaps there's a way to list those currently implemented?
  - The nesting of functions within modules means needing to drill 3 or 4 levels down to get to function explanations; This could be flattened a bit for readability.
