### 2023-11-03 vergauwenthomas/MetObs_toolkit#389 @Zeitsperre
[JOSS] Package Health
The comments here aren't necessarily blocking, but it would be good to implement some of these suggestions:
- The `pyproject.toml` file:
  - The `documentation` URL points to the documentation page of `Poetry`
  - There should be some classifiers listed to help identify the package based on its contents (see: https://pypi.org/classifiers/)
  - Don't bother trying to expand support for Python3.8 (currently end-of-life)
- `environment.yml`: It might be just my experience, but unless you have a specific implementation (e.g. a production-level docker service), you shouldn't need to pin the binaries to their hashes. I can see this being a problem for some platforms/systems down the line. Please correct me if I'm wrong!
- Packaged data is kind of enormous: The PyPI wheel/sources are around 13 MB when archive is inflated, with the majority of that space being used by a Shapefile. Is there a way of fetching a subset of that data remotely, maybe? To help bring down the space used, the wheel should only package the data needed to run (not test/demonstrate) the library.
- `.readthedocs.yml` - The `os` could be updated to `ubuntu-22.04`
- Linting
   - The repository has a disabled `black` action. Have you considered using [pre-commit](https://pre-commit.com/) or [pre-commit.ci](https://pre-commit.ci/)?
   - The `pyproject.toml` file could be used to set the linting and style configuration for the code base (using `flake8` or `ruff`). (example from a project I maintain called `xclim`: https://github.com/Ouranosinc/xclim/blob/2d981abc8291645da235c52479162953ef2ab6d6/pyproject.toml#L215)
   - The `pyproject.toml` file could be cleaned up a bit (https://github.com/macisamuele/language-formatters-pre-commit-hooks)

Context: https://github.com/openjournals/joss-reviews/issues/5916
