+++ examples/introduction_example.py
@@ -1,144 +1,109 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-Created on Fri Sep 23 12:01:35 2022
+Introduction example script
 
 @author: thoverga
 """
-import os
-from pathlib import Path
 
-main_folder = Path(__file__).resolve().parents[1]
-testdata_file = os.path.join(
-    str(main_folder), "tests", "test_data", "vlinderdata_small.csv"
-)
-metadata = os.path.join(str(main_folder), "static_data", "vlinder_metadata.csv")
-templatefile = os.path.join(
-    str(main_folder),
-    "metobs_toolkit",
-    "data_templates",
-    "template_defaults",
-    "default_template.csv",
-)
-
-import sys
-
-sys.path.append(str(main_folder))
+# This is an introduction to get started with the MetObs toolkit. Once the MetObs toolkit package is installed,
+# you can import its functionality by:
 
 import metobs_toolkit
-
 
 # =============================================================================
 # Dataset
 # =============================================================================
 
 # A dataset is a collection of all observational data. Most of the methods are
-# applied directly on a dataset. Always start by initializing a dataset object:
+# applied directly on a dataset. Start by creating an empty dataset object:
 
-aug_2020_all_vlinders = metobs_toolkit.Dataset()
+your_dataset = metobs_toolkit.Dataset()
 
-# All default observationtypes present in the MetObs toolkit can be shown by:
-print(metobs_toolkit.observation_types)
+# The most relevant attributes of a Dataset are:
+#    * .df --> a pandas DataFrame where all the observation data is stored
+#    * .metadf --> a pandas DataFrame where all the metadata for each station is stored
+#    * .settings --> a Settings object to store all specific settings.
+#    * .missing_obs and .gaps --> here the missing records and gaps are stored if present.
+
+# Note that each Dataset will be equipped with the default settings.
 
 
-# we created an dataset and stored in under the variable 'aug_2020_all_vlinder'.
+# We created a dataset and stored in under the variable 'aug_2020_all_vlinder'.
 # The show function prints out an overview of data in the dataset:
-aug_2020_all_vlinders.show() # or .get_info()
+your_dataset.show() # or .get_info()
+
+
+# TIP: to get an extensive overview of an object, call the .show() method on it.
+
 
 # =============================================================================
 # Importing data
 # =============================================================================
 
-# To import data, you must specify where your data (and additional templates) are.
-# All this information is stored in the Settings of a dataset. To get an overview
-# of all the settings:
-aug_2020_all_vlinders.show_settings()
 
-# you can see that the settings contain default values. But there are no default
-# values for where your data is. So you have to update the settings.
+# To import your data into a Dataframe, the following files are required:
+
+#    * data file: This is the csv file containing the observations
+#    * (optional) metadata file: The csv file containing metadata for each station.
+#    * template file: This is a csv file that is used to interpret your data file, and metadata file if present.
+
+# In practice you need to start by creating a template file for your data. More
+# information on the creation of the template can be found in documentation
+# (under "Mapping to the toolkit").
+
+# TIP: Use the template assistant of the toolkit for creating a template file by using:
+
+# metobs_toolkit.build_template_prompt()
 
 
-# Start by updating the settings of your dataset so it knows here the data files are:
 
+# To import data, you must specify the paths to your data, metadata and template are.
+# For this example we use the demo data, metadata and template that comes with
+# the toolkit.
 
-aug_2020_all_vlinders.update_settings(
-    input_data_file=testdata_file,  # A demo data file, downloaded with brian tool: https://vlinder.ugent.be/vlinderdata/multiple_vlinders.php
-    input_metadata_file=metadata,
-    output_folder="/home/$USER/output/",
[…]
