### 2023-11-06 74b50cc
update the lcz example
