+++ examples/introduction_example.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-Introduction example script
+Introduction example script.
 
 @author: thoverga
 """
