+++ examples/lcz_example.py
@@ -1,112 +1,108 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-Created on Fri Sep 23 12:01:35 2022
+Example script on GEE interactions.
 
 @author: thoverga
 """
-import os
-from pathlib import Path
-
-main_folder = Path(__file__).resolve().parents[1]
-testdata_file = os.path.join(
-    str(main_folder), "tests", "test_data", "vlinderdata_small.csv"
-)
-metadata = os.path.join(str(main_folder), "static_data", "vlinder_metadata.csv")
-
-
-# import sys
-# sys.path.append(str(main_folder))
-
-
 import metobs_toolkit
 
 
-# %%
 
 # =============================================================================
-# Import data
+# Import data (See previous example script for explanations)
 # =============================================================================
-
-
-# 1. Importing a dataset containing mulitple different stations is a function in the Dataset class. First we need to initiate a Dataset with a name of choise.
-aug_2020_all_vlinders = metobs_toolkit.Dataset()
-aug_2020_all_vlinders.update_settings(
-    input_data_file=testdata_file,  # A demo data file, downloaded with brian tool: https://vlinder.ugent.be/vlinderdata/multiple_vlinders.php
-    input_metadata_file=metadata,
+your_dataset = metobs_toolkit.Dataset()
+your_dataset.update_settings(
+    input_data_file=metobs_toolkit.demo_datafile, # path to the data file
+    input_metadata_file=metobs_toolkit.demo_metadatafile,
+    template_file=metobs_toolkit.demo_template,
 )
-# Coordinates needed to extract LCZ!
-aug_2020_all_vlinders.import_data_from_file()
 
-
-# The metadata is stored in the metadf attribute:
-
-print(aug_2020_all_vlinders.metadf.head())
-
-# you can see that there is no LCZ information yet. Als long as the coordinates are present, the lcz can be extracted.
+your_dataset.import_data_from_file()
 
 
 # =============================================================================
-#  Get LCZ
+# Using the Google Earth Engine
 # =============================================================================
 
-# To use the LCZ functions, you need a google develeopers account to make use of google earth engine (gee).
-# Creating first such an account, thans simply use this function to extract the LCZ for all stations in your metadata
 
-lcz_values = aug_2020_all_vlinders.get_lcz()
+# The Google Earth Engine (GEE) is used to extract geospatial information at the
+# locations of the stations.
+# To make use of the GEE, you need to setup a google developers account (free of charge).
+# The details and steps are documented at the following page: https://vergauwenthomas.github.io/MetObs_toolkit/gee_authentication.html#
 
+# Follow the steps on the 'Using Google Earth Engine' page and then continue.
+
+# =============================================================================
+# Extracting LCZ from GEE
+# =============================================================================
+# The metadata of your stations is stored in the .metadf attribute of your dataset.
+
+print(your_dataset.metadf)
+
+# In order to extract geospatial information for you stations, the lat and lon (latitude and longitude)
+# of your stations must be present in the metadf. If so, than geospatial
+# information will be extracted from GEE at these locations.
+
+# To extract the Local Climate Zones (LCZ's) of your stations:
+
+lcz_values = your_dataset.get_lcz()
+
+# The first time, in each session, you are asked to authenticated by Google.
+# Select your google account and billing project that you have set up and accept
+# the terms of condition.
+# NOTE: For small datarequest the read-only scopes are sufficient, for large
+# datarequests this is insufficient because the data will be written directly to your google Drive.
+
+
+# The LCZ's for all your stations are extracted
 print(lcz_values)
[…]
