### 2023-11-06 vergauwenthomas/MetObs_toolkit#388 @vergauwenthomas
@Zeitsperre Thanks for this comment, this is much-appreciated advice! I will add this to my todo's.
