### 2023-11-07 965b6a6
add nbsphinx to extensions
