+++ docs/examples.rst
@@ -1,40 +1,0 @@
-*******************
-Examples
-*******************
-.. contents:: Table of Contents
-
-
-The following examples can be used to get you started using the basic functionallities
-of the toolkit. These examples are making use of the demo data files that are comes with the toolkit.
-
-Making a Dataset and inspecting it
-========================================
-
-.. literalinclude:: ../examples/introduction_example.py
-   :language: python
-
-
-Physiography information
-========================
-
-.. literalinclude:: ../examples/lcz_example.py
-   :language: python
-
-
-Applying Qualitycontrol
-========================
-
-.. literalinclude:: ../examples/qualitycontrol_example.py
-   :language: python
-   :linenos:
-   :lines: 21-159
-
-
-Making timeseries plots
-========================
-
-.. literalinclude:: ../examples/plotting_example.py
-   :language: python
-   :linenos:
-   :lines: 46-102
-
