### 2023-11-07 574e53b
converted examples into notebooks
