+++ docs/examples/index.rst
@@ -1,0 +1,10 @@
+####################
+  Examples
+####################
+
+.. toctree::
+   :maxdepth: 2
+
+   doc_example.ipynb
+   gee_example.ipynb
+   qc_example.ipynb
