### 2023-11-08 038c5e7
created a concat_save function to fix the Futurewarning of pandas for concating empyt and NA-entrie dataframes
