+++ metobs_toolkit/dataset.py
@@ -84,7 +84,8 @@ from metobs_toolkit.df_helpers import (
     conv_applied_qc_to_df,
     get_freqency_series,
     value_labeled_doubleidxdf_to_triple_idxdf,
-    xs_save
+    xs_save,
+    concat_save
 )
 
 from metobs_toolkit.analysis import Analysis
@@ -189,7 +190,7 @@ class Dataset:
         other.df = other.df.drop(common_indexes)
 
         # set new df
-        new.df = pd.concat([self.df, other.df])
+        new.df = concat_save([self.df, other.df])
         new.df = new.df.sort_index()
 
         #  ----- outliers df ---------
@@ -197,7 +198,7 @@ class Dataset:
         other_outliers = other.outliersdf.reset_index()
         other_outliers = other_outliers[other_outliers['obstype'].isin(self_obstypes)]
         other_outliers = other_outliers.set_index(['name', 'datetime', 'obstype'])
-        new.outliersdf = pd.concat([self.outliersdf, other_outliers])
+        new.outliersdf = concat_save([self.outliersdf, other_outliers])
         new.outliersdf = new.outliersdf.sort_index()
 
         #  ------- Gaps -------------
@@ -214,7 +215,7 @@ class Dataset:
 
         # ---------- metadf -----------
         # Use the metadf from self and add new rows if they are present in other
-        new.metadf = pd.concat([self.metadf, other.metadf])
+        new.metadf = concat_save([self.metadf, other.metadf])
         new.metadf = new.metadf.drop_duplicates(keep='first')
         new.metadf = new.metadf.sort_index()
 
@@ -268,7 +269,7 @@ class Dataset:
 
         checknames = ["duplicated_timestamp"]  # KEEP order
 
-        new._applied_qc = pd.concat(
+        new._applied_qc = concat_save(
             [
                 new._applied_qc,
                 conv_applied_qc_to_df(
@@ -1020,7 +1021,7 @@ class Dataset:
         filldf_debias = make_gapfill_df(debias_gaps)
 
         # combine both fill df's
-        comb_df = pd.concat([filldf_interp, filldf_debias])
+        comb_df = concat_save([filldf_interp, filldf_debias])
 
         # update attr
         self.gapfilldf = comb_df
@@ -1425,10 +1426,10 @@ class Dataset:
                 # update the dataset and outliers
                 self.df = obsdf
                 if not outl_df.empty:
-                    self.outliersdf = pd.concat([self.outliersdf, outl_df])
+                    self.outliersdf = concat_save([self.outliersdf, outl_df])
 
                 # add this check to the applied checks
-                self._applied_qc = pd.concat(
+                self._applied_qc = concat_save(
                     [
                         self._applied_qc,
                         conv_applied_qc_to_df(
@@ -1454,10 +1455,10 @@ class Dataset:
                 # update the dataset and outliers
                 self.df = obsdf
                 if not outl_df.empty:
-                    self.outliersdf = pd.concat([self.outliersdf, outl_df])
+                    self.outliersdf = concat_save([self.outliersdf, outl_df])
 
                 # add this check to the applied checks
-                self._applied_qc = pd.concat(
+                self._applied_qc = concat_save(
                     [
                         self._applied_qc,
                         conv_applied_qc_to_df(
@@ -1483,10 +1484,10 @@ class Dataset:
                 # update the dataset and outliers
                 self.df = obsdf
                 if not outl_df.empty:
-                    self.outliersdf = pd.concat([self.outliersdf, outl_df])
+                    self.outliersdf = concat_save([self.outliersdf, outl_df])
 
                 # add this check to the applied checks
-                self._applied_qc = pd.concat(
+                self._applied_qc = concat_save(
                     [
                         self._applied_qc,
                         conv_applied_qc_to_df(
[…]
