### 2023-11-08 6861333
fixed the futurwarning when appending an emtpy multiindex
