### 2023-11-08 768cc08
add new examples to the documentation
