+++ docs/examples/index.rst
@@ -8,3 +8,5 @@
    doc_example.ipynb
    gee_example.ipynb
    qc_example.ipynb
+   filling_example.ipynb
+   analysis_example.ipynb
