### 2023-11-08 fc656c8
minor bugfix when a column is not present
