### 2023-11-08 ed2a1cc
minor bugfix to ensure unit conversion
