### 2023-11-09 46662ce
run documentation examples as tests + build the documentation as test
