### 2023-11-09 453fd63
Make sure the meta columns are kept if not metadata is available (solves #392)
