### 2023-11-09 537e982
add groups to the pyproject so a poetry-documentation env can be used
