+++ pyproject.toml
@@ -22,11 +22,25 @@ mapclassify = "^2.4.0"
 earthengine-api = "^0.1.340"
 geemap = '^0.20.0'
 
-
 titanlib = { version = "^0.3", optional = true }
+
+
+
+[tool.poetry.group.documentation.dependencies]
+#Group of dep packages for building the documentation
+
+sphinx = '^7.2'
+nbsphinx = '^0.9'
+sphinx-copybutton = '^0.5.1'
+sphinx-rtd-theme = '^1.3.0'
+myst_parser = '^2.0.0'
+
+
+
 [tool.poetry.extras]
 titanlib = ["titanlib"]
 #titanlib = "^0.3" #can give errors due to no default C-compiler in windows ....
+
 
 
 [tool.poetry.group.dev.dependencies]
