### 2023-11-09 17a5d91
avoid using the requirements file, now directly using poetry in documentation group
