+++ .github/workflows/build_and_deploy_doc_dev.yml
@@ -1,4 +1,4 @@
-name: Build and deploy documentation
+name: Build and deploy documentation (dev)
 on:
   push:
     branches:
@@ -13,15 +13,11 @@ jobs:
       - uses: actions/setup-python@v3
       - name: Install dependencies
         run: |
-          pip install sphinx==6.2.1 sphinx_rtd_theme sphinx-copybutton myst_parser
           pip install poetry
-      - name: poetry create requirements
+      - name: poetry create doc environment
         run: |
           poetry update
-          poetry export -f requirements.txt --output requirements.txt
-      - name: install package using pip
-        run: |
-          pip install -r requirements.txt
+          poetry install --with documentation,dev,titan
       - name: Sphinx build
         run: |
           sphinx-build -a -E docs _build
