### 2023-11-09 1125d94
gee map is always returned see codereview
