### 2023-11-09 c317726
the old examples are removed so no need to scan this folder
