### 2023-11-09 48ba0ab
update the introduction example to use the demo datafiles + documentation update (#390)
* update the introdction example to use the demo datafiles

* update the lcz example

* add nbsphinx to extensions

* examples in seperate folder

* converted examples into notebooks

* created a concat_save function to fix the Futurewarning of pandas for concating empyt and NA-entrie dataframes

* fixed the futurwarning when appending an emtpy multiindex

* add notebook checkpoints to the ignore

* Updated the examples for the documentation.

* delete checkpoints from git

* add new examples to the documen
[…]
