+++ metobs_toolkit/df_helpers.py
@@ -78,6 +78,17 @@ def xs_save(df, key, level, drop_level=True):
     return pd.DataFrame(index=idx, columns=columns)
 
 
+def concat_save(df_list, **kwargs):
+    """Concat dataframes row-wise without triggering the Futurwarning of concating empyt df's."""
+
+    if all([isinstance(df, pd.DataFrame) for df in df_list]):
+        # This line will filter columns with all NAN values (so empty dfs + all NA entries are filtered out)
+        return pd.concat([df.dropna(axis=1, how='all') for df in df_list], **kwargs)
+    if all([isinstance(df, pd.Series) for df in df_list]):
+        # This line will filter out empty series
+        return pd.concat([ser for ser in df_list if not ser.empty], **kwargs)
+    sys.exit('Cannot concat Dataframes and Series together')
+
 def init_multiindex():
     """Construct a name-datetime pandas multiindex."""
     return pd.MultiIndex(
@@ -173,7 +184,7 @@ def value_labeled_doubleidxdf_to_triple_idxdf(df ,value_col_name='value',
         subdf = subdf.set_index(['name', 'datetime', 'obstype'])
         subdf = subdf.rename(columns={obstype + '_final_label': label_col_name})
 
-        labelsdf = pd.concat([labelsdf, subdf])
+        labelsdf = concat_save([labelsdf, subdf])
 
     values[label_col_name] = labelsdf[label_col_name]
 
@@ -236,7 +247,15 @@ def metadf_to_gdf(df, crs=4326):
         coordsdf, geometry=gpd.points_from_xy(coordsdf.lon, coordsdf.lat)
     )
     geodf = geodf.set_crs(epsg=crs)
-    geodf = pd.concat([geodf, missing_coords_df])
+    metadata_columns = geodf.columns
+    geodf = concat_save([geodf, missing_coords_df])
+
+    # Because empyt and Nan columns are skipped in the concat save, add them
+    # again if needed
+    for col in metadata_columns:
+        if col not in geodf:
+            geodf[col] = np.nan
+
 
     geodf = geodf.sort_index()
     return geodf
