+++ metobs_toolkit/__init__.py
@@ -84,5 +84,5 @@ from metobs_toolkit.dataset_settings_updater import Dataset
 # =============================================================================
 
 # DO not change this manually!
-__version__ = "0.1.2"
+__version__ = "0.1.2beta"
 
