+++ docs/intro.rst
@@ -31,12 +31,12 @@ To install the PyPi version of the toolkit. To install the github versions one c
 
 
 For some advanced quality control methods, the `Titanlib <https://github.com/metno/titanlib>`_ package is used.
-Since the instalation of titanlib requires a c++ compiler, it is categorized as a *extra-dependency*. This means that
-the user must install titanlib manually if this functionallity is required or use the following command:
+Since the instalation of titanlib requires a c++ compiler, we have choosen not to include it in the toolkit. If your want to use
+the Titanlib functionality you must install both the toolkit and Titanlib:
 
 .. code-block:: console
 
-   pip3 install metobs-toolkit[titanlib]
+   pip3 install metobs-toolkit titanlib
 
 
 .. note::
