+++ .github/workflows/test_doc_build.yml
@@ -1,0 +1,28 @@
+name: Build and deploy documentation (testing)
+on:
+  push
+permissions:
+    contents: write
+jobs:
+  docs:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-python@v3
+      - name: Install dependencies
+        run: |
+          pip install poetry
+      - name: poetry create doc environment
+        run: |
+          poetry update
+          poetry install --with documentation,dev,titan
+      - name: Sphinx build
+        run: |
+          sphinx-build -a -E docs _build
+      - name: Deploy
+        uses: peaceiris/actions-gh-pages@v3
+        with:
+          publish_branch: gh-pages
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+          publish_dir: _build/
+          force_orphan: true
