### 2023-11-09 a5cc1b3
restrict poetry version
