+++ .github/workflows/test_doc_build.yml
@@ -11,13 +11,16 @@ jobs:
       - uses: actions/setup-python@v3
       - name: Install dependencies
         run: |
-          pip install poetry
+          pip install poetry==1.3.2
       - name: poetry create doc environment
         run: |
+          poetry --version
           poetry update
+          poetry --version
           poetry install --with documentation,dev,titan
       - name: Sphinx build
         run: |
+          poetry --version
           sphinx-build -a -E docs _build
       - name: Deploy
         uses: peaceiris/actions-gh-pages@v3
