### 2023-11-09 44a9cd5
try without explicit version control on numpy
