+++ pyproject.toml
@@ -13,7 +13,7 @@ keywords = ["meteorology", "observations", "urban climate"]
 [tool.poetry.dependencies]
 python = "^3.9" #can be converted to 3.8 but furher investigation needed
 pandas = "^1.3.0"
-numpy = "^1.17.3"
+#numpy = "^1.17.3"
 matplotlib = "^3.0.0"
 #mysql-connector-python = "^8.0.6"
 geopandas = "^0.9.0"
