### 2023-11-09 e9e7686
force numpy to 1.26.0 (as on my local env)
