+++ .github/workflows/test_doc_build.yml
@@ -12,6 +12,7 @@ jobs:
       - name: Install dependencies
         run: |
           pip install poetry==1.3.2
+          pip install numpy==1.26.0
       - name: poetry create doc environment
         run: |
           poetry --version
