+++ .github/workflows/test_doc_build.yml
@@ -15,9 +15,8 @@ jobs:
       - name: poetry create doc environment
         run: |
           poetry --version
-          poetry update
-          poetry --version
           poetry install --only documentation
+          poetry show
       - name: Sphinx build
         run: |
           poetry --version
