### 2023-11-09 e63c227
new test with recent poetry version
