+++ .github/workflows/test_doc_build.yml
@@ -11,10 +11,10 @@ jobs:
       - uses: actions/setup-python@v3
       - name: Install dependencies
         run: |
-          pip install poetry==1.3.2
+          pip install poetry==1.7.0
       - name: poetry create doc environment
         run: |
-          poetry --version
+          poetry update
           poetry install --with documentation
           poetry show
       - name: Sphinx build
