+++ pyproject.toml
@@ -13,7 +13,7 @@ keywords = ["meteorology", "observations", "urban climate"]
 [tool.poetry.dependencies]
 python = "^3.9"
 pandas = "^1.3.0"
-numpy = "^1.17.3"
+#numpy = "^1.17.3"
 matplotlib = "^3.0.0"
 geopandas = "^0.9.0"
 pyproj = "~3.4"
