+++ .github/workflows/test_doc_build.yml
@@ -1,4 +1,4 @@
-name: Build and deploy documentation (testing)
+name: Build and deploy documentation (test)
 on:
   push
 permissions:
@@ -11,17 +11,18 @@ jobs:
       - uses: actions/setup-python@v3
       - name: Install dependencies
         run: |
-          pip install poetry==1.7.0
-      - name: poetry create doc environment
+          pip install sphinx==6.2.1 sphinx_rtd_theme sphinx-copybutton myst_parser nbsphinx
+          pip install poetry
+      - name: poetry create requirements
         run: |
           poetry update
-          poetry install --with documentation
-          poetry show
+          poetry export -f requirements.txt --output requirements.txt
+      - name: install package using pip
+        run: |
+          pip install -r requirements.txt
       - name: Sphinx build
         run: |
-          poetry --version
-          poetry run sphinx-build --version
-          poetry run sphinx-build -a -E docs _build
+          sphinx-build -a -E docs _build
       - name: Deploy
         uses: peaceiris/actions-gh-pages@v3
         with:
