+++ .github/workflows/test_doc_build.yml
@@ -11,8 +11,8 @@ jobs:
       - uses: actions/setup-python@v3
       - name: Install dependencies
         run: |
-          pip install sphinx==6.2.1 sphinx_rtd_theme sphinx-copybutton myst_parser nbsphinx setuptools
-          pip install poetry
+          python3 -m pip install sphinx==6.2.1 sphinx_rtd_theme sphinx-copybutton myst_parser nbsphinx setuptools
+          python3 -m pip install poetry
       - name: poetry create requirements file
         run: |
           poetry --version
@@ -20,7 +20,7 @@ jobs:
           poetry export -f requirements.txt --output requirements.txt
       - name: install package using pip
         run: |
-          pip install -r requirements.txt
+          python3 -m pip install -r requirements.txt
       - name: Sphinx build
         run: |
           sphinx-build -a -E docs _build
