### 2023-11-10 d043e15
test doc building in the python matrix of the master test script
