+++ .github/workflows/test_doc_build.yml
@@ -11,6 +11,7 @@ jobs:
       - uses: actions/setup-python@v3
       - name: Install dependencies
         run: |
+          python3 --version
           python3 -m pip install sphinx==6.2.1 sphinx_rtd_theme sphinx-copybutton myst_parser nbsphinx setuptools
           python3 -m pip install poetry
       - name: poetry create requirements file
