### 2023-11-10 9ba0e51
add pandoc to doc dep group
