+++ pyproject.toml
@@ -41,7 +41,7 @@ nbsphinx = '^0.9'
 sphinx-copybutton = '^0.5.1'
 sphinx-rtd-theme = '^1.3.0'
 myst_parser = '^2.0.0'
-pandoc = '^3.1'
+pandoc = '^3.1.3'
 
 
 [tool.poetry.group.dev.dependencies]
