+++ pyproject.toml
@@ -41,7 +41,7 @@ nbsphinx = '^0.9'
 sphinx-copybutton = '^0.5.1'
 sphinx-rtd-theme = '^1.3.0'
 myst_parser = '^2.0.0'
-pandoc = '^2'
+pandoc = '^2' #Check version on PyPi (not in local conda env)
 
 
 [tool.poetry.group.dev.dependencies]
