+++ .github/workflows/test_doc_build.yml
@@ -17,7 +17,6 @@ jobs:
 
     steps:
       - uses: actions/checkout@v3
-      - uses: docker://pandoc/core:2.9
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v4
         with:
