+++ .github/workflows/test_doc_build.yml
@@ -40,3 +40,12 @@ jobs:
         run: |
             poetry show
             poetry run sphinx-build -a -E docs _build
+
+      - name: Deploy
+        uses: peaceiris/actions-gh-pages@v3
+        with:
+          publish_branch: gh-pages
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+          publish_dir: _build/
+          force_orphan: true
+
