+++ .github/workflows/test_doc_build.yml
@@ -1,51 +1,0 @@
-name: test Build documentation
-
-on: [push]
-
-jobs:
-  build:
-
-    runs-on: ${{ matrix.os }}
-    strategy:
-      # You can use PyPy versions in python-version.
-      # For example, pypy-2.7 and pypy-3.8
-      matrix:
-        python-version: ["3.9"]
-        # os: [ubuntu-latest, windows-latest]
-        os: [ubuntu-latest]
-        node-version: [12.x]
-
-    steps:
-      - uses: actions/checkout@v3
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v4
-        with:
-          python-version: ${{ matrix.python-version }}
-          #cache: 'poetry'
-      # You can test your matrix by printing the current Python version
-      - name: Display Python version
-        run: python -c "import sys; print(sys.version)"
-      - name: install pandoc (system wide)
-        run: |
-            sudo apt-get -y install pandoc
-      - name: Install poetry
-        run: |
-            pip install poetry
-      - name: Install dependencies and build
-        run: |
-            poetry update
-            poetry install --all-extras
-            poetry build
-      - name: Build documentation
-        run: |
-            poetry show
-            poetry run sphinx-build -a -E docs _build
-
-      - name: Deploy
-        uses: peaceiris/actions-gh-pages@v3
-        with:
-          publish_branch: gh-pages
-          github_token: ${{ secrets.GITHUB_TOKEN }}
-          publish_dir: _build/
-          force_orphan: true
-
