+++ .github/workflows/master_test.yml
@@ -48,6 +48,7 @@ jobs:
     #needs: run-tests
     runs-on: ubuntu-latest
     steps:
+      - uses: actions/checkout@v3
       - name: navigate
         run: |
           pwd
