+++ .github/workflows/master_test.yml
@@ -60,6 +60,7 @@ jobs:
           echo "current version (init) = ${CURRENT_VERSION}"
           PYPROJECT_VERSION="$(grep -E 'version'  pyproject.toml | cut -d ' ' -f3)"
           echo "current version (pyproject) = ${PYPROJECT_VERSION}"
+      - name: sync versions
         if: ${{ steps.version_info.outputs.CURRENT_VERSION }} != ${{ steps.version_info.outputsPYPROJECT_VERSION }}
         run: exit 1
 
