+++ .github/workflows/master_test.yml
@@ -61,8 +61,11 @@ jobs:
           PYPROJECT_VERSION="$(grep -E 'version'  pyproject.toml | cut -d ' ' -f3)"
           echo "current version (pyproject) = ${PYPROJECT_VERSION}"
       - name: sync versions
-        if: ${{ steps.version_info.outputs.CURRENT_VERSION }} != ${{ steps.version_info.outputsPYPROJECT_VERSION }}
-        run: exit 1
+        if: ${{ steps.version_info.outputs.CURRENT_VERSION }} != ${{ steps.version_info.outputs.PYPROJECT_VERSION }}
+        run: |
+          echo ${steps.version_info.outputs.CURRENT_VERSION}
+          echo ${steps.version_info.outputs.PYPROJECT_VERSION}
+          exit 1
 
 
   # doctests:
