+++ .github/workflows/master_test.yml
@@ -60,12 +60,9 @@ jobs:
           echo "current version (init) = ${CURRENT_VERSION}"
           PYPROJECT_VERSION="$(grep -E 'version'  pyproject.toml | cut -d ' ' -f3)"
           echo "current version (pyproject) = ${PYPROJECT_VERSION}"
-
-      - name: version is correct
-         if: ${{ steps.version_info.outputs.CURRENT_VERSION == steps.version_info.outputs.CURRENT_VERSION }}
-         run: echo "version tag is alined"
-
-
+      - name: version-is-correct
+        if: ${{ steps.version_info.outputs.CURRENT_VERSION == steps.version_info.outputs.CURRENT_VERSION }}
+        run: echo "version tag is alined"
   # doctests:
   #   name: documentation-test
   #   #needs: run-tests
