### 2023-11-10 c37c1ec
with some gpt help maybe?
