+++ .github/workflows/master_test.yml
@@ -32,7 +32,7 @@ jobs:
        - name: Install dependencies and build
          run: |
              poetry update
-             poetry install --all-extras
+             poetry install
              poetry build
        - name: Run tests
          run: |
