+++ pyproject.toml
@@ -13,17 +13,13 @@ keywords = ["meteorology", "observations", "urban climate"]
 [tool.poetry.dependencies]
 python = "^3.9"
 pandas = "^1.3.0"
-#numpy = "^1.17.3"
+#numpy = "^1.17.3" #in pandas
 matplotlib = "^3.0.0"
 geopandas = "^0.9.0"
 pyproj = "~3.4"
 mapclassify = "^2.4.0"
 earthengine-api = "^0.1.340"
 geemap = '^0.20.0'
-
-[tool.poetry.group.setup.dependencies]
-#setuptools = '^68.2'
-future = '0.18.1'
 
 
 
@@ -35,13 +31,12 @@ titanlib = '^0.3'
 
 [tool.poetry.group.documentation.dependencies]
 #Group of dep packages for building the documentation
-
 sphinx = '^7.2'
 nbsphinx = '^0.9'
 sphinx-copybutton = '^0.5.1'
 sphinx-rtd-theme = '^1.3.0'
 myst_parser = '^2.0.0'
-pandoc = '^2' #Check on PyPi (not in local conda env)
+pandoc = '^2' #Check on PyPi (not in local conda env) + with system wide install
 
 
 [tool.poetry.group.dev.dependencies]
