+++ .github/workflows/master_test.yml
@@ -71,7 +71,7 @@ jobs:
 
    doctests:
      name: documentation-test
-     #needs: run-tests
+     #needs: versiontest
      runs-on: ubuntu-latest
      steps:
        - uses: actions/checkout@v3
