+++ .github/workflows/master_test.yml
@@ -109,10 +109,10 @@ jobs:
       steps:
         - uses: peaceiris/actions-gh-pages@v3
            with:
-              publish_branch: gh-pages
-              github_token: ${{ secrets.GITHUB_TOKEN }}
-              publish_dir: _build/
-              force_orphan: true
+             publish_branch: gh-pages
+             github_token: ${{ secrets.GITHUB_TOKEN }}
+             publish_dir: _build/
+             force_orphan: true
 
    deploy_doc_dev:
       name: Deploy main documentation
@@ -122,10 +122,10 @@ jobs:
       steps:
         - uses: peaceiris/actions-gh-pages@v3
            with:
-              publish_branch: gh-pages-dev
-              github_token: ${{ secrets.GITHUB_TOKEN }}
-              publish_dir: _build/
-              force_orphan: true
+             publish_branch: gh-pages-dev
+             github_token: ${{ secrets.GITHUB_TOKEN }}
+             publish_dir: _build/
+             force_orphan: true
 
 
 
