+++ .github/workflows/master_test.yml
@@ -103,12 +103,14 @@ jobs:
 #---- Deploy documentation -----#
    deploy_doc:
       name: Deploy main documentation
-      needs: doctest
+      needs: doctests
       runs-on: ubuntu-latest
       if: github.ref == 'refs/heads/master'
       steps:
-          uses: peaceiris/actions-gh-pages@v3
-          with:
+         - uses: actions/checkout@v3
+         - name: deploy artifact
+           uses: peaceiris/actions-gh-pages@v3
+           with:
             publish_branch: gh-pages
             github_token: ${{ secrets.GITHUB_TOKEN }}
             publish_dir: _build/
