+++ .github/workflows/main_workflow.yml
@@ -43,6 +43,57 @@ jobs:
              poetry run python tests/push_test/breaking_test.py
              #poetry run python tests/push_test/gui_launch_test.py
              poetry run python tests/push_test/analysis_test.py
+
+#--- Package os installation ---
+   os_install_testing:
+   name: test package installation on multiple os
+   if: (github.ref == 'refs/heads/dev' || github.ref == 'refs/heads/master')
+   strategy:
+     fail-fast: true
+     matrix:
+       os: [ "ubuntu-latest", "macos-latest", "windows-latest" ]
+       python-version: ["3.9", "3.10"]
+   defaults:
+     run:
+       shell: bash
+   runs-on: ${{ matrix.os }}
+   steps:
+     #---
+     #       check-out repo and set-up python
+     #----
+     - name: Check out repository
+       uses: actions/checkout@v3
+     - name: Set up python ${{ matrix.python-version }}
+       id: setup-python
+       uses: actions/setup-python@v4
+       with:
+         python-version: ${{ matrix.python-version }}
+     #--
+     #  -----  install & configure poetry  -----
+     #--
+     - name: Install Poetry
+       uses: snok/install-poetry@v1
+       with:
+         version: 1.3.2
+         virtualenvs-create: true
+         virtualenvs-in-project: true
+     - name: Load cached venv
+       id: cached-pip-wheels
+       uses: actions/cache@v3
+       with:
+         path: ~/.cache
+         key: venv-${{ runner.os }}-${{ steps.setup-python.outputs.python-version }}-${{ hashFiles('**/poetry.lock') }}
+     - name: Update lock
+       run: poetry update
+     - name: Install dependencies
+       run: poetry install --no-interaction --no-root
+     - name: Install library
+       run: poetry install --no-interaction
+     - name: Run tests
+       run: |
+         source $VENV
+         python tests/push_test/import_test.py
+
 #---- Version Control -----#
    versiontest:
      name: check if version is valid
@@ -103,7 +154,7 @@ jobs:
 #---- Deploy documentation -----#
    deploy_doc:
       name: Deploy main documentation
-      needs: [doctests,run-tests,versiontest]
+      needs: [doctests,run-tests,versiontest,os_install_testing]
       runs-on: ubuntu-latest
       if: github.ref == 'refs/heads/master'
       steps:
@@ -130,6 +181,8 @@ jobs:
             github_token: ${{ secrets.GITHUB_TOKEN }}
             publish_dir: _build/
             force_orphan: true
+
+
 
 
 
