+++ .github/workflows/main_workflow.yml
@@ -71,28 +71,17 @@ jobs:
         #--
         #  -----  install & configure poetry  -----
         #--
-        - name: Install Poetry
-          uses: snok/install-poetry@v1
-          with:
-            version: 1.3.2
-            virtualenvs-create: true
-            virtualenvs-in-project: true
-        - name: Load cached venv
-          id: cached-pip-wheels
-          uses: actions/cache@v3
-          with:
-            path: ~/.cache
-            key: venv-${{ runner.os }}-${{ steps.setup-python.outputs.python-version }}-${{ hashFiles('**/poetry.lock') }}
-        - name: Update lock
-          run: poetry update
-        - name: Install dependencies
-          run: poetry install --no-interaction --no-root
-        - name: Install library
-          run: poetry install --no-interaction
-        - name: Run tests
+        - name: Install poetry
           run: |
-            source $VENV
-            python tests/push_test/import_test.py
+              pip install poetry
+        - name: Install the toolkit
+          run: |
+              poetry update
+              poetry install
+        - name: Run import test
+          run: |
+              poetry run python tests/push_test/import_test.py
+
 
 #---- Version Control -----#
    versiontest:
