+++ .github/workflows/main_workflow.yml
@@ -53,7 +53,7 @@ jobs:
         fail-fast: true
         matrix:
           os: [ "ubuntu-latest", "macos-latest", "windows-latest" ]
-          python-version: ["3.9", "3.10"]
+          python-version: ["3.9"]
       defaults:
         run:
           shell: bash
@@ -77,7 +77,8 @@ jobs:
         - name: Install the toolkit
           run: |
               poetry update
-              poetry install
+              poetry install --without titan,documentation
+              poetry show
         - name: Run import test
           run: |
               poetry run python tests/push_test/import_test.py
