### 2023-11-10 vergauwenthomas/MetObs_toolkit#384 @Zeitsperre
Nicely done!
