### 2023-11-10 cc31a23
trigger doc builder (#393)
* trigger doc builder

* restrict poetry version

* try without explicit version control on numpy

* force numpy to 1.26.0 (as on my local env)

* try without dev group

* with --only arg

* without the update

* explicit add poetry run for sphinx build

* poetry with default deps

* new test with recent poetry version

* without np explicit

* set back to requirements

* pfff again utils error ... is this a poetry bug?

* add python3 -m explicit before pip

* test doc building in the python matrix of the master test script

* add pandoc to doc dep group

* pandoc ve
[…]
