+++ .github/workflows/master_test.yml
@@ -1,48 +1,0 @@
-name: Test pushed version
-
-on: [push]
-
-jobs:
-  build:
-
-    runs-on: ${{ matrix.os }}
-    strategy:
-      # You can use PyPy versions in python-version.
-      # For example, pypy-2.7 and pypy-3.8
-      matrix:
-        python-version: ["3.9", "3.10"]
-        # os: [ubuntu-latest, windows-latest]
-        os: [ubuntu-latest]
-        node-version: [12.x]
-
-    steps:
-      - uses: actions/checkout@v3
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v4
-        with:
-          python-version: ${{ matrix.python-version }}
-          #cache: 'poetry'
-      # You can test your matrix by printing the current Python version
-      - name: Display Python version
-        run: python -c "import sys; print(sys.version)"
-      - name: Install poetry
-        run: |
-            pip install poetry
-
-      - name: Install dependencies and build
-        run: |
-            poetry update
-            poetry install --all-extras
-            poetry build
-      - name: Run tests
-        run: |
-            cd tests
-            cd push_test
-            poetry run python gap_and_fill_test.py
-            poetry run python IO_test.py
-            poetry run python qc_test.py
-            poetry run python breaking_test.py
-            #poetry run python gui_launch_test.py
-            poetry run python analysis_test.py
-
-
