### 2023-11-12 #5916 @ashwinvis
The fix broke the docs, I believe. Waiting for it again... https://github.com/vergauwenthomas/MetObs_toolkit/issues/394
