+++ .github/workflows/main_workflow.yml
@@ -159,7 +159,9 @@ jobs:
       name: Deploy dev documentation
       needs: [doctests]
       runs-on: ubuntu-latest
-      if: github.ref == 'refs/heads/dev'
+      if: |
+         github.ref == 'refs/heads/dev' || "!contains(github.event.head_commit.message, 'build doc')"
+
       steps:
          - uses: actions/checkout@v3
          - name: deploy artifact
