+++ .github/workflows/main_workflow.yml
@@ -161,7 +161,6 @@ jobs:
       runs-on: ubuntu-latest
       if: |
          github.ref == 'refs/heads/dev' || "!contains(github.event.head_commit.message, 'build doc')"
-
       steps:
          - uses: actions/checkout@v3
          - name: deploy artifact
