+++ .github/workflows/main_workflow.yml
@@ -160,7 +160,7 @@ jobs:
       needs: [doctests]
       runs-on: ubuntu-latest
       if: |
-         github.ref == 'refs/heads/dev' || "!contains(github.event.head_commit.message, 'build doc')"
+         github.ref == 'refs/heads/dev' || !contains(github.event.head_commit.message, 'build doc')
       steps:
          - uses: actions/checkout@v3
          - name: deploy artifact
