+++ .github/workflows/main_workflow.yml
@@ -4,109 +4,109 @@ on: [push]
 
 jobs:
 #---- Run tests -----#
-   run-tests:
-     name: run test scripts
-     runs-on: ${{ matrix.os }}
-     strategy:
-       # You can use PyPy versions in python-version.
-       # For example, pypy-2.7 and pypy-3.8
-       matrix:
-         python-version: ["3.9", "3.10"]
-         # os: [ubuntu-latest, windows-latest]
-         os: [ubuntu-latest]
-         node-version: [12.x]
+#    run-tests:
+#      name: run test scripts
+#      runs-on: ${{ matrix.os }}
+#      strategy:
+#        # You can use PyPy versions in python-version.
+#        # For example, pypy-2.7 and pypy-3.8
+#        matrix:
+#          python-version: ["3.9", "3.10"]
+#          # os: [ubuntu-latest, windows-latest]
+#          os: [ubuntu-latest]
+#          node-version: [12.x]
 
-     steps:
-       - uses: actions/checkout@v3
-       - name: Set up Python ${{ matrix.python-version }}
-         uses: actions/setup-python@v4
-         with:
-           python-version: ${{ matrix.python-version }}
-           #cache: 'poetry'
-       # You can test your matrix by printing the current Python version
-       - name: Display Python version
-         run: python -c "import sys; print(sys.version)"
-       - name: Install poetry
-         run: |
-             pip install poetry
+#      steps:
+#        - uses: actions/checkout@v3
+#        - name: Set up Python ${{ matrix.python-version }}
+#          uses: actions/setup-python@v4
+#          with:
+#            python-version: ${{ matrix.python-version }}
+#            #cache: 'poetry'
+#        # You can test your matrix by printing the current Python version
+#        - name: Display Python version
+#          run: python -c "import sys; print(sys.version)"
+#        - name: Install poetry
+#          run: |
+#              pip install poetry
 
-       - name: Install dependencies and build
-         run: |
-             poetry update
-             poetry install
-       - name: Run tests
-         run: |
-             poetry run python tests/push_test/gap_and_fill_test.py
-             poetry run python tests/push_test/IO_test.py
-             poetry run python tests/push_test/qc_test.py
-             poetry run python tests/push_test/breaking_test.py
-             #poetry run python tests/push_test/gui_launch_test.py
-             poetry run python tests/push_test/analysis_test.py
+#        - name: Install dependencies and build
+#          run: |
+#              poetry update
+#              poetry install
+#        - name: Run tests
+#          run: |
+#              poetry run python tests/push_test/gap_and_fill_test.py
+#              poetry run python tests/push_test/IO_test.py
+#              poetry run python tests/push_test/qc_test.py
+#              poetry run python tests/push_test/breaking_test.py
+#              #poetry run python tests/push_test/gui_launch_test.py
+#              poetry run python tests/push_test/analysis_test.py
 
-#--- Package os installation ---
-   mac_install_testing:
-      name: Installation on Mac latest
-      runs-on: macos-latest
-      steps:
-       - uses: actions/checkout@v3
-       - name: Set up Python39
-         uses: actions/setup-python@v4
-         with:
-           python-version: "3.9"
-       - name: Install poetry
-         run: |
-             pip install poetry
-       - name: Install dependencies and build
-         run: |
-             poetry update --without titan,documentation
-             poetry install --without titan,documentation
-       - name: Run tests
-         run: |
-              poetry run python tests/push_test/import_test.py
+# #--- Package os installation ---
+#    mac_install_testing:
[…]
