+++ .github/workflows/main_workflow.yml
@@ -139,6 +139,10 @@ jobs:
              poetry show
              poetry run sphinx-build -a -E docs _build
        - name: save documentation build
+         if: |
+           github.ref == 'refs/heads/dev' ||
+           github.ref == 'refs/head/main' ||
+           contains(github.event.head_commit.message, 'build doc')
          uses: actions/upload-artifact@v3
          with:
            name: doc_build
@@ -176,8 +180,6 @@ jobs:
          - name: check _build
            run: |
                pwd
-               ls
-               cd _build
                ls
          - name: deploy artifact
            uses: peaceiris/actions-gh-pages@v3
