+++ .github/workflows/main_workflow.yml
@@ -3,6 +3,7 @@ name: Test pushed version
 on: [push]
 
 jobs:
+# --- Build the package and save as artifacts----
    packageing:
      name: Build the package
      runs-on: ubuntu-latest
@@ -25,224 +26,503 @@ jobs:
              pip install poetry
        - name: Install dependencies and build
          run: |
-             poetry update
-             poetry install --all-extras
+             poetry update --without titan,documentation
+             poetry install --without titan,documentation
              poetry build
        - name: save package as artifact
          uses: actions/upload-artifact@v3
          with:
            name: package_build
            path: dist/*.tar.gz
-
-
-
-   testing:
-     name: test the package
-     needs: packageing
-     runs-on: ubuntu-latest
-     steps:
-       - uses: actions/checkout@v3
-       - name: Set up Python39
-         uses: actions/setup-python@v4
-         with:
-           python-version: '3.9'
-       - name: Display Python version
-         run: python -c "import sys; print(sys.version)"
-       - uses: actions/checkout@v3
-       - name: print
+       - name: build package with all extras
          run: |
-             pwd
-             ls
-       - name: Download builded documentation
+             poetry update
+             poetry install --all-extras
+        - name: export documentation requirement group
+          run: |
+              poetry export -f requirements.txt --without-hashes --only documentation -o only_doc_req.txt
+          uses: actions/upload-artifact@v3
+          with:
+            name: documentation_requirements
+            path: only_doc_req.txt
+
+        - name: export titan requirement group
+          run: |
+              poetry export -f requirements.txt --without-hashes --only titan -o only_titan_req.txt
+          uses: actions/upload-artifact@v3
+          with:
+            name: titan_requirements
+            path: only_titan_req.txt
+
+
+
+   # packageing_all_groups:
+   #   name: Build the package for seperate groups
+   #   needs: packageing
+   #   runs-on: ubuntu-latest
+   #   steps:
+   #     - uses: actions/checkout@v3
+   #     - name: Set up Python39
+   #       uses: actions/setup-python@v4
+   #       with:
+   #         python-version: '3.9'
+   #         #cache: 'poetry'
+   #     # You can test your matrix by printing the current Python version
+   #     - name: Display Python version
+   #       run: python -c "import sys; print(sys.version)"
+   #     - name: remove previous builds
+   #       run: |
+   #           rm -f dist/metobs_toolkit-*.whl
+   #           rm -f dist/metobs_toolkit-*.tar.gz
+   #     - name: Install poetry
+   #       run: |
+   #           pip install poetry
+   #     - name: Install dependencies and build
+   #       run: |
+   #           poetry update
+   #           poetry install --all-extras
+   #           poetry build
+   #     - name: save package as artifact
+   #       uses: actions/upload-artifact@v3
+   #       with:
+   #         name: package_build_all_groups
+   #         path: dist/*.tar.gz
+
+
+
+   # testing:
+   #   name: test the package
+   #   needs: packageing
+   #   runs-on: ubuntu-latest
+   #   steps:
+   #     - uses: actions/checkout@v3
+   #     - name: Set up Python39
+   #       uses: actions/setup-python@v4
+   #       with:
+   #         python-version: '3.9'
+   #     - name: Display Python version
+   #       run: python -c "import sys; print(sys.version)"
+   #     - uses: actions/checkout@v3
+   #     - name: print
+   #       run: |
+   #           pwd
+   #           ls
+   #     - name: Download builded documentation
+   #       uses: actions/download-artifact@v3
+   #       with:
+   #         name: package_build
+   #     - name: print
+   #       run: |
+   #           pwd
+   #           ls
+
+
+
+#---- Run tests -----#
+    run-tests:
[…]
