+++ .github/workflows/main_workflow.yml
@@ -38,19 +38,18 @@ jobs:
          run: |
              poetry update
              poetry install --all-extras
-        - name: export documentation requirement group
-          run: |
-              poetry export -f requirements.txt --without-hashes --only documentation -o only_doc_req.txt
-          uses: actions/upload-artifact@v3
-          with:
+       - name: export documentation requirement group
+         run: |
+             poetry export -f requirements.txt --without-hashes --only documentation -o only_doc_req.txt
+         uses: actions/upload-artifact@v3
+         with:
             name: documentation_requirements
             path: only_doc_req.txt
-
-        - name: export titan requirement group
-          run: |
-              poetry export -f requirements.txt --without-hashes --only titan -o only_titan_req.txt
-          uses: actions/upload-artifact@v3
-          with:
+       - name: export titan requirement group
+         run: |
+             poetry export -f requirements.txt --without-hashes --only titan -o only_titan_req.txt
+         uses: actions/upload-artifact@v3
+         with:
             name: titan_requirements
             path: only_titan_req.txt
 
@@ -119,44 +118,44 @@ jobs:
 
 
 #---- Run tests -----#
-    run-tests:
-      name: run test scripts
-      needs: packageing
-      runs-on: ${{ matrix.os }}
-      strategy:
+   run-tests:
+       name: run test scripts
+       needs: packageing
+       runs-on: ${{ matrix.os }}
+       strategy:
         # You can use PyPy versions in python-version.
         # For example, pypy-2.7 and pypy-3.8
-        matrix:
-          python-version: ["3.9", "3.10"]
+         matrix:
+           python-version: ["3.9", "3.10"]
           # os: [ubuntu-latest, windows-latest]
-          os: [ubuntu-latest]
-          node-version: [12.x]
-      steps:
-        - uses: actions/checkout@v3
-        - name: Set up Python ${{ matrix.python-version }}
-          uses: actions/setup-python@v4
-          with:
-            python-version: ${{ matrix.python-version }}
+           os: [ubuntu-latest]
+           node-version: [12.x]
+       steps:
+         - uses: actions/checkout@v3
+         - name: Set up Python ${{ matrix.python-version }}
+           uses: actions/setup-python@v4
+           with:
+             python-version: ${{ matrix.python-version }}
             #cache: 'poetry'
         # You can test your matrix by printing the current Python version
-        - name: Display Python version
-          run: python -c "import sys; print(sys.version)"
-        - name: Download the package build
-          uses: actions/download-artifact@v3
-          with:
-            name: package_build
-        - name: Install the package
-          run: |
-              python3 -m pip install ./metobs_toolkit-*.tar.gz
-       - name: get titanlib requirements
-         uses: actions/download-artifact@v3
-         with:
-           name: titan_requirements
-        - name: install titanlib
+         - name: Display Python version
+           run: python -c "import sys; print(sys.version)"
+         - name: Download the package build
+           uses: actions/download-artifact@v3
+           with:
+             name: package_build
+         - name: Install the package
+           run: |
+               python3 -m pip install ./metobs_toolkit-*.tar.gz
+         - name: get titanlib requirements
+           uses: actions/download-artifact@v3
+           with:
+             name: titan_requirements
+         - name: install titanlib
            run: |
                pip install -r only_titan_req.txt
-        - name: Run tests
-          run: |
+         - name: Run tests
+           run: |
               python tests/push_test/gap_and_fill_test.py
[…]
