+++ .github/workflows/main_workflow.yml
@@ -64,118 +64,118 @@ jobs:
             path: only_titan_req.txt
 
 
-# #---- Run tests -----#
-#    run-tests:
-#        name: run test scripts
-#        needs: packageing
-#        runs-on: ${{ matrix.os }}
-#        strategy:
-#         # You can use PyPy versions in python-version.
-#         # For example, pypy-2.7 and pypy-3.8
-#          matrix:
-#            python-version: ["3.9", "3.10"]
-#           # os: [ubuntu-latest, windows-latest]
-#            os: [ubuntu-latest]
-#            node-version: [12.x]
-#        steps:
-#          - uses: actions/checkout@v3
-#          - name: Set up Python ${{ matrix.python-version }}
-#            uses: actions/setup-python@v4
-#            with:
-#              python-version: ${{ matrix.python-version }}
-#             #cache: 'poetry'
-#         # You can test your matrix by printing the current Python version
-#          - name: Display Python version
-#            run: python -c "import sys; print(sys.version)"
-#          - name: Download the package build
-#            uses: actions/download-artifact@v3
-#            with:
-#              name: package_build
-#          - name: Install the package
-#            run: |
-#                python3 -m pip install ${{needs.packageing.outputs.pkg_file}}
-#          - name: get titanlib requirements
-#            uses: actions/download-artifact@v3
-#            with:
-#              name: titan_requirements
-#          - name: install titanlib
-#            run: |
-#                pip install -r only_titan_req.txt
-#          - name: Run tests
-#            run: |
-#               python tests/push_test/gap_and_fill_test.py
-#               python tests/push_test/IO_test.py
-#               python tests/push_test/qc_test.py
-#               python tests/push_test/breaking_test.py
-#               #python tests/push_test/gui_launch_test.py
-#               python tests/push_test/analysis_test.py
+#---- Run tests -----#
+   run-tests:
+       name: run test scripts
+       needs: packageing
+       runs-on: ${{ matrix.os }}
+       strategy:
+       # You can use PyPy versions in python-version.
+       # For example, pypy-2.7 and pypy-3.8
+         matrix:
+           python-version: ["3.9", "3.10"]
+         # os: [ubuntu-latest, windows-latest]
+           os: [ubuntu-latest]
+           node-version: [12.x]
+       steps:
+         - uses: actions/checkout@v3
+         - name: Set up Python ${{ matrix.python-version }}
+           uses: actions/setup-python@v4
+           with:
+             python-version: ${{ matrix.python-version }}
+           #cache: 'poetry'
+       # You can test your matrix by printing the current Python version
+         - name: Display Python version
+           run: python -c "import sys; print(sys.version)"
+         - name: Download the package build
+           uses: actions/download-artifact@v3
+           with:
+             name: package_build
+         - name: Install the package
+           run: |
+               python3 -m pip install ${{needs.packageing.outputs.pkg_file}}
+         - name: get titanlib requirements
+           uses: actions/download-artifact@v3
+           with:
+             name: titan_requirements
+         - name: install titanlib
+           run: |
+               pip install -r only_titan_req.txt
+         - name: Run tests
+           run: |
+             python tests/push_test/gap_and_fill_test.py
+             python tests/push_test/IO_test.py
+             python tests/push_test/qc_test.py
+             python tests/push_test/breaking_test.py
+             #python tests/push_test/gui_launch_test.py
+             python tests/push_test/analysis_test.py
 
-# #--- Package os installation ---
-#    mac_install_testing:
-#       name: Installation on Mac latest
[…]
