+++ .github/workflows/main_workflow.yml
@@ -274,7 +274,7 @@ jobs:
 # ---- delete artifacts that are not for storage -----
    cleanup_artifacts:
      name: delete artifacts
-     needs: [doctests,run-tests,versiontest,mac_install_testing, windows_install_testing,deploy_doc_dev, deploy_doc]
+     needs: [doctests,run-tests,versiontest,mac_install_testing, windows_install_testing, deploy_doc]
      runs-on: ubuntu-latest
      steps:
      - uses: actions/checkout@v3
