### 2023-11-13 f92a197
Fix the pages deployment of the documentation (#396)
* test build doc

* Update main_workflow.yml

* Update main_workflow.yml build doc

* build doc testing

* Update main_workflow.yml build doc

* with artifact test

* Update main_workflow.yml build doc

* Update main_workflow.yml build doc

* Update main_workflow.yml build doc

* test deploy doc

* test workflow

* Update main_workflow.yml

* ready for syntax fixes

* Update main_workflow.yml

* Update main_workflow.yml

* Update main_workflow.yml

* Update main_workflow.yml

* Update main_workflow.yml

* Update main_workflow.yml

* Update m
[…]
