+++ .github/workflows/main_workflow.yml
@@ -3,93 +3,162 @@ name: Test pushed version
 on: [push]
 
 jobs:
+#--- Build the package and save as artifacts----
+   packageing:
+      name: Build the package
+      runs-on: ubuntu-latest
+      outputs:
+        pkg_file: ${{ steps.base_def.outputs.filename }}
+      steps:
+        - uses: actions/checkout@v3
+        - name: Set up Python39
+          uses: actions/setup-python@v4
+          with:
+            python-version: '3.9'
+            #cache: 'poetry'
+        # You can test your matrix by printing the current Python version
+        - name: Display Python version
+          run: python -c "import sys; print(sys.version)"
+        - name: remove previous builds
+          run: |
+              rm -f dist/metobs_toolkit-*.whl
+              rm -f dist/metobs_toolkit-*.tar.gz
+        - name: Install poetry
+          run: |
+              pip install poetry
+        - name: Install dependencies and build
+          run: |
+              poetry update --without titan,documentation
+              poetry install --without titan,documentation
+              poetry build
+        - name: get distribution file name
+          id: base_def
+          run: |
+              rel_path=$(find dist/ -name 'metobs_toolkit-*.tar.gz' -print -quit)
+              echo "filename=$(basename "$rel_path")" >> $GITHUB_OUTPUT
+
+        - name: save package as artifact
+          uses: actions/upload-artifact@v3
+          with:
+            name: package_build
+            path: dist/*.tar.gz
+        - name: build package with all extras
+          run: |
+              poetry update
+              poetry install --all-extras
+        - name: export documentation requirement group
+          run: |
+              poetry export -f requirements.txt --without-hashes --only documentation -o only_doc_req.txt
+        - name: convert doc requirments to artifact
+          uses: actions/upload-artifact@v3
+          with:
+            name: documentation_requirements
+            path: only_doc_req.txt
+        - name: export titan requirement group
+          run: |
+              poetry export -f requirements.txt --without-hashes --only titan -o only_titan_req.txt
+        - name: convert titan requirements to artifact
+          uses: actions/upload-artifact@v3
+          with:
+            name: titan_requirements
+            path: only_titan_req.txt
+
+
 #---- Run tests -----#
    run-tests:
-     name: run test scripts
-     runs-on: ${{ matrix.os }}
-     strategy:
+       name: run test scripts
+       needs: packageing
+       runs-on: ${{ matrix.os }}
+       strategy:
        # You can use PyPy versions in python-version.
        # For example, pypy-2.7 and pypy-3.8
-       matrix:
-         python-version: ["3.9", "3.10"]
+         matrix:
+           python-version: ["3.9", "3.10"]
          # os: [ubuntu-latest, windows-latest]
-         os: [ubuntu-latest]
-         node-version: [12.x]
-
-     steps:
-       - uses: actions/checkout@v3
-       - name: Set up Python ${{ matrix.python-version }}
-         uses: actions/setup-python@v4
-         with:
-           python-version: ${{ matrix.python-version }}
+           os: [ubuntu-latest]
+           node-version: [12.x]
+       steps:
+         - uses: actions/checkout@v3
+         - name: Set up Python ${{ matrix.python-version }}
+           uses: actions/setup-python@v4
+           with:
+             python-version: ${{ matrix.python-version }}
            #cache: 'poetry'
        # You can test your matrix by printing the current Python version
-       - name: Display Python version
-         run: python -c "import sys; print(sys.version)"
-       - name: Install poetry
-         run: |
-             pip install poetry
-
-       - name: Install dependencies and build
-         run: |
-             poetry update
-             poetry install
-       - name: Run tests
-         run: |
-             poetry run python tests/push_test/gap_and_fill_test.py
[…]
