### 2023-11-14 #5916 @vergauwenthomas
> The fix broke the docs, I believe. Waiting for it again... [vergauwenthomas/MetObs_toolkit#394](https://github.com/vergauwenthomas/MetObs_toolkit/issues/394)

The [docs](https://vergauwenthomas.github.io/MetObs_toolkit/) are back online.
