+++ CONTRIBUTING.md
@@ -1,0 +1,177 @@
+============
+Contributing
+============
+
+All contributions are welcome, and they are greatly appreciated! Every little bit helps, and credit will always be given.
+
+You can contribute in many ways:
+
+Types of Contributions
+----------------------
+
+Feature Requests
+~~~~~~~~~~~~~~~~
+
+If you are interested in a new Feature, you can add it as an [issue](https://github.com/vergauwenthomas/MetObs_toolkit/issues)
+with the *enhanced* label.
+
+In the issues describe clearly new feature or functionallity you want to see implemented in the toolkit. If the request is specific for your application/data, make shure to
+add a sample of your data, a template and a description.
+
+Assign yourself to the issue if you want to (help) implement the new request.
+
+
+Report Bugs
+~~~~~~~~~~~
+
+Report bugs at the as a new [issue](https://github.com/vergauwenthomas/MetObs_toolkit/issues) with the *enhanced* label.
+
+If you are reporting a bug, please include:
+
+* Your operating system name and version.
+* The version the MetObs-toolkit.
+* Any details about your local setup that might be helpful in troubleshooting.
+* Detailed steps to reproduce the bug.
+* If possible, provide a pickled version of your dataset (use [.save_dataset()](https://vergauwenthomas.github.io/MetObs_toolkit/_autosummary/metobs_toolkit.dataset.Dataset.html#metobs_toolkit.dataset.Dataset.save_dataset)). Limit the size of the dataset as much as possible.
+
+Fix Bugs
+~~~~~~~~
+
+Look through the GitHub issues for bugs. Anything tagged with "bug" and "invalid" is open to whoever wants to implement it.
+If you found yourself not so familiar with python you can start by filtering to the "easy"-labels.
+
+The "tricky"-label indicates that this issue might affect multiple modules of the toolkit, the data structures or is technical more challenging. Contact @vergauwenthomas to discuss a plan-of-attack in advance.
+
+
+Implement Features
+~~~~~~~~~~~~~~~~~~
+
+Look through the GitHub issues for features. Anything tagged with "enhancement" and "feature" is open to whoever wants to implement it.
+
+Write Documentation
+~~~~~~~~~~~~~~~~~~~
+
+The MetObs-toolkit could always use more documentation or spell checkers. Use the "documentation"-flag to indicate that your issue is documentation related.
+
+Submit Feedback
+~~~~~~~~~~~~~~~
+
+The best way to send feedback is to file an issue at https://github.com/CSHS-CWRA/ravenpy/issues.
+
+If you are proposing a feature:
+
+* Explain in detail how it would work.
+* Keep the scope as narrow as possible, to make it easier to implement.
+* Remember that this is a volunteer-driven project, and that contributions
+  are welcome :)
+
+Get Started!
+------------
+
+Ready to contribute? Here's how to set up `ravenpy` for local development.
+
+1. Fork the `ravenpy` repo on GitHub.
+2. Clone your fork locally::
+
+    $ git clone git@github.com:your_name_here/ravenpy.git
+
+3. Install your local copy into a virtualenv. Assuming you have virtualenvwrapper installed, this is how you set up your fork for local development::
+
+    $ mkvirtualenv ravenpy
+    $ cd ravenpy/
+    $ pip install -e ".[dev]"
+
+4. To ensure a consistent style, please install the pre-commit hooks to your repo::
+
+    $ pre-commit install
+
+   Special style and formatting checks will be run when you commit your changes. You
+   can always run the hooks on their own with::
+
+    $ pre-commit run -a
+
+5. Create a branch for local development::
+
+    $ git checkout -b name-of-your-bugfix-or-feature
+
+   Now you can make your changes locally.
+
+6. When you're done making changes, check that your changes pass flake8, black, and the
[…]
