+++ CONTRIBUTING.md
@@ -1,19 +1,17 @@
-============
-Contributing
-============
+
+# Contributing
+
 
 All contributions are welcome, and they are greatly appreciated! Every little bit helps, and credit will always be given.
 
 You can contribute in many ways:
 
-Types of Contributions
-----------------------
+## Types of Contributions
 
-Feature Requests
-~~~~~~~~~~~~~~~~
+### Feature Requests
 
 If you are interested in a new Feature, you can add it as an [issue](https://github.com/vergauwenthomas/MetObs_toolkit/issues)
-with the *enhanced* label.
+with the https://github.com/vergauwenthomas/MetObs_toolkit/labels/enhancement label.
 
 In the issues describe clearly new feature or functionallity you want to see implemented in the toolkit. If the request is specific for your application/data, make shure to
 add a sample of your data, a template and a description.
@@ -21,10 +19,8 @@ add a sample of your data, a template and a description.
 Assign yourself to the issue if you want to (help) implement the new request.
 
 
-Report Bugs
-~~~~~~~~~~~
-
-Report bugs at the as a new [issue](https://github.com/vergauwenthomas/MetObs_toolkit/issues) with the *enhanced* label.
+### Report Bugs
+Report bugs at the as a new [issue](https://github.com/vergauwenthomas/MetObs_toolkit/issues) with the https://github.com/vergauwenthomas/MetObs_toolkit/labels/bug label.
 
 If you are reporting a bug, please include:
 
@@ -34,144 +30,92 @@ If you are reporting a bug, please include:
 * Detailed steps to reproduce the bug.
 * If possible, provide a pickled version of your dataset (use [.save_dataset()](https://vergauwenthomas.github.io/MetObs_toolkit/_autosummary/metobs_toolkit.dataset.Dataset.html#metobs_toolkit.dataset.Dataset.save_dataset)). Limit the size of the dataset as much as possible.
 
-Fix Bugs
-~~~~~~~~
+### Fix Bugs
 
-Look through the GitHub issues for bugs. Anything tagged with "bug" and "invalid" is open to whoever wants to implement it.
-If you found yourself not so familiar with python you can start by filtering to the "easy"-labels.
+Look through the GitHub issues for bugs. Anything tagged with https://github.com/vergauwenthomas/MetObs_toolkit/labels/bug and https://github.com/vergauwenthomas/MetObs_toolkit/labels/invalid is open to whoever wants to implement it.
+If you find yourself not so familiar with Python you can start by filtering to the https://github.com/vergauwenthomas/MetObs_toolkit/labels/easy and https://github.com/vergauwenthomas/MetObs_toolkit/labels/good_first_issue labels.
 
-The "tricky"-label indicates that this issue might affect multiple modules of the toolkit, the data structures or is technical more challenging. Contact @vergauwenthomas to discuss a plan-of-attack in advance.
-
-
-Implement Features
-~~~~~~~~~~~~~~~~~~
-
-Look through the GitHub issues for features. Anything tagged with "enhancement" and "feature" is open to whoever wants to implement it.
-
-Write Documentation
-~~~~~~~~~~~~~~~~~~~
-
-The MetObs-toolkit could always use more documentation or spell checkers. Use the "documentation"-flag to indicate that your issue is documentation related.
-
-Submit Feedback
-~~~~~~~~~~~~~~~
-
-The best way to send feedback is to file an issue at https://github.com/CSHS-CWRA/ravenpy/issues.
-
-If you are proposing a feature:
-
-* Explain in detail how it would work.
-* Keep the scope as narrow as possible, to make it easier to implement.
-* Remember that this is a volunteer-driven project, and that contributions
-  are welcome :)
-
[…]
