### 2023-11-14 b17bf78
remove the rm -f poetry lock because lock is ignored
