+++ CONTRIBUTING.md
@@ -1,10 +1,10 @@
 
 # Contributing
 
-
 All contributions are welcome, and they are greatly appreciated! Every little bit helps, and credit will always be given.
 
 You can contribute in many ways:
+
 
 ## Types of Contributions
 
@@ -13,11 +13,10 @@ You can contribute in many ways:
 If you are interested in a new Feature, you can add it as an [issue](https://github.com/vergauwenthomas/MetObs_toolkit/issues)
 with the https://github.com/vergauwenthomas/MetObs_toolkit/labels/enhancement label.
 
-In the issues describe clearly new feature or functionallity you want to see implemented in the toolkit. If the request is specific for your application/data, make shure to
-add a sample of your data, a template and a description.
+In the issues describe clearly the new feature or functionality you want to see implemented in the toolkit. If the request is specific to your application/data, make sure to
+add a sample of your pickled Dataset (use [.save_dataset()](https://vergauwenthomas.github.io/MetObs_toolkit/_autosummary/metobs_toolkit.dataset.Dataset.html#metobs_toolkit.dataset.Dataset.save_dataset)).
 
-Assign yourself to the issue if you want to (help) implement the new request.
-
+Assign yourself to the issue if you want to (help) implement the new request. All help is much appreciated!
 
 ### Report Bugs
 Report bugs at the as a new [issue](https://github.com/vergauwenthomas/MetObs_toolkit/issues) with the https://github.com/vergauwenthomas/MetObs_toolkit/labels/bug label.
@@ -25,7 +24,7 @@ Report bugs at the as a new [issue](https://github.com/vergauwenthomas/MetObs_to
 If you are reporting a bug, please include:
 
 * Your operating system name and version.
-* The version the MetObs-toolkit.
+* The version of the MetObs-toolkit (use `metobs_toolkit.__version__`).
 * Any details about your local setup that might be helpful in troubleshooting.
 * Detailed steps to reproduce the bug.
 * If possible, provide a pickled version of your dataset (use [.save_dataset()](https://vergauwenthomas.github.io/MetObs_toolkit/_autosummary/metobs_toolkit.dataset.Dataset.html#metobs_toolkit.dataset.Dataset.save_dataset)). Limit the size of the dataset as much as possible.
@@ -35,8 +34,7 @@ If you are reporting a bug, please include:
 Look through the GitHub issues for bugs. Anything tagged with https://github.com/vergauwenthomas/MetObs_toolkit/labels/bug and https://github.com/vergauwenthomas/MetObs_toolkit/labels/invalid is open to whoever wants to implement it.
 If you find yourself not so familiar with Python you can start by filtering to the https://github.com/vergauwenthomas/MetObs_toolkit/labels/easy and https://github.com/vergauwenthomas/MetObs_toolkit/labels/good_first_issue labels.
 
-The https://github.com/vergauwenthomas/MetObs_toolkit/labels/tricky label indicates that this issue might affect multiple modules of the toolkit, the data structures or is technical more challenging. Contact @vergauwenthomas to discuss a plan-of-attack in advance.
-
+The https://github.com/vergauwenthomas/MetObs_toolkit/labels/tricky label indicates that this issue might affect multiple modules of the toolkit, the data structures or is technical more challenging. Contact @vergauwenthomas to discuss a *plan-of-attack* in advance.
 
 ### Implement Features
 
@@ -49,10 +47,12 @@ The MetObs-toolkit could always use more documentation or spell checkers. Use th
 Any form of feedback is much appreciated. The best way to send feedback is to file [an issue](https://github.com/vergauwenthomas/MetObs_toolkit/issues). If you cannot find a suitable label, you do not have to specify one.
 
 
+
 ## Get Started
[…]
