+++ CONTRIBUTING.md
@@ -1,4 +1,3 @@
-
 # Contributing
 
 All contributions are welcome, and they are greatly appreciated! Every little bit helps, and credit will always be given.
