+++ docs/index.rst
@@ -11,7 +11,8 @@ Welcome to MetObs-Toolkit's documentation!
    gee_authentication
    special_topics
    gui
-
+   readme_link.md
+   contributing_link.md
 
 MetObs toolkit Documentation
 -----------------------------
@@ -19,6 +20,7 @@ MetObs toolkit Documentation
    :maxdepth: 2
 
    MetObs_documentation
+
 
 Metobs for developpers
 -----------------------------
