### 2023-11-16 8489ef1
add contributing and logo to the docs
