+++ docs/conf.py
@@ -136,3 +136,11 @@ html_theme_options = {
 # so a file named "default.css" will overwrite the builtin "default.css".
 
 html_static_path = ["_static"]
+
+html_logo = "logo_small.svg"
+html_theme_options = {
+    'logo_only': True,
+    'display_version': False,
+}
+
+
