### 2023-11-16 cbe4135
Add a Contribution file  (#398)
* use RavenPy as a template

* specify pandoc minor version + remove poetry.lock and add to gitignore

* Update CONTRIBUTING.md

* Update CONTRIBUTING.md

* remove the rm -f poetry lock because lock is ignored

* Update README.md

* Update CONTRIBUTING.md

* Update CONTRIBUTING.md

* include readme and contributing to the docs, test if labels are rendered correctly on github side

* trigger build doc

* Update CONTRIBUTING.md

* Update README.md

* Update intro.rst

* Update CONTRIBUTING.md

* add contributing and logo to the docs

* minor codereview
