+++ pyproject.toml
@@ -36,7 +36,7 @@ nbsphinx = '^0.9'
 sphinx-copybutton = '^0.5.1'
 sphinx-rtd-theme = '^1.3.0'
 myst_parser = '^2.0.0'
-pandoc = '^2' #Check on PyPi (not in local conda env) + with system wide install
+pandoc = '^2.1' #Check on PyPi (not in local conda env) + with system wide install
 
 
 [tool.poetry.group.dev.dependencies]
