### 2023-11-16 vergauwenthomas/MetObs_toolkit#385 @vergauwenthomas
A CONTRIBUTING file is added to the repo + is rendered in the documentation.
