### 2023-11-20 ab0075b
attempt to include paper and build using workflow
