+++ docs/JOSS-paper/paper.md
@@ -1,0 +1,106 @@
+---
+title: 'MetObs - a Python toolkit for using non-traditional meteorological observations'
+tags:
+  - Python
+  - Meteorology
+  - Urban climate
+  - Observations
+authors:
+  - name: Thomas Vergauwen
+    orcid: 0000-0003-2899-9218
+    equal-contrib: false
+    affiliation: "1, 2" # (Multiple affiliations must be quoted)
+  - name: Michiel Vieijra
+    orcid: 0000-0003-0817-2846
+    equal-contrib: False # (This is how you can denote equal contributions between multiple authors)
+    affiliation: 1
+  - name: Andrei Covaci
+    orcid: 0000-0001-5147-2460
+    equal-contrib: False # (This is how you can denote equal contributions between multiple authors)
+    affiliation: 3
+  - name: Amber Jacobs
+    orcid: 0000-0002-4628-3988
+    equal-contrib: False # (This is how you can denote equal contributions between multiple authors)
+    affiliation: 2
+  - name: Sara Top
+    orcid: 0000-0003-1281-790X
+    equal-contrib: False # (This is how you can denote equal contributions between multiple authors)
+    affiliation: 2
+  - name: Wout Dewettinck
+    orcid: 0000-0002-0728-5331
+    equal-contrib: False # (This is how you can denote equal contributions between multiple authors)
+    affiliation: 2
+  - name: Kobe Vandelanotte
+    orcid: 0009-0001-1252-7315
+    equal-contrib: False # (This is how you can denote equal contributions between multiple authors)
+    affiliation: "1, 2"
+  - name: Ian Hellebosch
+    orcid: 0000-0003-0150-529X
+    equal-contrib: False # (This is how you can denote equal contributions between multiple authors)
+    affiliation: "2, 4"
+  - name: Steven Caluwaerts
+    orcid: 0000-0001-7456-3891
+    equal-contrib: False # (This is how you can denote equal contributions between multiple authors)
+    affiliation: "1, 2"
+affiliations:
+ - name: Royal Meteorological Institute of Belgium, Brussels, Belgium
+   index: 1
+ - name: Ghent University department of Physics and Astronomy, Ghent, Belgium
+   index: 2
+ - name: Vrije Universiteit Brussel (VUB), Brussels, Belgium
+   index: 3
+ - name: VITO, Mol, Belgium
+   index: 4
+date: 24 August 2023
+bibliography: paper.bib
+
+---
+
+
+# Summary
+In-situ meteorological observations are highly important for weather and climate research. The evolution towards more affordable sensor technology and data communication has resulted in the emergence of novel meteorological networks alongside the traditional high-quality measurement networks of meteorological institutions. Examples include urban measurement networks intended to study the impact of cities [@mocca] and networks consisting of devices of weather enthusiasts [@crowdsourcing_status]. However, exploiting the data of such non-traditional networks comes with significant challenges [@crowdsourcing]. Firstly, sensors and data communication protocols are usually low-cost, and this in general results in an increase of measurement errors, biases and data gaps. Secondly, data storage formats and temporal measurement frequencies are often not consistent or compatible. Finally, metadata, such as land use around a station and elevation, are not easily accessible or documented.
+
+The MetObs-toolkit is a Python package developed to address these issues and facilitate the use of non-traditional observations. The package provides automated quality control (QC) techniques to identify and flag erroneous observations, and includes methods to fill data gaps. Additionally, the package offers tools for analyzing the data, e.g. linkage with popular land-use datasets [@worldcover; @lcz_map] is included such that microclimate effects can be investigated with the MetObs-toolkit.
+
+
+# Statement of need
+The primary objective of the MetObs-toolkit is  to enable scientists to process meteorological observations into datasets ready for analysis. The data cleaning process involves three steps: 
+
[…]
