+++ docs/JOSS-paper/paper.bib
@@ -1,0 +1,218 @@
+@Article{titan2020,
+    AUTHOR = {B{\aa}serud, L. and Lussana, C. and Nipen, T. N. and Seierstad, I. A. and Oram, L. and Aspelien, T.},
+    TITLE = {TITAN automatic spatial quality control of meteorological in-situ observations},
+    JOURNAL = {Advances in Science and Research},
+    VOLUME = {17},
+    YEAR = {2020},
+    PAGES = {153--163},
+    URL = {https://asr.copernicus.org/articles/17/153/2020/},
+    DOI = {10.5194/asr-17-153-2020}
+    }
+
+@article{mocca,
+abstract     = {As urban environments have a specific climate that poses extra challenges (e.g. increased heat stress during heat waves), gaining detailed insight into the urban climate is important. This paper presents the high-accuracy MOCCA (MOnitoring the City's Climate and Atmosphere) network, which is monitoring the urban climate of the city of Ghent since July 2016. The study illustrates the complementarity between modelling and observing the urban climate. Two different modelling approaches are used: 1 km resolution runs of the SURFEX land surface model and 100 m resolution runs of the computationally cheaper UrbClim boundary layer model. On the one hand, urban models are able to simulate the spatial variability of the urban climate. As such, these models serve as a tool to help deciding on the locations of the measurement stations. On the other hand, the MOCCA observations are used to validate the high-resolution urban model experiments for the summer (July-August-September) of 2016. Our results demonstrate that the models capture the nighttime intra-urban temperature differences, but they are not able to reproduce the observed daytime temperature differences which are determined by the micro-scale environment.},
+articleno    = {100565},
+author       = {Caluwaerts, Steven and Hamdi, Rafiq and Top, Sara and Lauwaet, Dirk and Berckmans, Julie and Degrauwe, Daan and Dejonghe, Herwig and De Ridder, Koen and De Troch, Rozemien and Duchêne, Francois and Maiheu, Bino and Van Ginderachter, Michiel and Verdonck, Marie-Leen and Vergauwen, Thomas and Wauters, Guy and Termonia, Piet},
+issn         = {2212-0955},
+journal      = {Urban Climate},
+keywords     = {High-resolution urban modelling,Urban heat island,Urban monitoring network,ALARO-0,SURFEX,UrbClim,HEAT-ISLAND,MODEL,CONVECTION,SCHEME,ENERGY,LAND,ADVECTION,EVOLUTION,BRUSSELS,SYSTEM},
+language     = {eng},
+pages        = {19},
+title        = {The urban climate of Ghent, Belgium : a case study combining a high-accuracy monitoring network with numerical simulations},
+url          = {http://dx.doi.org/10.1016/j.uclim.2019.100565},
+doi          = {10.1016/j.uclim.2019.100565},
+volume       = {31},
+year         = {2020}
+}
+
+
+@ARTICLE{CrowdQC,
+AUTHOR={Fenner, Daniel and Bechtel, Benjamin and Demuzere, Matthias and Kittner, Jonas and Meier, Fred},
+TITLE={CrowdQC+—A Quality-Control for Crowdsourced Air-Temperature Observations Enabling World-Wide Urban Climate Applications},
+JOURNAL={Frontiers in Environmental Science},
+VOLUME={9},
+YEAR={2021},
+URL={https://www.frontiersin.org/articles/10.3389/fenvs.2021.720747},
+DOI={10.3389/fenvs.2021.720747},
+ISSN={2296-665X},
[…]
