+++ .github/workflows/main_workflow.yml
@@ -270,6 +270,36 @@ jobs:
             publish_dir: _build/
             force_orphan: true
 
+# ---- build the JOSS - paper -------
+   paper:
+     if: |
+            github.ref == 'refs/heads/dev' ||
+            github.ref == 'refs/head/main' ||
+            contains(github.event.head_commit.message, 'build paper')
+     runs-on: ubuntu-latest
+     name: Paper Draft
+     steps:
+       - name: Checkout
+         uses: actions/checkout@v3
+       - name: Build draft PDF
+         uses: openjournals/openjournals-draft-action@master
+         with:
+           journal: joss
+           # This should be the path to the paper within your repo.
+           paper-path: docs/JOSS-paper/paper.md
+       - name: Upload
+         uses: actions/upload-artifact@v1
+         with:
+           name: paper
+           # This is the output path where Pandoc will write the compiled
+           # PDF. Note, this should be the same directory as the input
+           # paper.md
+           path: docs/paper/paper.pdf
+       - name: save pdf to repo
+         uses: stefanzweifel/git-auto-commit-action@v4
+         with:
+           commit_message: Saved new PDF of paper
+
 
 # ---- delete artifacts that are not for storage -----
    cleanup_artifacts:
