+++ .github/workflows/main_workflow.yml
@@ -294,7 +294,7 @@ jobs:
            # This is the output path where Pandoc will write the compiled
            # PDF. Note, this should be the same directory as the input
            # paper.md
-           path: docs/paper/paper.pdf
+           path: docs/JOSS-paper/paper.pdf
        - name: save pdf to repo
          uses: stefanzweifel/git-auto-commit-action@v4
          with:
