+++ docs/paper/paper.bib
@@ -119,6 +119,7 @@ year = {2020}
   version      = {v2.0.0},
   doi          = {10.5281/zenodo.7794821},
   url          = {https://doi.org/10.5281/zenodo.7794821}
+  author-in-text ={The pandas team}
 }
 
 @InProceedings{scipy,
