+++ docs/paper/paper.bib
@@ -118,8 +118,8 @@ year = {2020}
   publisher    = {Zenodo},
   version      = {v2.0.0},
   doi          = {10.5281/zenodo.7794821},
-  url          = {https://doi.org/10.5281/zenodo.7794821}
-  author-in-text ={The pandas team}
+  url          = {https://doi.org/10.5281/zenodo.7794821},
+  author-in-text = {The pandas team}
 }
 
 @InProceedings{scipy,
