### 2023-11-26 #5916 @ashwinvis
I tried to verify the functionality of the software today and I realize that a key attractive feature of this software which is gap-filling relies on Google Earth Engine to access ERA5 reanalysis dataset, a choice which severely restricts the usability (and to some extent the openness) of the software.

While the use of [proprietary software IDEs](https://joss.readthedocs.io/en/latest/submitting.html#submissions-using-proprietary-languages-development-environments) have been mentioned in the JOSS docs, I ask what is the stance for reliance on proprietary network services, @hugoledoux ?

Also I should ask @vergauwenthomas why did you opt for Google Earth Engine for a dataset easily accessible via Copernicus? Would it be easy to add support for Copernicus's Python API `cdsapi`?

https://climate.copernicus.eu/climate-reanalysis
https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-single-levels?tab=form
