### 2023-11-26 #5916 @ashwinvis
Other comments:

- Reproducibility: I believe the results shown in Figure 1 is based on the Examples section in the docs. The author should point out that.
- References: Fix `Caluwaerts et al.` title `The urban climate of ghent, belgium...` and the citation for Pandas.
