### 2023-11-27 fe7fff3
test citeproc format (workarround)
