+++ docs/paper/paper.bib
@@ -110,7 +110,7 @@ year = {2020}
 }
 
 @software{pandas,
-  author       = {The pandas development team},
+  author       = {The development pandas-team},
   title        = {pandas-dev/pandas: Pandas},
   month        = apr,
   year         = 2023,
