+++ docs/paper/paper.bib
@@ -110,7 +110,7 @@ year = {2020}
 }
 
 @software{pandas,
-  author       = {The development pandas-team},
+  author       = {pandas-dev-team},
   title        = {pandas-dev/pandas: Pandas},
   month        = apr,
   year         = 2023,
