### 2023-11-27 6d57de8
Add script for the creation of the figures of the paper.
