+++ docs/paper/paper_figures.py
@@ -1,0 +1,238 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Script to create the figures used in the JOSS publication of the Metobs-Toolkit
+
+@author: thoverga
+"""
+
+import logging
+import math
+import os
+import sys
+import time
+from pathlib import Path
+
+import pandas as pd
+
+lib_folder = Path(__file__).resolve().parents[2]
+sys.path.insert(0,str(lib_folder))
+
+# add testdata paths
+from tests.push_test.test_data_paths import testdata
+
+tmp_pickle=os.path.join(lib_folder, 'development', 'tmp', 'dev_pickle.pkl')
+
+import metobs_toolkit
+
+# handle logs to spyder console
+toolkit_logger = metobs_toolkit.loggers
+toolkit_logger.addHandler(logging.StreamHandler())
+
+
+# =============================================================================
+# Creation of the Dataset
+# =============================================================================
+
+datadf = pd.read_csv(metobs_toolkit.demo_datafile, sep=';')
+metadf = pd.read_csv(metobs_toolkit.demo_metadatafile, sep=',')
+
+# Subset to regio ghent
+ghent_stations = [ 'vlinder24', 'vlinder25', 'vlinder05', 'vlinder27',
+                  'vlinder02', 'vlinder01', 'vlinder28']
+
+
+datadf = datadf[datadf['Vlinder'].isin(ghent_stations)]
+metadf = metadf[metadf['Vlinder'].isin(ghent_stations)]
+
+# subset period
+datadf['dummy_dt'] = datadf['Datum'] + datadf['Tijd (UTC)']
+datadf['dummy_dt'] = pd.to_datetime(datadf['dummy_dt'], format='%Y-%m-%d%H:%M:%S')
+
+#Subset to period
+from datetime import datetime
+startdt = datetime(2022, 9, 1)
+enddt = datetime(2022, 9, 10)
+datadf = datadf[(datadf['dummy_dt'] >= startdt) & (datadf['dummy_dt'] <= enddt)]
+datadf = datadf.drop(columns=['dummy_dt'])
+
+# Inducing outliers as demo
+datadf = datadf.drop(index=datadf.iloc[180:200, :].index.tolist())
+
+
+# save in paper folder
+folder = os.path.join(lib_folder, 'tests', 'test_data', 'paper_dataset')
+datadf.to_csv(os.path.join(folder, 'datafile.csv'))
+metadf.to_csv(os.path.join(folder, 'metadatafile.csv'))
+
+#Importing raw data
+use_dataset = 'paper_dataset'
+dataset = metobs_toolkit.Dataset()
+dataset.update_settings(output_folder=folder,
+                        input_data_file=testdata[use_dataset]['datafile'],
+                        input_metadata_file=testdata[use_dataset]['metadatafile'],
+                        template_file=testdata[use_dataset]['template'],
+                        )
+
+dataset.import_data_from_file(**testdata[use_dataset]['kwargs'])
+
+#%% Make raw data timeseries plot
+import matplotlib.pyplot as plt
+
+
+# =============================================================================
+# Set styling and formats for the plots
+# =============================================================================
+
+
+#Shape figure and resolution
+def shape_timeseries_fig_for_paper(ax, filename):
+
+    fig = ax.get_figure()
+    fig.set_size_inches(12, 5, forward=True)
+    fig.set_dpi(200)
+
+    # fig.tight_layout()
+    figpath = os.path.join(folder, filename)
+    plt.savefig(fname=figpath)
+
+
+# change color for printing (avoid yellow!)
+dataset.settings.app['plot_settings']['color_mapper']['gross_value'] = "#fc0303"
+
+
+# =============================================================================
[…]
