### 2023-11-27 bf79e15
add the paper_figure notebook to the documentation.
