### 2023-11-27 #5916 @vergauwenthomas
> Also I should ask @vergauwenthomas why did you opt for Google Earth Engine for a dataset easily accessible via Copernicus? Would it be easy to add support for Copernicus's Python API `cdsapi`?
>

@ashwinvis, this is indeed a relevant topic. At the beginning of this project, I considered using the Copernicus API, but the Google Earth Engine integration seems a better choice to me because of:

- The presence of the global LCZ map and the worldcover (10m-res) in the GEE.
- The transfer speed of GEE is better than Copernicus API (= colleagues advice @saratop-ugent).
- The availability of semi-real-time products at the GEE has potential weather-related features.

Since the use of the GEE is limited to 1) static geospatial datasets and 2) time-series, I have never reached the limit of the free computational-units.

If you feel like including the Copernicus API is a valuable addition, I can open a discussion on this.
