### 2023-11-27 2be2eaf
removed the 'expected features' from the paper
