### 2023-11-27 #5916 @vergauwenthomas
FYI, I have created a new label `JOSS publication blocking.` You can add this to issues if you found them blocking your approval.

I will make a priority of these issues and try to solve them asap.
