### 2023-11-27 #5916 @ashwinvis
Personally, I am concerned the use of Google Earth Engine would discourage the future users of MetObs toolkit from contributing data to weather forecast agencies and thus completely undermining the spirit of open source and open data. See the terms of GEE

> For all other government agencies, Earth Engine is free of charge if it is used for Government Research Activities. For your use to qualify as Government Research Activities all of the following requirements must be met:

>    The project work is question-driven.

>    The project work has a stated start & (estimated) end date.

>    The project work has a limit on the amount of compute and data, such that it isn't “operational,” or repeated periodically over time. If Government Research Activities are repeated periodically over time, that would then be considered an operational use case. As an example, a government agency’s use of Earth Engine to run the same algorithm on a monthly or annual basis would not qualify as Government Research Activities.

>    The project work is meant to result in scholarly output, e.g. manuscripts in a peer reviewed scholarly journal, white papers, or other documentation about the scientific methods. Findings should be made public: e.g. code libraries or datasets (results and training/validation) shared with the scientific community.

However it is nevertheless true that the source code remains open, so technically the project may _qualify_ . But, I want to hear from the JOSS editors what their policy is in the scope of the paper.
