### 2023-11-28 #5916 @hugoledoux
This is a valid point @ashwinvis and I don't want to answer without asking other editors. I'll get back to you soon (and we should indeed update our docs for such issues).
