### 2023-11-28 fe91582
forced the Mocca paper to use capital letters for Gent and Belgiu
