+++ docs/paper/paper.bib
@@ -18,7 +18,7 @@ journal      = {Urban Climate},
 keywords     = {High-resolution urban modelling,Urban heat island,Urban monitoring network,ALARO-0,SURFEX,UrbClim,HEAT-ISLAND,MODEL,CONVECTION,SCHEME,ENERGY,LAND,ADVECTION,EVOLUTION,BRUSSELS,SYSTEM},
 language     = {eng},
 pages        = {19},
-title        = {The urban climate of Ghent, Belgium : a case study combining a high-accuracy monitoring network with numerical simulations},
+title        = {The urban climate of {G}hent, {B}elgium : a case study combining a high-accuracy monitoring network with numerical simulations},
 url          = {http://dx.doi.org/10.1016/j.uclim.2019.100565},
 doi          = {10.1016/j.uclim.2019.100565},
 volume       = {31},
