+++ docs/paper/paper.bib
@@ -110,7 +110,7 @@ year = {2020}
 }
 
 @software{pandas,
-  author       = {pandas-dev-team},
+  author       = {{the pandas dev team}},
   title        = {pandas-dev/pandas: Pandas},
   month        = apr,
   year         = 2023,
