+++ docs/paper/paper.bib
@@ -110,7 +110,7 @@ year = {2020}
 }
 
 @software{pandas,
-  author       = {{the pandas dev team}},
+  author       = {{The pandas development team}},
   title        = {pandas-dev/pandas: Pandas},
   month        = apr,
   year         = 2023,
@@ -119,7 +119,7 @@ year = {2020}
   version      = {v2.0.0},
   doi          = {10.5281/zenodo.7794821},
   url          = {https://doi.org/10.5281/zenodo.7794821},
-  author-in-text = {The pandas team}
+  author-in-text = {{pandas team}}
 }
 
 @InProceedings{scipy,
