### 2023-12-08 47f97b2
Add the ONLY-metadata functionallity
