### 2023-12-12 #5916 @hugoledoux
@ashwinvis we don't have anything against the use of GEE. We will update the docs to reflect this in the future, thanks for pointing this out.

That being said, I agree that if an open-data alternative is feasible, it would be nicer to have that option in the software, but this is not something we will enforce.
