### 2023-12-12 vergauwenthomas/MetObs_toolkit#387 @Zeitsperre
@vergauwenthomas The latest draft of the paper addresses many of the issues I raised (and flows much better). There are a handful of small grammatical errors that I can fix if you'd like (let me know), but I consider this issue addressed!
