### 2023-12-13 c680993
Paper (#400)
* paper draft creation

* Update paper.bib

* Update paper.bib

* Update paper.bib

* Update paper.md

* add overview figure

* pandoc action to make pdf

* Update paper.md

* Update paper.md

* try new ref method

* easier naming of references

* Update paper.md 

update references

* fixe some bad structured bib items (prob. more to come)

* Update paper.md

* Update paper.md

try nocite all

* Update paper.md

add hyperlink to fairness

* Update paper.md

reposition figure

* Update paper.md

update orcid Andrei and Kobe

* restructure location of paper draft under docs

* Upda
[…]
