### 2023-12-13 vergauwenthomas/MetObs_toolkit#387 @vergauwenthomas
@Zeitsperre thanks!  If you could pinpoint the grammatical errors that would be great, but only if you have time for it. (If you do, and prefer to fix the errors in the paper.md you are free to do so.)

btw #400 is merged to master.
