### 2023-12-25 vergauwenthomas/MetObs_toolkit#408 @ashwinvis
Open data alternative to Google Earth Engine
As mentioned [here](https://github.com/openjournals/joss-reviews/issues/5916#issuecomment-1828426214) Google Earth Engine (GEE) can be replaced with fallback data sources like Copernicus which do not come with restrictive terms of use.

From the perspective of MetObs_toolkit users, this would be really good to have, **and** essential  who would be willing to contribute the weather data to meteorological agencies (which is not allowed by GEE).

**NOTE:** this is not a blocker for the JOSS paper as clarified by the editor.
