### 2024-01-03 73ec146
Gee example fix (#409)
* fix gee example

* Saved new PDF of paper

---------

Co-authored-by: vergauwenthomas <vergauwenthomas@users.noreply.github.com>
