### 2024-01-03 6a40ab8
should typo (#411)
* fix docstring in sync_observations

* should typo

* Saved new PDF of paper

---------

Co-authored-by: vergauwenthomas <vergauwenthomas@users.noreply.github.com>
