### 2024-01-03 cd91cfa
Update template_mapping.rst
Updated standard name for 'wind speed' i.e, wind_speed in docs.
