### 2024-01-09 #5916 @vergauwenthomas
Hi @Zeitsperre, @ashwinvis, @hugoledoux

I want to proceed with the reviewing process since some scientific publication drafts want to add a reference to the MetObs toolkit.  If I am not mistaken, all the blocking issues are resolved.

_(Please correct me if i am wrong, or create a new issue and i try to fix them ASAP.)_
