### 2024-01-25 3a027a6
new example on obstypes and added in the documentation
