+++ docs/examples/index.rst
@@ -10,5 +10,6 @@
    qc_example.ipynb
    filling_example.ipynb
    analysis_example.ipynb
+   using_obstypes.ipynb
 
 
