### 2024-01-25 a7476da
code review fixes + remove Stamen tiles
