+++ metobs_toolkit/dataset.py
@@ -967,7 +967,6 @@ class Dataset:
             categorical_fields=self.settings.app["categorical_fields"],
             static_fields=self.settings.app["static_fields"],
             display_name_mapper=self.settings.app["display_name_mapper"],
-            world_boundaries_map=self.settings.app["world_boundary_map"],
             data_template=self.data_template,
             boundbox=boundbox
         )
