### 2024-01-25 fdadffd
Without worldboundaries data (#423)
* replace world boundaries files by using cartopy package

* fix test

* Saved new PDF of paper

---------

Co-authored-by: vergauwenthomas <vergauwenthomas@users.noreply.github.com>
