+++ docs/paper/paper.bib
@@ -214,6 +214,3 @@ year = {2020}
   doi          = {10.5281/zenodo.6364594},
   url          = {https://doi.org/10.5281/zenodo.6364594}
 }
-
-
-
