+++ metobs_toolkit/plotting_functions.py
@@ -39,9 +39,16 @@ from metobs_toolkit.df_helpers import xs_save
 logger = logging.getLogger(__name__)
 
 
-def folium_plot(mapinfo, band, vis_params, labelnames, layername,
-                basemap='SATELLITE', legendname=None,
-                legendpos='bottomleft'):
+def folium_plot(
+    mapinfo,
+    band,
+    vis_params,
+    labelnames,
+    layername,
+    basemap="SATELLITE",
+    legendname=None,
+    legendpos="bottomleft",
+):
     """Make an interactive folium plot of an Image."""
     # get the ee.Image
     im = get_ee_obj(mapinfo, band)
@@ -52,22 +59,23 @@ def folium_plot(mapinfo, band, vis_params, labelnames, layername,
         MAP.add_basemap(basemap)
     MAP.add_layer(im, vis_params, layername)
     if legendname:
-        MAP.add_legend(title=legendname,
-                       labels=labelnames,
-                       colors=vis_params.get('palette'),
-                       position=legendpos)
+        MAP.add_legend(
+            title=legendname,
+            labels=labelnames,
+            colors=vis_params.get("palette"),
+            position=legendpos,
+        )
 
     return MAP
 
 
 def add_stations_to_folium_map(Map, metadf):
     """Add stations as markers to the folium map."""
-    points = metadf['geometry'].to_crs("epsg:4326")
+    points = metadf["geometry"].to_crs("epsg:4326")
     for station, point in points.items():
-        folium.Marker(location=[point.y, point.x],
-                      fill_color='#43d9de',
-                      popup=station,
-                      radius=8).add_to(Map)
+        folium.Marker(
+            location=[point.y, point.x], fill_color="#43d9de", popup=station, radius=8
+        ).add_to(Map)
 
     return Map
 
@@ -110,7 +118,9 @@ def make_cat_colormapper(catlist, cmapname):
 
     # check number of colors in the cmap
     if cmap.N < len(catlist):
-        logger.warning(f'colormap: {cmapname}, is not well suited to color {len(catlist)} categories.')
+        logger.warning(
+            f"colormap: {cmapname}, is not well suited to color {len(catlist)} categories."
+        )
         same_col_n_groups = np.ceil(len(catlist) / cmap.N)
 
         # group cateogries and color them by group
@@ -139,22 +149,34 @@ def make_cat_colormapper(catlist, cmapname):
 # Plotters
 # =============================================================================
 
-def make_folium_html_plot(gdf, variable_column, var_display_name, var_unit,
-                          label_column, label_col_map,
-                          vmin=None, vmax=None,
-                          radius=13, fill_alpha=0.6,
-                          mpl_cmap_name='viridis',
-                          max_fps=4,
-                          dt_disp_fmt='%Y-%m-%d %H:%M'):
+
+def make_folium_html_plot(
+    gdf,
+    variable_column,
+    var_display_name,
+    var_unit,
+    label_column,
+    label_col_map,
+    vmin=None,
+    vmax=None,
+    radius=13,
+    fill_alpha=0.6,
+    mpl_cmap_name="viridis",
+    max_fps=4,
+    dt_disp_fmt="%Y-%m-%d %H:%M",
+):
 
     # create a map
-    m = folium.Map(location=_get_init_mapcenter(gdf),
-                   tiles="cartodbpositron", zoom_start=10,
-                   attr="<a href=https://github.com/vergauwenthomas/MetObs_toolkit </a>")
+    m = folium.Map(
+        location=_get_init_mapcenter(gdf),
+        tiles="cartodbpositron",
+        zoom_start=10,
+        attr="<a href=https://github.com/vergauwenthomas/MetObs_toolkit </a>",
+    )
 
[…]
