+++ metobs_toolkit/obstypes.py
@@ -10,6 +10,7 @@ import logging
 from collections.abc import Iterable
 
 import numpy as np
+
 logger = logging.getLogger(__name__)
 
 
@@ -18,52 +19,63 @@ logger = logging.getLogger(__name__)
 # =============================================================================
 
 tlk_std_units = {
-    "temp": 'Celsius',
-    "radiation_temp": 'Celsius',
-    "humidity": '%',
-    "precip": 'mm/m²',
-    "precip_sum": 'mm/m² from midnight',
-    "wind_speed": 'm/s',
-    "wind_gust": 'm/s',
-    "wind_direction": '° from north (CW)',
-    "pressure": 'pa',
-    "pressure_at_sea_level": 'pa'}
+    "temp": "Celsius",
+    "radiation_temp": "Celsius",
+    "humidity": "%",
+    "precip": "mm/m²",
+    "precip_sum": "mm/m² from midnight",
+    "wind_speed": "m/s",
+    "wind_gust": "m/s",
+    "wind_direction": "° from north (CW)",
+    "pressure": "pa",
+    "pressure_at_sea_level": "pa",
+}
 
 
 # =============================================================================
 # Aliases for units
 # =============================================================================
 
-temp_aliases = {'Celsius': ['celsius', '°C', '°c', 'celcius', 'Celcius'],  # for the dyselectic developper..
-                'Kelvin': ['K', 'kelvin'],
-                'Farenheit': ['farenheit']}
-pressure_aliases = {'pa': ['Pascal', 'pascal', 'Pa'],
-                    'hpa': ['hecto pascal', 'hPa'],
-                    'psi': ['Psi'],
-                    'bar': ['Bar']}
+temp_aliases = {
+    "Celsius": [
+        "celsius",
+        "°C",
+        "°c",
+        "celcius",
+        "Celcius",
+    ],  # for the dyselectic developper..
+    "Kelvin": ["K", "kelvin"],
+    "Farenheit": ["farenheit"],
+}
+pressure_aliases = {
+    "pa": ["Pascal", "pascal", "Pa"],
+    "hpa": ["hecto pascal", "hPa"],
+    "psi": ["Psi"],
+    "bar": ["Bar"],
+}
 
-precip_aliases = {'mm/m²': ['mm', 'liter', 'liters', 'l/m²', "milimeter"]}
+precip_aliases = {"mm/m²": ["mm", "liter", "liters", "l/m²", "milimeter"]}
 
-wind_aliases = {'m/s': ['meters/second', 'm/sec'],
-                'km/h': ['kilometers/hour', 'kph'],
-                'mph': ['miles/hour']}
-direction_aliases = {'° from north (CW)': ['°', 'degrees']}
+wind_aliases = {
+    "m/s": ["meters/second", "m/sec"],
+    "km/h": ["kilometers/hour", "kph"],
+    "mph": ["miles/hour"],
+}
+direction_aliases = {"° from north (CW)": ["°", "degrees"]}
 
 
 # conversion between standard-NAMES and aliases
 all_units_aliases = {
     "temp": temp_aliases,
     "radiation_temp": temp_aliases,
-    "humidity": {
-        "%": ['percent', 'percentage']
-    },
-    'pressure': pressure_aliases,
-    'pressure_at_sea_level': pressure_aliases,
-    'precip': precip_aliases,
-    'precip_sum': precip_aliases,
-    'wind_speed': wind_aliases,
-    'wind_gust': wind_aliases,
-    'wind_direction': direction_aliases,
+    "humidity": {"%": ["percent", "percentage"]},
+    "pressure": pressure_aliases,
+    "pressure_at_sea_level": pressure_aliases,
+    "precip": precip_aliases,
+    "precip_sum": precip_aliases,
+    "wind_speed": wind_aliases,
[…]
