+++ .pre-commit-config.yaml
@@ -9,3 +9,8 @@ repos:
     rev: 22.10.0
     hooks:
     -   id: black
+- repo: https://github.com/macisamuele/language-formatters-pre-commit-hooks
+  rev: v2.12.0
+  hooks:
+  - id: pretty-format-toml
+    args: [--autofix]
