### 2024-01-25 a4006e8
added flake and toml checks to pre-commit
