+++ pyproject.toml
@@ -1,51 +1,46 @@
+[build-system]
+build-backend = "poetry.core.masonry.api"
+requires = ["poetry-core"]
+
 [tool.poetry]
-name = "MetObs-toolkit"
-version = "0.1.3a"
-description = "A Meteorological observations toolkit for scientists"
 authors = ["Thomas Vergauwen <thomas.vergauwen@meteo.be>"]
-maintainers = ["Thomas Vergauwen <thomas.vergauwen@meteo.be>"]
-license = "LICENSE"
-readme = "README.md"
+description = "A Meteorological observations toolkit for scientists"
 documentation = "https://vergauwenthomas.github.io/MetObs_toolkit/"
-packages = [{include = "metobs_toolkit"}]
 keywords = ["meteorology", "observations", "urban climate"]
+license = "LICENSE"
+maintainers = ["Thomas Vergauwen <thomas.vergauwen@meteo.be>"]
+name = "MetObs-toolkit"
+packages = [{include = "metobs_toolkit"}]
+readme = "README.md"
+version = "0.1.3a"
 
 [tool.poetry.dependencies]
-python = "^3.9"
-pandas = "^1.3.0"
-#numpy = "^1.17.3" #in pandas
-matplotlib = "^3.0.0"
-geopandas = "^0.9.0"
-pyproj = "~3.4"
 cartopy = '^0.20'
-mapclassify = "^2.4.0"
 earthengine-api = "^0.1.340"
 geemap = '^0.20.0'
-
-
-
-[tool.poetry.group.titan.dependencies]
-#titanlib requires C-compilers, which are not by default present on windows.
-#Make a seperate group for titan
-titanlib = '^0.3'
-
-
-[tool.poetry.group.documentation.dependencies]
-#Group of dep packages for building the documentation
-sphinx = '^7.2'
-nbsphinx = '^0.9'
-sphinx-copybutton = '^0.5.1'
-sphinx-rtd-theme = '^1.3.0'
-myst_parser = '^2.0.0'
-pandoc = '^2.1' #Check on PyPi (not in local conda env) + with system wide install
-
+geopandas = "^0.9.0"
+mapclassify = "^2.4.0"
+# numpy = "^1.17.3" #in pandas
+matplotlib = "^3.0.0"
+pandas = "^1.3.0"
+pyproj = "~3.4"
+python = "^3.9"
 
 [tool.poetry.group.dev.dependencies]
-#To run poetry tests
+# To run poetry tests
 poetry = "^1.7"
 pre-commit = "^3.6"
 
+[tool.poetry.group.documentation.dependencies]
+myst_parser = '^2.0.0'
+nbsphinx = '^0.9'
+pandoc = '^2.1'  # Check on PyPi (not in local conda env) + with system wide install
+# Group of dep packages for building the documentation
+sphinx = '^7.2'
+sphinx-copybutton = '^0.5.1'
+sphinx-rtd-theme = '^1.3.0'
 
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+[tool.poetry.group.titan.dependencies]
+# titanlib requires C-compilers, which are not by default present on windows.
+# Make a seperate group for titan
+titanlib = '^0.3'
