+++ .pre-commit-config.yaml
@@ -15,12 +15,13 @@ repos:
   hooks:
   - id: pretty-format-toml
     args: [--autofix]
-- repo: https://github.com/pycqa/flake8
-  rev: 6.1.0
-  hooks:
-  - id: flake8
-    additional_dependencies: ['flake8-alphabetize', 'flake8-rst-docstrings']
-    args: ['--config=setup.cfg']
+
+#- repo: https://github.com/pycqa/flake8
+#  rev: 6.1.0
+#  hooks:
+#  - id: flake8
+#    additional_dependencies: ['flake8-alphabetize', 'flake8-rst-docstrings']
+#    args: ['--config=setup.cfg']
 
 ci:
   autofix_commit_msg: |
