+++ pyproject.toml
@@ -15,7 +15,7 @@ readme = "README.md"
 version = "0.1.3a"
 
 [tool.poetry.dependencies]
-cartopy = '^0.20'
+cartopy = '^0.21'
 earthengine-api = "^0.1.340"
 geemap = '^0.20.0'
 geopandas = "^0.9.0"
