### 2024-01-25 1175284
Pre commit with black (#424)
* Saved new PDF of paper

* added an ran pre-commit

* Saved new PDF of paper

* install pre-commit format for toml file

* added flake and toml checks to pre-commit

* run flake8

* Saved new PDF of paper

* formatted wit flake

* Saved new PDF of paper

---------

Co-authored-by: vergauwenthomas <vergauwenthomas@users.noreply.github.com>
