+++ metobs_toolkit/qc_statistics.py
@@ -45,7 +45,7 @@ def get_freq_statistics(comb_df, obstype, checks_info, gaps_info, applied_qc_ord
     """
     outlier_labels = [qc["outlier_flag"] for qc in checks_info.values()]
 
-    final_counts = comb_df['label'].value_counts()
+    final_counts = comb_df["label"].value_counts()
 
     # add missing labels
     # QC labels
