+++ metobs_toolkit/obstype_modeldata.py
@@ -11,10 +11,7 @@ import numpy as np
 import logging
 from metobs_toolkit.obstypes import Obstype
 
-from metobs_toolkit.obstypes import (temperature,
-                                     pressure,
-                                     wind,
-                                     direction_aliases)
+from metobs_toolkit.obstypes import temperature, pressure, wind, direction_aliases
 
 logger = logging.getLogger(__name__)
 
@@ -22,24 +19,41 @@ logger = logging.getLogger(__name__)
 # Standard modeldata equivalents
 # =============================================================================
 tlk_std_modeldata_obstypes = {
-    'temp': {"ERA5_hourly": {'name': 'temperature_2m', 'units': 'Kelvin',
-                             'band_desc': "Temperature of air at 2m above the surface of land, sea or in-land waters. 2m temperature is calculated by interpolating between the lowest model level and the Earth's surface, taking account of the atmospheric conditions."}},
-
-    'pressure': {'ERA5_hourly': {'name': 'surface_pressure', 'units': 'pa',
-                                 'band_desc': "Pressure (force per unit area) of the atmosphere on the surface of land, sea and in-land water. It is a measure of the weight of all the air in a column vertically above the area of the Earth's surface represented at a fixed point. Surface pressure is often used in combination with temperature to calculate air density. The strong variation of pressure with altitude makes it difficult to see the low and high pressure systems over mountainous areas, so mean sea level pressure, rather than surface pressure, is normally used for this purpose. The units of this variable are Pascals (Pa). Surface pressure is often measured in hPa and sometimes is presented in the old units of millibars, mb (1 hPa = 1 mb = 100 Pa)."}},
-    'u_wind': {'ERA5_hourly': {'name': 'u_component_of_wind_10m', 'units': 'm/s',
-                                 'band_desc': "Eastward component of the 10m wind. It is the horizontal speed of air moving towards the east, at a height of ten meters above the surface of the Earth, in meters per second. Care should be taken when comparing this variable with observations, because wind observations vary on small space and time scales and are affected by the local terrain, vegetation and buildings that are represented only on average in the ECMWF Integrated Forecasting System. This variable can be combined with the V component of 10m wind to give the speed and direction of the horizontal 10m wind."}},
-    'v_wind': {'ERA5_hourly': {'name': 'v_component_of_wind_10m', 'units': 'm/s',
-                                  'band_desc': "Northward component of the 10m wind. It is the horizontal speed of air moving towards the north, at a height of ten meters above the surface of the Earth, in meters per second. Care should be taken when comparing this variable with observations, because wind observations vary on small space and time scales and are affected by the local terrain, vegetation and buildings that are represented only on average in the ECMWF Integrated Forecasting System. This variable can be combined with the U component of 10m wind to give the speed and direction of the horizontal 10m wind."}},
-
+    "temp": {
+        "ERA5_hourly": {
+            "name": "temperature_2m",
+            "units": "Kelvin",
+            "band_desc": "Temperature of air at 2m above the surface of land, sea or in-land waters. 2m temperature is calculated by interpolating between the lowest model level and the Earth's surface, taking account of the atmospheric conditions.",
+        }
+    },
+    "pressure": {
+        "ERA5_hourly": {
+            "name": "surface_pressure",
+            "units": "pa",
[…]

[…]
