+++ pyproject.toml
@@ -19,6 +19,7 @@ cartopy = '^0.21'
 earthengine-api = "^0.1.340"
 geemap = '^0.20.0'
 geopandas = "^0.9.0"
+geos = '^3.7'
 mapclassify = "^2.4.0"
 # numpy = "^1.17.3" #in pandas
 matplotlib = "^3.0.0"
