+++ pyproject.toml
@@ -15,11 +15,11 @@ readme = "README.md"
 version = "0.1.3a"
 
 [tool.poetry.dependencies]
-cartopy = '^0.21'
+cartopy = '^0.21.1'
 earthengine-api = "^0.1.340"
 geemap = '^0.20.0'
 geopandas = "^0.9.0"
-geos = '^3.7'
+# geos = '^3.7'
 mapclassify = "^2.4.0"
 # numpy = "^1.17.3" #in pandas
 matplotlib = "^3.0.0"
