### 2024-01-25 299af4e
add shapley and pygeos versions explicitly to pyproject
