+++ pyproject.toml
@@ -28,6 +28,7 @@ pygeos = "^0.14"
 pyproj = "~3.4"
 python = "^3.9"
 shapely = "^2.0.2"
+xarray = "^2024.1.1"
 
 [tool.poetry.group.dev.dependencies]
 # To run poetry tests
