### 2024-01-25 2ca7abb
rely on xarray for cartopy
