+++ pyproject.toml
@@ -15,7 +15,7 @@ readme = "README.md"
 version = "0.1.3a"
 
 [tool.poetry.dependencies]
-cartopy = '^0.21.1'
+# cartopy = '^0.21.1'
 earthengine-api = "^0.1.340"
 geemap = '^0.20.0'
 geopandas = "^0.9.0"
@@ -24,8 +24,8 @@ mapclassify = "^2.4.0"
 # numpy = "^1.17.3" #in pandas
 matplotlib = "^3.0.0"
 pandas = "^1.3.0"
-pygeos = "^0.14"
-pyproj = "~3.4"
+# pygeos = "^0.14"
+# pyproj = "~3.4"
 python = "^3.9"
 shapely = "^2.0.2"
 xarray = "^2024.1.1"
