+++ pyproject.toml
@@ -15,10 +15,12 @@ readme = "README.md"
 version = "0.1.3a"
 
 [tool.poetry.dependencies]
+cartopy = "^0.22.0"
 # cartopy = '^0.21.1'
 earthengine-api = "^0.1.340"
 geemap = '^0.20.0'
 geopandas = "^0.9.0"
+geos = "^0.2.3"
 # geos = '^3.7'
 mapclassify = "^2.4.0"
 # numpy = "^1.17.3" #in pandas
@@ -28,7 +30,6 @@ pandas = "^1.3.0"
 # pyproj = "~3.4"
 python = "^3.9"
 shapely = "^2.0.2"
-xarray = "^2024.1.1"
 
 [tool.poetry.group.dev.dependencies]
 # To run poetry tests
