### 2024-01-25 e125e89
add geos to dep test (#425)
* add geos to dep test

* Saved new PDF of paper

* change cartopy version

* Saved new PDF of paper

* add shapley and pygeos versions explicitly to pyproject

* Saved new PDF of paper

* add xarray

* Saved new PDF of paper

* rely on xarray for cartopy

* Saved new PDF of paper

* test geos dep

* Saved new PDF of paper

---------

Co-authored-by: vergauwenthomas <vergauwenthomas@users.noreply.github.com>
