### 2024-01-25 62f5559
Dev (#421)
* Siebe bug (#364)

* * add Siebe dataset to testing framework
* fixed bug so that outliers in combined df are in obs time frequency
* add obstype in qc plot title
* fixed bug that qc settings can be update on non existing default obstypes
* extended function to check if a qc can be applied to check if all required settings are available.

* fixed docstring

* build on dev branch

* Titanlib extra (#367)

* add titanlib in extension group

* add documentation + catchers if titanlib is not installed

* add all extras in the github actions

* less restrictive on titanlib version

* up
[…]
