+++ metobs_toolkit/data_import.py
@@ -14,14 +14,13 @@ import pandas as pd
 # from mysql.connector import errorcode
 from pytz import all_timezones
 
-from metobs_toolkit import observation_types
-
 logger = logging.getLogger(__name__)
 
 
 # =============================================================================
 # Helpers
 # =============================================================================
+
 
 def _remove_keys_from_dict(dictionary, keys):
     for key in keys:
@@ -105,6 +104,7 @@ def _read_csv_to_df(filepath, kwargsdict):
 # Template
 # =============================================================================
 
+
 def check_template_compatibility(template, df_columns, filetype):
     """Log template compatiblity with dataframe columns.
 
@@ -123,11 +123,13 @@ def check_template_compatibility(template, df_columns, filetype):
 
     """
     # ignore datetime because this is already mapped
-    present_cols = [col for col in df_columns if col != 'datetime']
-    assumed_cols = [key for key in template.keys() if key != 'datetime']
+    present_cols = [col for col in df_columns if col != "datetime"]
+    assumed_cols = [key for key in template.keys() if key != "datetime"]
 
     # in mapper but not in df
-    unmapped_assumed = [templ_var for templ_var in assumed_cols if templ_var not in present_cols]
+    unmapped_assumed = [
+        templ_var for templ_var in assumed_cols if templ_var not in present_cols
+    ]
 
     if len(unmapped_assumed) > 0:
         logger.info(
@@ -139,7 +141,8 @@ def check_template_compatibility(template, df_columns, filetype):
     unmapped_appearing = [col for col in present_cols if col not in assumed_cols]
     if len(unmapped_appearing) > 0:
         logger.info(
-            f"The following columns in the {filetype} cannot be mapped with the template: {unmapped_appearing}.")
+            f"The following columns in the {filetype} cannot be mapped with the template: {unmapped_appearing}."
+        )
 
     # check if at least one column is mapped
     if len(list(set(present_cols) - set(assumed_cols))) == len(present_cols):
@@ -148,13 +151,16 @@ def check_template_compatibility(template, df_columns, filetype):
         )
 
 
-def extract_options_from_template(templ):
+def extract_options_from_template(templ, known_obstypes):
     """Filter out options settings from the template dataframe.
 
     Parameters
     ----------
     templ : pandas.DataFrame()
         Template in a dataframe structure
+    known_obstypes : list
+        A list of known observation types. These consist of the default
+        obstypes and the ones added by the user.
 
     Returns
     -------
@@ -165,68 +171,77 @@ def extract_options_from_template(templ):
 
     """
     opt_kwargs = {}
-    if 'options' in templ.columns:
-        if 'options_values' in templ.columns:
-            opt = templ[['options', 'options_values']]
+    if "options" in templ.columns:
+        if "options_values" in templ.columns:
+            opt = templ[["options", "options_values"]]
             # drop nan columns
-            opt = opt[opt['options'].notna()]
+            opt = opt[opt["options"].notna()]
             # convert to dict
-            opt = opt.set_index('options')['options_values'].to_dict()
+            opt = opt.set_index("options")["options_values"].to_dict()
 
             # check options if valid
-            possible_options = {'data_structure': ['long', 'wide', 'single_station'],
-                                'stationname': '_any_',
-                                'obstype': observation_types,
-                                'obstype_unit': '_any_',
-                                'obstype_description': '_any_',
-                                'timezone': all_timezones
-                                }
+            possible_options = {
+                "data_structure": ["long", "wide", "single_station"],
+                "stationname": "_any_",
+                "obstype": known_obstypes,
+                "obstype_unit": "_any_",
+                "obstype_description": "_any_",
[…]

[…]
