### 2024-01-26 dc4cee7
setup gee structure + static point extractors
