+++ metobs_toolkit/landcover_functions.py
@@ -22,12 +22,12 @@ logger = logging.getLogger(__name__)
 # =============================================================================
 
 
-def connect_to_gee():
-    """Authenticate to GEE if needed."""
-    if not ee.data._credentials:  # check if ee connection is initialized
-        ee.Authenticate()
-        ee.Initialize()
-    return
+# def connect_to_gee():
+#     """Authenticate to GEE if needed."""
+#     if not ee.data._credentials:  # check if ee connection is initialized
+#         ee.Authenticate()
+#         ee.Initialize()
+#     return
 
 
 # =============================================================================
@@ -35,83 +35,83 @@ def connect_to_gee():
 # =============================================================================
 
 
-def lcz_extractor(metadf, mapinfo):
-    """Extract LCZ for all stations in the metadf."""
-    # make return in case something went wrong
-    default_return = pd.Series(
-        index=metadf.index, data="Location_unknown", name="lcz", dtype=object
-    )
-    # test if metadata is suitable
-    if not _validate_metadf(metadf):
-        logger.warning(f"Metadf is not suitable for GEE extractiond: {metadf}")
-        return default_return
+# def lcz_extractor(metadf, mapinfo):
+#     """Extract LCZ for all stations in the metadf."""
+#     # make return in case something went wrong
+#     default_return = pd.Series(
+#         index=metadf.index, data="Location_unknown", name="lcz", dtype=object
+#     )
+#     # test if metadata is suitable
+#     if not _validate_metadf(metadf):
+#         logger.warning(f"Metadf is not suitable for GEE extractiond: {metadf}")
+#         return default_return
 
-    relevant_metadf = metadf.reset_index()[["name", "lat", "lon"]]
+#     relevant_metadf = metadf.reset_index()[["name", "lat", "lon"]]
 
-    lcz_df = extract_pointvalues(
-        metadf=relevant_metadf, mapinfo=mapinfo, output_column_name="lcz"
-    )
-    if lcz_df.empty:
-        return pd.Series(dtype=object)
-    return lcz_df["lcz"]  # return series
-
-
-def lc_fractions_extractor(metadf, mapinfo, buffer, agg):
-    """Get landcover fractions for all buffers from GEE."""
-    # make return in case something went wrong
-    default_return = (pd.DataFrame(index=metadf.index), buffer)
-
-    # test if metadata is suitable
-    if not _validate_metadf(metadf):
-        logger.warning(f"Metadf is not suitable for GEE extractiond: {metadf}")
-        return default_return
-
-    relevant_metadf = metadf.reset_index()[["name", "lat", "lon"]]
-
-    freqs_df = extract_buffer_frequencies(
-        metadf=relevant_metadf, mapinfo=mapinfo, bufferradius=buffer
-    )
-
-    # apply aggregation if required
-    if agg:
-        logger.info(f"Using aggregation scheme: {mapinfo['aggregation']}")
-        agg_df = pd.DataFrame()
-        for agg_name, agg_classes in mapinfo["aggregation"].items():
-            present_agg_classes = [
-                str(num) for num in agg_classes if str(num) in freqs_df.columns
-            ]
-            agg_df[agg_name] = freqs_df[present_agg_classes].sum(axis=1)
-
-        return agg_df, buffer
-
-    else:
-        # map numeric classes to human
-        mapper = {
-            str(num): human for num, human in mapinfo["categorical_mapper"].items()
-        }
-        freqs_df = freqs_df.rename(columns=mapper)
-
-        return freqs_df, buffer
+#     lcz_df = extract_pointvalues(
+#         metadf=relevant_metadf, mapinfo=mapinfo, output_column_name="lcz"
+#     )
+#     if lcz_df.empty:
+#         return pd.Series(dtype=object)
[…]
