+++ metobs_toolkit/dataset.py
@@ -26,10 +26,7 @@ from metobs_toolkit.data_import import (
 
 from metobs_toolkit.printing import print_dataset_info
 from metobs_toolkit.landcover_functions import (
-    connect_to_gee,
-    lcz_extractor,
-    height_extractor,
-    lc_fractions_extractor,
+    # connect_to_gee,
     _validate_metadf,
 )
 
@@ -95,6 +92,7 @@ from metobs_toolkit.obstypes import Obstype as Obstype_class
 
 from metobs_toolkit.analysis import Analysis
 from metobs_toolkit.modeldata import Modeldata
+from metobs_toolkit.gee_extractor import GeeExtractor
 
 
 logger = logging.getLogger(__name__)
@@ -3363,6 +3361,7 @@ station with the default name: {self.settings.app["default_name"]}.'
         wudapt global LCZ map on the Google engine for all stations.
 
         A 'LCZ' column will be added to the metadf, and series is returned.
+        If this column already exists, it will be overwritten.
 
         Returns
         -------
@@ -3370,23 +3369,16 @@ station with the default name: {self.settings.app["default_name"]}.'
             A series with the stationnames as index and the LCZ as values.
 
         """
-        # connect to gee
-        connect_to_gee()
-
-        # Extract LCZ for all stations
-        lcz_series = lcz_extractor(
-            metadf=self.metadf,
-            mapinfo=self.settings.gee["gee_dataset_info"]["global_lcz_map"],
-        )
-
         # drop column if it was already present
         if "lcz" in self.metadf:
             self.metadf = self.metadf.drop(columns=["lcz"])
 
-        # update metadata
-        self.metadf = self.metadf.merge(
-            lcz_series.to_frame(), how="left", left_index=True, right_index=True
+        lcz_extractor = GeeExtractor()
+        lcz_extractor.activate_LCZ()
+        lcz_series = lcz_extractor.extract_static_point_values(
+            metadf=self.metadf, aggregate=False
         )
+        self.metadf["lcz"] = lcz_series
         return lcz_series
 
     def get_altitude(self):
@@ -3396,6 +3388,7 @@ station with the default name: {self.settings.app["default_name"]}.'
         global map on the Google engine for all stations.
 
         A 'altitude' column will be added to the metadf, and series is returned.
+        If this column already exists, it will be overwritten.
 
         Returns
         -------
@@ -3403,27 +3396,19 @@ station with the default name: {self.settings.app["default_name"]}.'
             A series with the stationnames as index and the altitudes as values.
 
         """
-        # connect to gee
-        connect_to_gee()
-
-        # Extract LCZ for all stations
-        altitude_series = height_extractor(
-            metadf=self.metadf, mapinfo=self.settings.gee["gee_dataset_info"]["DEM"]
-        )
-
         # drop column if it was already present
         if "altitude" in self.metadf:
             self.metadf = self.metadf.drop(columns=["altitude"])
 
-        # update metadata
-        self.metadf = self.metadf.merge(
-            altitude_series.to_frame(), how="left", left_index=True, right_index=True
+        alt_extractor = GeeExtractor()
+        alt_extractor.activate_DEM()
+        altitude_series = alt_extractor.extract_static_point_values(
+            metadf=self.metadf, aggregate=False
         )
+        self.metadf["altitude"] = altitude_series
         return altitude_series
 
-    def get_landcover(
-        self, buffers=[100], aggregate=True, overwrite=True, gee_map="worldcover"
-    ):
+    def get_landcover(self, buffers=[100], aggregate=True, overwrite=True):
         """Extract landcover for all stations.
 
         Extract the landcover fractions in a buffer with a specific radius for
@@ -3445,10 +3430,6 @@ station with the default name: {self.settings.app["default_name"]}.'
         overwrite : bool, optional
[…]

[…]
