### 2024-01-29 a20ab0e
work in progress on seting up new interaction schemes between observations and modeldata
