+++ metobs_toolkit/gee_extractor.py
@@ -15,6 +15,7 @@ import ee
 from metobs_toolkit.landcover_functions import (
     extract_pointvalues,
     extract_buffer_frequencies,
+    gee_extract_timeseries,
 )
 
 logger = logging.getLogger(__name__)
@@ -51,7 +52,7 @@ class GeeExtractor:
         if self.location is None:
             return f"GeeExtractor with unknown target gee location."
         else:
-            return f"GeeExtractor with for {self.usage} use: \n \
+            return f"GeeExtractor for {self.usage} use: \n \
   location : {self.location}"
 
     def __repr__(self):
@@ -86,6 +87,7 @@ class GeeExtractor:
         if not _is_metadf_valid(metadf):
             sys.exit()
 
+        # filter to non-nan stations
         filtered_metadf = metadf[metadf[["lat", "lon"]].notnull().all(1)]
 
         # authenticate to GEE
@@ -202,6 +204,84 @@ class GeeExtractor:
         valuesdf = valuesdf.fillna(self.missing_value_label)
         return valuesdf
 
+    def extract_timeseries(
+        self, metadf, bandnames, start_utc, end_utc, gdrive_filename="era5_data"
+    ):
+
+        # check if self is valid
+        if not self.dynamical:
+            sys.exit(
+                "Time series gee extraction is only valid for dynamical (time dep.) dataset"
+            )
+        if self.time_res is None:
+            sys.exit(f"No time_res specified for {self}")
+
+        # Check if metadf is valid as coordinates
+        _ = _is_metadf_valid(metadf)
+
+        # check if bandnames is valid
+        if isinstance(bandnames, str):
+            bandnames = [bandnames]
+
+        # check if datetimes are valid
+        assert (
+            end_utc > start_utc
+        ), f"end_utc ({end_utc}) not later than start_utc ({start_utc})."
+        assert end_utc > start_utc + pd.Timedelta(
+            self.time_res
+        ), f"end_utc ({end_utc}) not later than start_utc + time res ({start_utc} + {pd.Timedelta(self.time_res)})."
+
+        # filter to non-nan stations
+        filtered_metadf = metadf[metadf[["lat", "lon"]].notnull().all(1)]
+
+        # authenticate to GEE
+        self.connect_to_gee()
+
+        # Extract timeseries
+        df = gee_extract_timeseries(
+            metadf=filtered_metadf,
+            bandnames=bandnames,
+            startdt=start_utc,
+            enddt=end_utc,
+            scale=self.scale,
+            timeres=self.time_res,
+            trg_gee_loc=self.location,
+            is_imagecollection=self.is_imagecollection,
+            is_image=self.is_image,
+            gdrive_filename=gdrive_filename,
+        )
+        if not df.empty:
+            # Format dataframe
+            df["datetime"] = pd.to_datetime(df["datetime"], format="%Y%m%d%H%M%S")
+            # set timezone
+            df["datetime"] = df["datetime"].dt.tz_localize("UTC")
+
+            # format index
+            df = df.set_index(["feature_idx", "datetime"])
+            df = df.sort_index()
+
+            # Add missing stations
+            missing_stations = list(
+                set(metadf.index) - set(df.index.get_level_values("feature_idx"))
+            )
+            unique_datetimes = df.index.get_level_values("datetime").unique()
+            for missing_sta in missing_stations:
+                missing_df = pd.DataFrame(index=unique_datetimes, columns=df.columns)
+                missing_df["feature_idx"] = str(missing_sta)
+                missing_df = missing_df.reset_index().set_index(
+                    ["feature_idx", "datetime"]
+                )
+                df = pd.concat([df, missing_df])
+
+            # fix the index to the original naming + datetime as multiindex
+            df = (
+                df.reset_index()
[…]
