+++ metobs_toolkit/modeldata.py
@@ -23,13 +23,14 @@ from metobs_toolkit.df_helpers import (
 
 from metobs_toolkit.plotting_functions import model_timeseries_plot, timeseries_plot
 
-# from metobs_toolkit.obstypes import tlk_obstypes
+
 from metobs_toolkit.obstypes import Obstype as Obstype_class
+from metobs_toolkit.obstype_modeldata import era5_default_model_obstypes
 from metobs_toolkit.obstype_modeldata import (
-    model_obstypes,
     ModelObstype,
     ModelObstype_Vectorfield,
 )
+from metobs_toolkit.gee_extractor import GeeExtractor
 from metobs_toolkit.obstype_modeldata import compute_amplitude, compute_angle
 from metobs_toolkit.settings import Settings
 
@@ -43,33 +44,35 @@ logger = logging.getLogger(__name__)
 class Modeldata:
     """Class holding timeseries data and methods for modeldata."""
 
-    def __init__(self):
+    def __init__(self, metadf, extractor=None):
         """Initialize modeldata."""
         self.df = init_multiindexdf()  # Holds the timeseries coming from a model
-        self.metadf = (
-            pd.DataFrame()
-        )  # Holds the metadata for where modeldate is extracted
 
-        # self.modelname = modelname
+        _ = _is_metadf_valid(metadf)
+        self.metadf = metadf  # Holds the metadata for where modeldate is extracted
+        self.obstypes = []
+
+        self.extractor = extractor
+        if isinstance(extractor, GeeExtractor):
+            # Default obstypes Dict name: Obstype-instance
+            self.obstypes = era5_default_model_obstypes.copy()
+        else:
+            self.obstypes = []  # only default era5 modelobstypes are defined
 
         self._settings = Settings()
-        # self.mapinfo = self._settings.gee["gee_dataset_info"]
-
         self.df_tz = "UTC"  # the timezone of the datetimes stored in the df
-        self.obstypes = model_obstypes  # Dict name: Obstype-instance
 
     def __str__(self):
         """Print overview information of the modeldata."""
         if self.df.empty:
-            return "Empty Modeldata instance."
+            return f"Empty Modeldata instance linked to {self.extractor}."
         n_stations = self.df.index.get_level_values("name").unique().shape[0]
         obstypes = self.df.columns.to_list()
         startdt = self.df.index.get_level_values("datetime").min()
         enddt = self.df.index.get_level_values("datetime").max()
         data_units = [self.obstypes[col].get_standard_unit() for col in self.df.columns]
 
-        return f"Modeldata instance containing: \n \
-    * Modelname: {self.modelname} \n \
+        return f"Modeldata instance linked to {self.extractor}, containing: \n \
     * {n_stations} timeseries \n \
     * The following obstypes are available: {obstypes} \n \
     * Data has these units: {data_units} \n \
@@ -83,11 +86,10 @@ class Modeldata:
         """Print out detailed information on the Modeldata."""
         print(str(self))
 
-        # print("\n ------ Known gee datasets -----------")
-        # self.list_gee_datasets()
-
-    def add_obstype(self, Obstype, bandname, band_units, band_description=None):
-        """Add a new Observation type for the current Modeldata.
+    def add_obstype(self, Modelobstype):
+        """
+        TODO: update docstring
+        Add a new Observation type for the current Modeldata.
 
 
         Parameters
@@ -107,337 +109,59 @@ class Modeldata:
         None.
 
         """
-        if not isinstance(Obstype, Obstype_class):
-            sys.exit(
-                f"{Obstype} is not an instance of metobs_toolkit.obstypes.Obstype."
-            )
+        if isinstance(Modelobstype, ModelObstype_Vectorfield):
[…]
