### 2024-01-30 aff5dc0
further work on new implementation of modeldata
