+++ metobs_toolkit/gee_extractor.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
 Created on Fri Jan 26 10:14:38 2024
@@ -7,6 +6,7 @@ Created on Fri Jan 26 10:14:38 2024
 """
 
 import sys
+import os
 import logging
 import numpy as np
 import pandas as pd
@@ -17,6 +17,8 @@ from metobs_toolkit.landcover_functions import (
     extract_buffer_frequencies,
     gee_extract_timeseries,
 )
+
+from metobs_toolkit.plotting_functions import folium_plot, add_stations_to_folium_map
 
 logger = logging.getLogger(__name__)
 
@@ -44,6 +46,7 @@ class GeeExtractor:
         self.categorical_map = (
             {}
         )  # dictionary to map categories {gee_Value: target_str}
+        self.colorscheme = {}
         self.categorical_aggregation = (
             {}
         )  # dictionary to aggregate categories to {target_str : [list of gee_values]}
@@ -64,6 +67,7 @@ class GeeExtractor:
 
     def connect_to_gee(self, **kwargs):
         """Authenticate to GEE if needed."""
+        logger.debug("Authenticate to gee")
         if not ee.data._credentials:  # check if ee connection is initialized
             ee.Authenticate(**kwargs)
             ee.Initialize()
@@ -184,6 +188,7 @@ class GeeExtractor:
                 sys.exit(
                     "Cannot aggregate categorical values, since no categorical_aggregation is defined."
                 )
+            logger.debug(f"Aggregate categories using {self.categorical_aggregation}.")
             aggdf = pd.DataFrame(index=valuesdf.index)
             for aggcat, coverlist in self.categorical_aggregation.items():
                 present_agg_classes = [
@@ -286,6 +291,7 @@ class GeeExtractor:
     # default targets
     # =============================================================================
     def activate_LCZ(self):
+        logger.info("Activate LCZ map on the GeeExtractor.")
         self.location = "RUB/RUBCLIM/LCZ/global_lcz_map/v1"
         self.usage = "LCZ"
         self.scale = 100
@@ -319,6 +325,7 @@ class GeeExtractor:
         self.categorical_aggregation = {}
 
     def activate_worldcover(self):
+        logger.info("Activate worldcover map on the GeeExtractor.")
         self.location = "ESA/WorldCover/v200"
         self.usage = "landcover"
         self.scale = 10
@@ -343,6 +350,20 @@ class GeeExtractor:
             95: "Mangroves",
             100: "Moss and lichen",
         }
+        self.colorscheme = {
+            10: "006400",
+            20: "ffbb22",
+            30: "ffff4c",
+            40: "f096ff",
+            50: "fa0000",
+            60: "b4b4b4",
+            70: "f0f0f0",
+            80: "0064c8",
+            90: "0096a0",
+            95: "00cf75",
+            100: "fae6a0",
+        }
+
         self.categorical_aggregation = {
             "water": [70, 80, 90, 95],
             "pervious": [10, 20, 30, 40, 60, 100],
@@ -350,6 +371,7 @@ class GeeExtractor:
         }
 
     def activate_ERA5(self):
+        logger.info("Activate ERA5 LAND on the GeeExtractor.")
         self.location = "ECMWF/ERA5_LAND/HOURLY"
         self.usage = "ERA5"
         self.scale = 2500
@@ -366,7 +388,7 @@ class GeeExtractor:
         self.categorical_aggregation = {}
 
     def activate_DEM(self):
-
+        logger.info("Activate DEM map on the GeeExtractor.")
         self.location = "CGIAR/SRTM90_V4"
         self.usage = "elevation"
         self.scale = 100
[…]
