+++ metobs_toolkit/gap_filling.py
@@ -158,7 +158,7 @@ def create_leading_trailing_debias_periods(
 
     # only datetimes are relevant
     obs = obs.reset_index()
-    obs = obs[["name", "datetime", obstype]]
+    obs = obs[["name", "datetime", obstype.name]]
 
     # Select all leading and all trailing obs
     leading_period = obs[obs["datetime"] < gap.startgap]
@@ -319,7 +319,7 @@ def create_leading_trailing_debias_periods(
     return leading_df, trailing_df
 
 
-def get_time_specific_biases(model, obs, obstype, period):
+def get_time_specific_biases(model, obs, obstypename, period):
     """Get hourly biases."""
     diff = model - obs
     diff = diff.reset_index().set_index("datetime")
@@ -327,8 +327,8 @@ def get_time_specific_biases(model, obs, obstype, period):
     diff["minutes"] = diff.index.minute
     diff["seconds"] = diff.index.second
 
-    biases = diff.groupby(["name", "hours", "minutes", "seconds"])[obstype].mean()
-    biases.name = obstype + "_bias_" + period
+    biases = diff.groupby(["name", "hours", "minutes", "seconds"])[obstypename].mean()
+    biases.name = obstypename + "_bias_" + period
 
     biases = biases.reset_index()
     return biases
@@ -341,12 +341,12 @@ def make_era_bias_correction(
     error_message = ""
     # 1. get lead timestamp biases
     lead_biases = get_time_specific_biases(
-        model=leading_model, obs=leading_obs, obstype=obstype, period="lead"
+        model=leading_model, obs=leading_obs, obstypename=obstype.name, period="lead"
     )
 
     # 2. get trailing timestamp biases
     trail_biases = get_time_specific_biases(
-        model=trailing_model, obs=trailing_obs, obstype=obstype, period="trail"
+        model=trailing_model, obs=trailing_obs, obstypename=obstype.name, period="trail"
     )
 
     # 3. apply bias correction on modeldata in gap
@@ -363,13 +363,15 @@ def make_era_bias_correction(
     gap_model = gap_model.reset_index()
 
     gap_model = gap_model.merge(
-        right=lead_biases[["hours", "minutes", "seconds", obstype + "_bias_lead"]],
+        right=lead_biases[["hours", "minutes", "seconds", obstype.name + "_bias_lead"]],
         how="left",
         on=["hours", "minutes", "seconds"],
     )
 
     gap_model = gap_model.merge(
-        right=trail_biases[["hours", "minutes", "seconds", obstype + "_bias_trail"]],
+        right=trail_biases[
+            ["hours", "minutes", "seconds", obstype.name + "_bias_trail"]
+        ],
         how="left",
         on=["hours", "minutes", "seconds"],
     )
@@ -380,22 +382,22 @@ def make_era_bias_correction(
     # use the debias corection (even if it is for a part of the gap!).
     # If either one or both are missing, than no bias correction is applied
     no_debias = gap_model[
-        (gap_model[obstype + "_bias_lead"].isnull())
-        | (gap_model[obstype + "_bias_trail"].isnull())
+        (gap_model[obstype.name + "_bias_lead"].isnull())
+        | (gap_model[obstype.name + "_bias_trail"].isnull())
     ].index
     if not no_debias.empty:
         error_message = f"No debias possible for these gap records: {no_debias},the gap will be filled by model data without bias correction. "
         logger.warning(error_message)
 
     # set weights to zero if not debias correction can be applied on that record
-    gap_model.loc[no_debias, obstype + "_bias_trail"] = 0.0
[…]
