+++ tests/push_test/gap_and_fill_test.py
@@ -219,12 +219,14 @@ dataset.coarsen_time_resolution(freq="30T")
 
 
 # offline mode
-era = metobs_toolkit.Modeldata("ERA5_hourly")
+era = metobs_toolkit.Modeldata(
+    metadf=dataset.metadf, extractor=metobs_toolkit.GeeExtractor()
+)
 
 era_datafile = os.path.join(str(lib_folder), "tests", "test_data", "era5_data.csv")
 
 
-era.set_model_from_csv(era_datafile)
+era.import_gee_data_from_csv(era_datafile)
 
 assert era.df.shape[0] == 5348, "Something wrong with importing era data from csv."
 
