+++ tests/push_test/modeldata_test.py
@@ -36,7 +36,10 @@ dataset.coarsen_time_resolution()
 # dataset.get_modeldata()
 
 #%% test adding gee information
-model_data = metobs_toolkit.Modeldata("ERA5_hourly")
+model_data = metobs_toolkit.Modeldata(
+    metadf=dataset.metadf, extractor=metobs_toolkit.GeeExtractor()
+)
+
 
 # Define a regular obstype
 new_obstype = metobs_toolkit.Obstype(
@@ -49,22 +52,39 @@ new_obstype = metobs_toolkit.Obstype(
     unit_conversions={"hpa": ["x * 100"]},
 )
 
-# add new obstype to model_data
-model_data.add_obstype(
-    Obstype=new_obstype,
-    bandname="surface_pressure",
-    band_units="hpa",
+# Convert to modelobstype
+new_obstype_modelform = metobs_toolkit.ModelObstype(
+    obstype=new_obstype,
+    band_name="surface_pressure",
+    band_unit="hpa",
+    band_description="oijmjimo",
 )
+
+
+# add new obstype to model_data
+model_data.add_obstype(new_obstype_modelform)
+
+assert set(model_data.obstypes.keys()) == set(
+    ["temp", "pressure", "wind", "special_pressure"]
+), "New obstype not added to the modeldata"
 
 
 model_data.get_info()
 from datetime import datetime
 
-tstart = datetime(2022, 10, 3, 23)
-tend = datetime(2022, 10, 4, 4)
-model_data = dataset.get_modeldata(
-    modeldata=model_data, obstype="special_pressure", startdt=tstart, enddt=tend
+tstart = datetime(2022, 9, 3, 23)
+tend = datetime(2022, 9, 4, 4)
+
+model_data.extractor.activate_ERA5()
+
+
+model_data.import_from_gee(
+    target_obstypes=model_data.obstypes["special_pressure"],
+    start_utc=tstart,
+    end_utc=tend,
+    gdrive_filename="era5_data",
 )
+
 
 assert (
     model_data.df.shape[0] == 168
@@ -76,47 +96,49 @@ assert model_data.df.columns.to_list() == [
 model_data.make_plot(obstype_model="special_pressure")
 #%% Test 2D vector fields
 
-model_data = dataset.get_modeldata(
-    modeldata=model_data, obstype="wind", startdt=tstart, enddt=tend
+model_data.import_from_gee(
+    target_obstypes=model_data.obstypes["wind"],
+    start_utc=tstart,
+    end_utc=tend,
 )
 
 print(model_data)
 
-assert model_data.df.columns.to_list() == [
-    "wind_amplitude",
-    "wind_direction",
-], "Something is wrong with column names"
-
+assert set(model_data.df.columns) == set(
+    ["u_comp_wind", "v_comp_wind", "wind_amplitude", "wind_direction"]
+), "something wrong with exploiting vector fields"
+assert model_data.df.sum().to_dict() == {
+    "u_comp_wind": 75.64337158203125,
+    "v_comp_wind": 359.2424736022949,
+    "wind_amplitude": 378.537614761677,
+    "wind_direction": 9559.611737920999,
+}, "something wrong with exploiting vector fields."
 
 #%% Testing multiple field extraction
-model_data.get_gee_dataset_data(
-    mapname=model_data.modelname,
-    metadf=dataset.metadf,
-    obstypes=["temp", "wind"],
-    startdt_utc=tstart,
-    enddt_utc=tend,
+model_data.import_from_gee(
+    target_obstypes=[model_data.obstypes["wind"], model_data.obstypes["temp"]],
+    start_utc=tstart,
+    end_utc=tend,
 )
 
-assert model_data.df.columns.to_list() == [
[…]
