+++ metobs_toolkit/__init__.py
@@ -52,7 +52,7 @@ demo_template = os.path.join(
 
 from metobs_toolkit.dataset import Dataset
 from metobs_toolkit.station import Station
-from metobs_toolkit.modeldata import Modeldata
+from metobs_toolkit.modeldata import Modeldata, import_modeldata
 from metobs_toolkit.obstypes import Obstype
 from metobs_toolkit.obstype_modeldata import ModelObstype, ModelObstype_Vectorfield
 from metobs_toolkit.gee_extractor import GeeExtractor

[…]
