### 2024-02-08 #5916 @hugoledoux
@ashwinvis there are still criteria in your list that are not ticked, could you let us know why? You forgot? You still think that the submission doesn't fulfill those? It would be nice if you elaborated
