### 2024-02-09 #5916 @ashwinvis
> @ashwinvis there are still criteria in your list that are not ticked, could you let us know why? You forgot? You still think that the submission doesn't fulfill those? It would be nice if you elaborated

I need to make that GEE account to verify the software. I will get to it soon
