### 2024-02-09 vergauwenthomas/MetObs_toolkit#388 @Zeitsperre
Hi @vergauwenthomas

The documentation is looking much better, but the API documentation should be flattened a bit if possible, specifically here: https://vergauwenthomas.github.io/MetObs_toolkit/MetObs_documentation_full.html. Addressing this would be enough for me to check off the remaining `Documentation` checks in the review.
