+++ metobs_toolkit/verification.py
@@ -14,12 +14,21 @@ import numpy as np
 import pandas as pd
 import geopandas as gpd
 import xarray as xr
+import matplotlib.pyplot as plt
+from matplotlib.lines import Line2D
+import cartopy.crs as ccrs
+import cartopy.feature as cfeature
+import folium
+import matplotlib
+import branca.colormap as brcm
+from folium import plugins as folium_plugins
 
 from metobs_toolkit.obstypes import Obstype
 from metobs_toolkit.analysis import _make_time_derivatives
 
-from metobs_toolkit.df_helpers import get_seasons, _make_time_derivatives
+from metobs_toolkit.df_helpers import get_seasons, _make_time_derivatives, metadf_to_gdf
 
+# from metobs_toolkit.plotting_functions import make_folium_html_plot
 # from metobs_toolkit.dataset import Dataset
 # from metobs_verif.verification_methods import get_basic_scores_dict
 # from metobs_verif.modeloutput import Modelfield
@@ -217,7 +226,29 @@ class Verification:
             model_obstype=self.modeldata.obstypes[model_obstype],
             groupby=groupby,
         )
-        return scoringdf
+        # return scoringdf
+        # import matplotlib.pyplot as plt
+        # return scoringdf
+        plotdf = scoringdf[["bias", "RMSE", "MAE", "cor"]]
+        plotdf.index = plotdf.index.droplevel("variabel")
+
+        fig, ax = plt.subplots()
+        ax = simple_multiline_plot(
+            df=plotdf,
+            ax=ax,
+            col_map_dict={
+                "bias": self.biascolor,
+                "RMSE": self.rmsecolor,
+                "MAE": self.maecolor,
+                "cor": self.corcolor,
+            },
+            xlabel=f"{groupby}",
+            ylabel=f"/ or {self.dataset.obstypes[observation_obstype].get_plot_y_label()}",
+            title=f"Point verif scores for {self.dataset.obstypes[observation_obstype].name} and {self.modeldata.obstypes[model_obstype].get_bandname()}.",
+            add_zero=True,
+        )
+        return ax
+
         # Find out which situation is applicable
         # cur_Obstype = self.dataset.obstypes[observation_obstype]
 
@@ -325,25 +356,6 @@ class Verification:
                 ]
             ).all(), f"Unknonw groupid in {groupby}."
 
-        # # format model variables
-        # if isinstance(model_variabels, str):
-        #     model_variabels = [model_variabels]
-        # if model_variabels is None:
-        #     #get all fields in the same observationtype
-        #     model_variabels = self._get_corresponding_model_var(obstype=observation_obstype)
-
-        # #Check if all model variables are known
-        # if not np.all([var in self.verifdf.columns for var in model_variabels]):
-        #     sys.exit(f'No all {model_variabels} are found in the verificatin table columns: {self.verifdf.columns}')
-        # for var in model_variabels:
-        #     assert var in self.verifdf.columns, f'{var} not found in the verification table columns: {self.verifdf.columns}'
-
-        # Construct dataframe to calculate grouped scores on
-        # relevant_columns =
-        # relevant_columns = model_variabels.copy()
-        # relevant_columns.append(observation_obstype)
-        # df = self.verifdf[relevant_columns].reset_index().rename(columns={'validate': 'datetime'})
-
         df = self._construct_mod_obs_df(
             obs_obstype=observation_obstype, mod_obstype=model_obstype, interp=True
         )
@@ -398,6 +410,166 @@ class Verification:
         scoringdf = pd.concat(scoringlist)
         return scoringdf
 
[…]
