### 2024-02-12 2bdb6e5
WIP on restructuring docs (similar as geopandas) + flattening the API doctree
