+++ metobs_toolkit/obstype_modeldata.py
@@ -9,7 +9,7 @@ import copy
 import math
 import numpy as np
 import logging
-from metobs_toolkit.obstypes import Obstype
+from metobs_toolkit.obstypes import Obstype, expression_calculator
 
 from metobs_toolkit.obstypes import temperature, pressure, wind, direction_aliases
 
@@ -393,7 +393,7 @@ class ModelObstype_Vectorfield(Obstype):
     def convert_to_standard_units(self, input_df, input_unit):
         """Convert data from a known unit to the standard unit.
 
-        The data c must be a pandas dataframe with both the u and v component
+        The data must be a pandas dataframe with both the u and v component
         prensent as columns.
 
         Parameters
