+++ docs/intro.rst
@@ -1,4 +1,3 @@
-
 *******************
 Introduction
 *******************
