+++ docs/reference/api/metobs_toolkit.Dataset.get_gaps_df.rst
@@ -1,0 +1,6 @@
+﻿metobs\_toolkit.Dataset.get\_gaps\_df
+=====================================
+
+.. currentmodule:: metobs_toolkit
+
+.. automethod:: Dataset.get_gaps_df
