+++ docs/reference/api/metobs_toolkit.Dataset.update_settings.rst
@@ -1,0 +1,6 @@
+﻿metobs\_toolkit.Dataset.update\_settings
+========================================
+
+.. currentmodule:: metobs_toolkit
+
+.. automethod:: Dataset.update_settings
