+++ docs/reference/api/metobs_toolkit.Dataset.get_modeldata.rst
@@ -1,0 +1,6 @@
+﻿metobs\_toolkit.Dataset.get\_modeldata
+======================================
+
+.. currentmodule:: metobs_toolkit
+
+.. automethod:: Dataset.get_modeldata
