+++ docs/reference/api/metobs_toolkit.Dataset.coarsen_time_resolution.rst
@@ -1,0 +1,6 @@
+﻿metobs\_toolkit.Dataset.coarsen\_time\_resolution
+=================================================
+
+.. currentmodule:: metobs_toolkit
+
+.. automethod:: Dataset.coarsen_time_resolution
