+++ docs/reference/api/metobs_toolkit.ModelObstype.get_bandname_mapper.rst
@@ -1,0 +1,6 @@
+﻿metobs\_toolkit.ModelObstype.get\_bandname\_mapper
+==================================================
+
+.. currentmodule:: metobs_toolkit
+
+.. automethod:: ModelObstype.get_bandname_mapper
