+++ docs/reference/api/metobs_toolkit.Dataset.get_analysis.rst
@@ -1,0 +1,6 @@
+﻿metobs\_toolkit.Dataset.get\_analysis
+=====================================
+
+.. currentmodule:: metobs_toolkit
+
+.. automethod:: Dataset.get_analysis
