+++ docs/reference/api/metobs_toolkit.Dataset.get_missing_obs_info.rst
@@ -1,0 +1,6 @@
+﻿metobs\_toolkit.Dataset.get\_missing\_obs\_info
+===============================================
+
+.. currentmodule:: metobs_toolkit
+
+.. automethod:: Dataset.get_missing_obs_info
