### 2024-02-12 vergauwenthomas/MetObs_toolkit#388 @vergauwenthomas
Hi @Zeitsperre

I am working (#434) on the documentation to
 1. Make the documentation API "flatter"
 2. Restructure pages (i am trying to mimic the documentation of [geopandas](https://geopandas.org/en/stable/)

For now i have added all methods-designed-for-users in the doc's. In addition, i have quite a lot of classes/functions/methods that should not be used by users. These are only relevant for developers. So i am not sure if it is good pracktiche to add them to the documentation or not. @Zeitsperre Do you prefer to add these _developer-specific-functions_ to the documentation?
