+++ docs/reference/api/metobs_toolkit.Modeldata.add_obstype.rst
@@ -1,0 +1,6 @@
+metobs\_toolkit.Modeldata.add\_obstype
+======================================
+
+.. currentmodule:: metobs_toolkit
+
+.. automethod:: Modeldata.add_obstype
