+++ docs/reference/api/metobs_toolkit.Dataset.fill_missing_obs_linear.rst
@@ -1,4 +1,4 @@
-﻿metobs\_toolkit.Dataset.fill\_missing\_obs\_linear
+metobs\_toolkit.Dataset.fill\_missing\_obs\_linear
 ==================================================
 
 .. currentmodule:: metobs_toolkit
