+++ docs/reference/api/metobs_toolkit.Dataset.fill_gaps_automatic.rst
@@ -1,4 +1,4 @@
-﻿metobs\_toolkit.Dataset.fill\_gaps\_automatic
+metobs\_toolkit.Dataset.fill\_gaps\_automatic
 =============================================
 
 .. currentmodule:: metobs_toolkit
