+++ docs/reference/api/metobs_toolkit.Obstype.get_standard_unit.rst
@@ -1,4 +1,4 @@
-﻿metobs\_toolkit.Obstype.get\_standard\_unit
+metobs\_toolkit.Obstype.get\_standard\_unit
 ===========================================
 
 .. currentmodule:: metobs_toolkit
