+++ docs/reference/api/metobs_toolkit.analysis.Analysis.aggregate_df.rst
@@ -1,4 +1,4 @@
-﻿metobs\_toolkit.analysis.Analysis.aggregate\_df
+metobs\_toolkit.analysis.Analysis.aggregate\_df
 ===============================================
 
 .. currentmodule:: metobs_toolkit.analysis
