### 2024-02-13 vergauwenthomas/MetObs_toolkit#388 @Zeitsperre
Hey @vergauwenthomas

If there are a handful of developer-specific functions, I wouldn't add them to the user API. They should be automatically covered by the `automodule` for sphinx already, so a power user would probably be able to find these functions if they're looking for them.

If you really wanted to check some boxes, I would suggest having a developer section in your API and placing some of these functions in there. Here's an example of how we do this in a project I maintain: https://xclim.readthedocs.io/en/stable/api.html#modules-for-xclim-developers. Feel free to take some inspiration, but there's no need to perform this if the effort is high IMO.
