+++ docs/reference/api/metobs_toolkit.analysis.Analysis.subset_period.rst
@@ -1,6 +1,0 @@
-metobs\_toolkit.analysis.Analysis.subset\_period
-================================================
-
-.. currentmodule:: metobs_toolkit.analysis
-
-.. automethod:: Analysis.subset_period
