+++ docs/reference/api/metobs_toolkit.Dataset.add_new_unit.rst
@@ -1,6 +1,0 @@
-metobs\_toolkit.Dataset.add\_new\_unit
-======================================
-
-.. currentmodule:: metobs_toolkit
-
-.. automethod:: Dataset.add_new_unit
