+++ docs/reference/api/metobs_toolkit.ModelObstype_Vectorfield.get_mapped_datasets.rst
@@ -1,6 +1,0 @@
-metobs\_toolkit.ModelObstype\_Vectorfield.get\_mapped\_datasets
-===============================================================
-
-.. currentmodule:: metobs_toolkit
-
-.. automethod:: ModelObstype_Vectorfield.get_mapped_datasets
