+++ docs/reference/api/metobs_toolkit.Modeldata.set_model_from_csv.rst
@@ -1,6 +1,0 @@
-metobs\_toolkit.Modeldata.set\_model\_from\_csv
-===============================================
-
-.. currentmodule:: metobs_toolkit
-
-.. automethod:: Modeldata.set_model_from_csv
