+++ docs/reference/api/metobs_toolkit.Obstype.get_standard_unit.rst
@@ -1,6 +1,0 @@
-metobs\_toolkit.Obstype.get\_standard\_unit
-===========================================
-
-.. currentmodule:: metobs_toolkit
-
-.. automethod:: Obstype.get_standard_unit
