+++ docs/reference/api/metobs_toolkit.Dataset.rst
@@ -1,58 +1,0 @@
-metobs\_toolkit.Dataset
-=======================
-
-.. currentmodule:: metobs_toolkit
-
-.. autoclass:: Dataset
-
-
-   .. automethod:: __init__
-
-
-   .. rubric:: Methods
-
-   .. autosummary::
-
-      ~Dataset.__init__
-      ~Dataset.add_new_observationtype
-      ~Dataset.add_new_unit
-      ~Dataset.apply_buddy_check
-      ~Dataset.apply_quality_control
-      ~Dataset.apply_titan_buddy_check
-      ~Dataset.apply_titan_sct_resistant_check
-      ~Dataset.coarsen_time_resolution
-      ~Dataset.combine_all_to_obsspace
-      ~Dataset.fill_gaps_automatic
-      ~Dataset.fill_gaps_era5
-      ~Dataset.fill_gaps_linear
-      ~Dataset.fill_missing_obs_linear
-      ~Dataset.get_altitude
-      ~Dataset.get_analysis
-      ~Dataset.get_gaps_df
-      ~Dataset.get_gaps_info
-      ~Dataset.get_info
-      ~Dataset.get_landcover
-      ~Dataset.get_lcz
-      ~Dataset.get_missing_obs_info
-      ~Dataset.get_modeldata
-      ~Dataset.get_qc_stats
-      ~Dataset.get_station
-      ~Dataset.import_data_from_file
-      ~Dataset.import_dataset
-      ~Dataset.make_gee_plot
-      ~Dataset.make_geo_plot
-      ~Dataset.make_interactive_plot
-      ~Dataset.make_plot
-      ~Dataset.save_dataset
-      ~Dataset.show
-      ~Dataset.show_settings
-      ~Dataset.sync_observations
-      ~Dataset.update_default_name
-      ~Dataset.update_gap_and_missing_fill_settings
-      ~Dataset.update_gaps_and_missing_from_outliers
-      ~Dataset.update_outliersdf
-      ~Dataset.update_qc_settings
-      ~Dataset.update_settings
-      ~Dataset.update_timezone
-      ~Dataset.update_titan_qc_settings
-      ~Dataset.write_to_csv
