+++ docs/MetObs_for_developers.rst
@@ -1,0 +1,30 @@
+
+********************************************
+MetObs toolkit Documentation for developers
+********************************************
+Here you can find the documentation on all classes, functions, and methods in
+the MetObs toolkit
+
+
+
+.. toctree::
+   :maxdepth: 2
+
+   full_module_summaries/gaps
+   full_module_summaries/gap_filling
+   full_module_summaries/missingobs
+   full_module_summaries/dataset
+   full_module_summaries/analysis
+   full_module_summaries/data_import
+   full_module_summaries/df_helpers
+   full_module_summaries/geometry
+   full_module_summaries/landcover_functions
+   full_module_summaries/missingobs
+   full_module_summaries/modeldata
+   full_module_summaries/obstypes
+   full_module_summaries/plotting
+   full_module_summaries/printing
+   full_module_summaries/qc
+   full_module_summaries/settings
+   full_module_summaries/template_build
+   full_module_summaries/writing
