### 2024-02-14 46cda56
Dev documentation as modlist + add docexamples + doctest tests
