+++ metobs_toolkit/dataset.py
@@ -158,7 +158,7 @@ class Dataset:
         if (not self.metadf["lat"].isnull().all()) & (
             not self.metadf["lon"].isnull().all()
         ):
-            add_info += "    *Coordinates are available for all stations. \n"
+            add_info += "    *Coordinates are available for all stations."
 
         return (
             f"Dataset instance containing: \n \
@@ -325,6 +325,7 @@ class Dataset:
 
             # Print out details
             dataset.show()
+
         """
         logger.info("Show basic info of dataset.")
 
@@ -346,6 +347,25 @@ class Dataset:
         -------
         None.
 
+        Examples
+        --------
+        .. code-block:: python
+
+            import metobs_toolkit
+
+            # Import data into a Dataset
+            dataset = metobs_toolkit.Dataset()
+            dataset.update_settings(
+                        input_data_file=metobs_toolkit.demo_datafile,
+                        input_metadata_file=metobs_toolkit.demo_metadatafile,
+                        template_file=metobs_toolkit.demo_template,
+                        )
+
+            dataset.import_data_from_file()
+
+            # Print out details
+            dataset.get_info()
+
         """
         self.show(show_all_settings, max_disp_n_gaps)
 
@@ -363,6 +383,27 @@ class Dataset:
         Returns
         -------
         None.
+
+        Examples
+        --------
+        .. code-block:: python
+
+            import metobs_toolkit
+            import os
+
+            # Import data into a Dataset
+            dataset = metobs_toolkit.Dataset()
+            dataset.update_settings(
+                        input_data_file=metobs_toolkit.demo_datafile,
+                        input_metadata_file=metobs_toolkit.demo_metadatafile,
+                        template_file=metobs_toolkit.demo_template,
+                        )
+
+            dataset.import_data_from_file()
+
+            # Save dataset to a .pkl file
+            dataset.save_dataset(outputfolder=os.getcwd(),
+                                 filename='your_saved_dataset.pkl')
 
         """
         # check if outputfolder is known and exists
@@ -405,6 +446,20 @@ class Dataset:
         metobs_toolkit.Dataset
             The Dataset instance.
 
+        Examples
+        --------
+        .. code-block:: python
+
+            import metobs_toolkit
+            import os
+
+            # Initialize an empty Dataset
+            empty_dataset = metobs_toolkit.Dataset()
+
+            # Import the dataset
+            dataset=empty_dataset.import_dataset(folder_path=os.getcwd(),
+                                                 filename='your_saved_dataset.pkl')
+
         """
         # check if folder_path is known and exists
         if folder_path is None:
@@ -437,11 +492,30 @@ class Dataset:
 
         Parameters
         ----------
-        Obstype : metobs_toolkit.obstype.Obstype
+
             The new Obstype to add.
         Returns
         -------
         None.
+
+        Examples
+        --------
+        .. code-block:: python
+
+            >>> import metobs_toolkit
+            >>> co2_concentration = metobs_toolkit.Obstype(obsname='co2',
+            ...                                            std_unit='ppm')
+            >>> #add other units to it (if needed)
+            >>> co2_concentration.add_unit(unit_name='ppb',
+            ...                            conversion=['x / 1000'], #1 ppb = 0.001 ppm
+            ...                           )
+            >>> #Set a description
+            >>> co2_concentration.set_description(desc='The CO2 concentration measured at 2m above surface')
+            >>> #Add it to a Dataset
+            >>> dataset = metobs_toolkit.Dataset()
+            >>> dataset.add_new_observationtype(co2_concentration)
+            >>> dataset.obstypes
[…]
