+++ docs/index.rst
@@ -28,7 +28,6 @@ such as PostGIS.
    Specific topics <specific_topics>
    Examples <examples/index>
    Documentation <MetObs_documentation_flat>
-   For Developers <MetObs_for_developers>
    Contributing <topics/contributing_link.md>
 
 .. container:: button
