+++ docs/MetObs_documentation_flat.rst
@@ -14,3 +14,11 @@ the MetObs toolkit to be used by a user and developer.
   Obstype <reference/obstype>
   Analysis <reference/analysis>
   Modeldata <reference/modeldata>
+
+********************************************
+Modules for MetObs toolkit Developers
+********************************************
+A lot of the funtions and methods of the MetObs toolkit are not meant to be called
+by the users. For developers these functions and methods are important building blocks.
+
+The :ref:`modindex` contains all functions and methods present in the MetObs toolkit.
