+++ docs/index.rst
@@ -12,12 +12,19 @@ MetObs Toolkit is an open source project to make working with meteorological obs
 Description
 -----------
 
-The goal of GeoPandas is to make working with geospatial data in
-python easier.  It combines the capabilities of pandas and shapely,
-providing geospatial operations in pandas and a high-level interface
-to multiple geometries to shapely.  GeoPandas enables you to easily do
-operations in python that would otherwise require a spatial database
-such as PostGIS.
+The MetObs-toolkit is a Python package developed to facilitate the use of non-traditional meteorological observations.
+The package provides automated quality control (QC) techniques to identify and
+flag erroneous observations, and includes methods to fill data gaps.
+Additionally, the package offers tools for analyzing the data, e.g. linkage
+with popular land-use datasets is included, trough the use of the Google Earth Engine, such
+that microclimate effects can be investigated with the MetObs-toolkit.
+
+
+The MetObs-toolkit provides a comprehensive framework for scientists to process
+raw meteorological data for analysis by making intensive use of the `Pandas <https://pandas.pydata.org/>`_
+and `GeoPandas <https://geopandas.org/en/stable/>`_ functionalities.
+
+
 
 .. toctree::
    :hidden:
@@ -34,13 +41,6 @@ such as PostGIS.
 
     :doc:`Getting started <intro>` :doc:`Specific topics <specific_topics>`
     :doc:`Examples <examples/index>` :doc:`Documentation <MetObs_documentation_flat>`
-
-
-Team
---------
-This package is created and maintained by the `Atmospheric Physics group <https://www.ugent.be/we/physics-astronomy/en/research/atmophys>`_ . of the
-Ghent university in Belgium.
-
 
 
 
