### 2024-02-14 vergauwenthomas/MetObs_toolkit#388 @vergauwenthomas
Okay, that is clear @Zeitsperre !

Here is a snapshot of the current API documentation (in #434):

![Screenshot from 2024-02-14 17-08-39](https://github.com/vergauwenthomas/MetObs_toolkit/assets/82087298/c71ae225-6ce9-4474-ac77-c17079318c7e)

- The API doc for the users is flattened, and the methods are grouped per application-ish.
- For the API doc for the developers (i.g. all functions/methods/classes) i have tested two approaches:
  1. creating a separate tab, and autosummary all the modules of the package.
  2. referring to the *Module Index*, which is a (recursive) list of all modules --> classes/methods/functions. (see screenshot above).

I prefer method *ii*, since it is the easiest approach and all relevant info can be found for the developers. @Zeitsperre are you okay with method *ii*?

(Since I can only deploy one version of the documentation, i have to take screenshots of my local host)

PS. I am still working on #434 for spellcheck + adding examples in the docstrings for the most common methods. I plan to merge it tomorrow, so to resolve this issue.
