### 2024-02-14 vergauwenthomas/MetObs_toolkit#388 @Zeitsperre
@vergauwenthomas

That looks fantastic! Much easier for users/developers to explore the library internals!

Using the module index is perfectly fine as far as I'm concerned; An experienced developer will know where to look.
