### 2024-02-15 4a23fff
further work on docstring examples and tests
