+++ metobs_toolkit/dataset.py
@@ -788,22 +788,22 @@ class Dataset:
         --------
         .. code-block:: python
 
-            import metobs_toolkit
-
-            # Import data into a Dataset
-            dataset = metobs_toolkit.Dataset()
-            dataset.update_settings(
-                        input_data_file=metobs_toolkit.demo_datafile,
-                        input_metadata_file=metobs_toolkit.demo_metadatafile,
-                        template_file=metobs_toolkit.demo_template,
-                        )
-
-            dataset.import_data_from_file()
-
-            # Make plot
-            dataset.make_plot(stationnames=['vlinder02', 'vlinder16'],
-                              obstype='temp',
-                              colorby='label')
+            >>> import metobs_toolkit
+            >>>
+            >>> # Import data into a Dataset
+            >>> dataset = metobs_toolkit.Dataset()
+            >>> dataset.update_settings(
+            ...                         input_data_file=metobs_toolkit.demo_datafile,
+            ...                         input_metadata_file=metobs_toolkit.demo_metadatafile,
+            ...                         template_file=metobs_toolkit.demo_template,
+            ...                         )
+            >>> dataset.import_data_from_file()
+            >>>
+            >>> # Make plot
+            >>> dataset.make_plot(stationnames=['vlinder02', 'vlinder16'],
+            ...                   obstype='temp',
+            ...                   colorby='label')
+            <Axes: ...
 
         """
 
@@ -949,20 +949,19 @@ class Dataset:
         --------
         .. code-block:: python
 
-            import metobs_toolkit
-
-            # Import data into a Dataset
-            dataset = metobs_toolkit.Dataset()
-            dataset.update_settings(
-                        input_data_file=metobs_toolkit.demo_datafile,
-                        input_metadata_file=metobs_toolkit.demo_metadatafile,
-                        template_file=metobs_toolkit.demo_template,
-                        )
-
-            dataset.import_data_from_file()
-
-            # Make default interactive plot
-            dataset.make_interactive_plot()
+            >>> import metobs_toolkit
+            >>>
+            >>> # Import data into a Dataset
+            >>> dataset = metobs_toolkit.Dataset()
+            >>> dataset.update_settings(
+            ...                         input_data_file=metobs_toolkit.demo_datafile,
+            ...                         input_metadata_file=metobs_toolkit.demo_metadatafile,
+            ...                         template_file=metobs_toolkit.demo_template,
+            ...                         )
+            >>> dataset.import_data_from_file()
+            >>>
+            >>> # Make default interactive geospatial plot
+            >>> dataset.make_interactive_plot()
 
         """
         # Check if obstype is known
@@ -1133,20 +1132,20 @@ class Dataset:
         --------
         .. code-block:: python
 
-            import metobs_toolkit
-
-            # Import data into a Dataset
-            dataset = metobs_toolkit.Dataset()
-            dataset.update_settings(
-                        input_data_file=metobs_toolkit.demo_datafile,
-                        input_metadata_file=metobs_toolkit.demo_metadatafile,
-                        template_file=metobs_toolkit.demo_template,
-                        )
-
-            dataset.import_data_from_file()
-
-            # Make default interactive plot
-            dataset.make_geo_plot()
+            >>> import metobs_toolkit
+            >>>
+            >>> # Import data into a Dataset
+            >>> dataset = metobs_toolkit.Dataset()
+            >>> dataset.update_settings(
+            ...                         input_data_file=metobs_toolkit.demo_datafile,
+            ...                         input_metadata_file=metobs_toolkit.demo_metadatafile,
[…]
