+++ metobs_toolkit/dataset.py
@@ -311,20 +311,24 @@ class Dataset:
         --------
         .. code-block:: python
 
-            import metobs_toolkit
+            >>> import metobs_toolkit
+            >>>
+            >>> #Create your Dataset
+            >>> dataset = metobs_toolkit.Dataset() #empty Dataset
 
-            # Import data into a Dataset
-            dataset = metobs_toolkit.Dataset()
-            dataset.update_settings(
-                        input_data_file=metobs_toolkit.demo_datafile,
-                        input_metadata_file=metobs_toolkit.demo_metadatafile,
-                        template_file=metobs_toolkit.demo_template,
-                        )
-
-            dataset.import_data_from_file()
-
-            # Print out details
-            dataset.show()
+            >>>
+            >>> #Add observations to the Dataset
+            >>> dataset.update_settings(
+            ...                         input_data_file=metobs_toolkit.demo_datafile,
+            ...                         input_metadata_file=metobs_toolkit.demo_metadatafile,
+            ...                         template_file=metobs_toolkit.demo_template,
+            ...                         )
+            >>> dataset.import_data_from_file()
+            >>>
+            >>> # Print out details
+            >>> dataset.show()
+            --------  General ---------
+            ...
 
         """
         logger.info("Show basic info of dataset.")
@@ -351,20 +355,24 @@ class Dataset:
         --------
         .. code-block:: python
 
-            import metobs_toolkit
+            >>> import metobs_toolkit
+            >>>
+            >>> #Create your Dataset
+            >>> dataset = metobs_toolkit.Dataset() #empty Dataset
 
-            # Import data into a Dataset
-            dataset = metobs_toolkit.Dataset()
-            dataset.update_settings(
-                        input_data_file=metobs_toolkit.demo_datafile,
-                        input_metadata_file=metobs_toolkit.demo_metadatafile,
-                        template_file=metobs_toolkit.demo_template,
-                        )
-
-            dataset.import_data_from_file()
-
-            # Print out details
-            dataset.get_info()
+            >>>
+            >>> #Add observations to the Dataset
+            >>> dataset.update_settings(
+            ...                         input_data_file=metobs_toolkit.demo_datafile,
+            ...                         input_metadata_file=metobs_toolkit.demo_metadatafile,
+            ...                         template_file=metobs_toolkit.demo_template,
+            ...                         )
+            >>> dataset.import_data_from_file()
+            >>>
+            >>> # Print out details
+            >>> dataset.get_info()
+            --------  General ---------
+            ...
 
         """
         self.show(show_all_settings, max_disp_n_gaps)
@@ -388,22 +396,22 @@ class Dataset:
         --------
         .. code-block:: python
 
-            import metobs_toolkit
-            import os
-
-            # Import data into a Dataset
-            dataset = metobs_toolkit.Dataset()
-            dataset.update_settings(
-                        input_data_file=metobs_toolkit.demo_datafile,
-                        input_metadata_file=metobs_toolkit.demo_metadatafile,
-                        template_file=metobs_toolkit.demo_template,
-                        )
-
-            dataset.import_data_from_file()
-
-            # Save dataset to a .pkl file
-            dataset.save_dataset(outputfolder=os.getcwd(),
-                                 filename='your_saved_dataset.pkl')
+            >>> import metobs_toolkit
+            >>> import os
+            >>>
+            >>> # Import data into a Dataset
+            >>> dataset = metobs_toolkit.Dataset()
+            >>> dataset.update_settings(
+            ...            input_data_file=metobs_toolkit.demo_datafile,
+            ...            input_metadata_file=metobs_toolkit.demo_metadatafile,
+            ...            template_file=metobs_toolkit.demo_template)
+            >>>
+            >>> dataset.import_data_from_file()
+            >>>
[…]
