+++ metobs_toolkit/dataset.py
@@ -557,7 +557,7 @@ class Dataset:
             type. The expression is a (list of) strings with simple algebraic
             operations, where x represent the value in the new unit, and the
             result is the value in the standard unit. Two examples for
-            temperature (with a standard unit in Celcius):
+            temperature (with a standard unit in Celsius):
 
                 ["x - 273.15"] #if the new_unit is Kelvin
                 ["x-32.0", "x/1.8"] #if the new unit is Farenheit
@@ -683,7 +683,7 @@ class Dataset:
 
         logger.info(f"Extract {stationname} from dataset.")
 
-        # important: make shure all station attributes are of the same time as dataset.
+        # important: make sure all station attributes are of the same time as dataset.
         # so that all methods can be inherited.
 
         try:
@@ -950,7 +950,7 @@ class Dataset:
         Note
         -------
         The figure will only appear when this is runned in notebooks. If you do
-        not run this in a notebook, make shure to save the html file, and open it
+        not run this in a notebook, make sure to save the html file, and open it
         with a browser.
 
         Examples
@@ -1281,7 +1281,7 @@ class Dataset:
         Note
         ------
         When extracting large amounts of data, the timeseries data will be
-        writen to a file and saved on your google drive. In this case, you need
+        written to a file and saved on your google drive. In this case, you need
         to provide the Modeldata with the data using the .set_model_from_csv()
         method.
 
@@ -1384,10 +1384,10 @@ class Dataset:
             )
 
         print(
-            f"(When using the .set_model_from_csv() method, make shure the modelname of your Modeldata is {modelname})"
+            f"(When using the .set_model_from_csv() method, make sure the modelname of your Modeldata is {modelname})"
         )
         logger.info(
-            f"(When using the .set_model_from_csv() method, make shure the modelname of your Modeldata is {modelname})"
+            f"(When using the .set_model_from_csv() method, make sure the modelname of your Modeldata is {modelname})"
         )
         return Modl
 
@@ -2292,13 +2292,13 @@ class Dataset:
         A final qualty control label for each
         quality-controlled-observation type can be added in the outputfile.
 
-        The file will be writen to the outputfolder specified in the settings.
+        The file will be written to the outputfolder specified in the settings.
 
         Parameters
         ----------
         obstype : string, optional
             Specify an observation type to subset all observations to. If None,
-            all available observation types are writen to file. The default is
+            all available observation types are written to file. The default is
             None.
         filename : string, optional
             The name of the output csv file. If none, a standard-filename
@@ -2318,7 +2318,7 @@ class Dataset:
             interpret these records as gaps/missing outliers if True. Else these
             will be interpreted as outliers. The default is True.
         seperate_metadata_file : bool, optional
-            If true, the metadat is writen to a seperate file, else the metadata
+            If true, the metadat is written to a seperate file, else the metadata
             is merged to the observation in one file. The default is True.
         Returns
         -------
@@ -3609,18 +3609,18 @@ class Dataset:
 
     def sync_observations(
         self,
-        tollerance,
+        tolerance,
         verbose=True,
         _force_resolution_minutes=None,
         _drop_target_nan_dt=False,
     ):
         """Simplify and syncronize the observation timestamps.
 
-        To simplify the resolution (per station), a tollerance is use to shift timestamps. The tollerance indicates the
+        To simplify the resolution (per station), a tolerance is use to shift timestamps. The tolerance indicates the
         maximum translation in time that can be applied to an observation.
 
[…]
