### 2024-02-15 f7a53c8
bugfix in testscript
