### 2024-02-15 3ea5ec2
restructuring docs (similar as geopandas) (#434)
* WIP on restructuring docs (similar as geopandas) + flattening the API doctree

* Saved new PDF of paper

* bulk commit for doc reformat

* Saved new PDF of paper

* add developers api to doc

* Saved new PDF of paper

* Saved new PDF of paper

* Dev documentation as modlist + add docexamples + doctest tests

* Saved new PDF of paper

* further doc cleanup

* Saved new PDF of paper

* further work on docstring examples and tests

* Saved new PDF of paper

* ready for spellingcheck

* Saved new PDF of paper

* spelcheck

* Saved new PDF of paper
[…]
