+++ docs/MetObs_documentation_flat.rst
@@ -1,0 +1,24 @@
+********************************************
+MetObs toolkit API Documentation
+********************************************
+
+Here you can find the documentation on the classes, functions, and methods in
+the MetObs toolkit to be used by a user and developer.
+
+
+.. toctree::
+  :maxdepth: 2
+
+  Dataset <reference/dataset>
+  Station <reference/station>
+  Obstype <reference/obstype>
+  Analysis <reference/analysis>
+  Modeldata <reference/modeldata>
+
+********************************************
+Modules for MetObs toolkit Developers
+********************************************
+A lot of the funtions and methods of the MetObs toolkit are not meant to be called
+by the users. For developers these functions and methods are important building blocks.
+
+The :ref:`modindex` contains all functions and methods present in the MetObs toolkit.
