+++ docs/specific_topics.rst
@@ -1,0 +1,14 @@
+============================
+Specific topics
+============================
+
+
+.. toctree::
+   :maxdepth: 3
+
+   topics/template_mapping
+   topics/gee_authentication
+   topics/create_obstypes
+   topics/irregular_timestamps
+   topics/gui
+   paper/index
