### 2024-02-15 82a43fd
add pydata theme to doc group
