+++ pyproject.toml
@@ -4,10 +4,6 @@ requires = ["poetry-core"]
 
 [tool.poetry]
 authors = ["Thomas Vergauwen <thomas.vergauwen@meteo.be>"]
-description = "A Meteorological observations toolkit for scientists"
-documentation = "https://vergauwenthomas.github.io/MetObs_toolkit/"
-keywords = ["meteorology", "observations", "urban climate"]
-license = "LICENSE"
 classifiers = [
   "Development Status :: 4 - Beta",
   "Intended Audience :: Science/Research",
@@ -21,6 +17,10 @@ classifiers = [
   "Programming Language :: Python :: 3.12",
   "Topic :: Scientific/Engineering :: Atmospheric Science"
 ]
+description = "A Meteorological observations toolkit for scientists"
+documentation = "https://vergauwenthomas.github.io/MetObs_toolkit/"
+keywords = ["meteorology", "observations", "urban climate"]
+license = "LICENSE"
 maintainers = ["Thomas Vergauwen <thomas.vergauwen@meteo.be>"]
 name = "MetObs-toolkit"
 packages = [{include = "metobs_toolkit"}]
@@ -53,10 +53,11 @@ pre-commit = "^3.6"
 myst_parser = '^2.0.0'
 nbsphinx = '^0.9'
 pandoc = '^2.1'  # Check on PyPi (not in local conda env) + with system wide install
+# sphinx-rtd-theme = '^1.3.0'
+pydata-sphinx-theme = '^0.15.2'
 # Group of dep packages for building the documentation
 sphinx = '^7.2'
 sphinx-copybutton = '^0.5.1'
-sphinx-rtd-theme = '^1.3.0'
 
 [tool.poetry.group.titan.dependencies]
 # titanlib requires C-compilers, which are not by default present on windows.
