+++ pyproject.toml
@@ -58,6 +58,7 @@ pydata-sphinx-theme = '^0.15.2'
 # Group of dep packages for building the documentation
 sphinx = '^7.2'
 sphinx-copybutton = '^0.5.1'
+sphinx-rtd-theme = '^1.3.0'
 
 [tool.poetry.group.titan.dependencies]
 # titanlib requires C-compilers, which are not by default present on windows.
