### 2024-02-15 vergauwenthomas/MetObs_toolkit#388 @vergauwenthomas
> Hi @vergauwenthomas
>
> The documentation is looking much better, but the API documentation should be flattened a bit if possible, specifically here: https://vergauwenthomas.github.io/MetObs_toolkit/MetObs_documentation_full.html. Addressing this would be enough for me to check off the remaining `Documentation` checks in the review.

@Zeitsperre, The new (flatted + docexamples) documentation is [online](https://vergauwenthomas.github.io/MetObs_toolkit/)! I have added docstring examples for the methods of the `Dataset`, and plan to do the same for the other classes.

I will not close this issue, since i will address your other remarks later.
