+++ metobs_toolkit/gap_filling.py
@@ -20,98 +20,270 @@ logger = logging.getLogger(__name__)
 
 
 # =============================================================================
+# Helpers
+# =============================================================================
+def _add_diurnal_timestamps(df):
+    df["hour"] = df.index.hour
+    df["minutes"] = df.index.minute
+    df["seconds"] = df.index.second
+    return df
+
+
+# =============================================================================
+# Modeldata gapfillers
+# =============================================================================
+
+
+def _raw_modeldata_fill(filldf):
+    """Return series of raw modeldata for gap records"""
+    gapfill_df = filldf[filldf["fill_method"] == "gap"]
+
+    # Construct returns
+    fillvalues = gapfill_df["modelvalues"]
+    msg = gapfill_df.apply(
+        lambda x: f"Modelvalue: {x['modelvalues']:.2f}, without correction.", axis=1
+    )
+    return fillvalues, msg
+
+
+# dit nu overal toepassen (de msg return + opvangen in gap module)
+def _debias_modeldata_fill(filldf):
+    # calculate one bias for full anchor period
+    anchors = filldf[filldf["fill_method"] != "gap"]
+    bias = (anchors["modelvalues"] - anchors["obsvalues"]).mean()
+
+    # correct the modelvalues with a single bias value
+    gapfill = filldf[filldf["fill_method"] == "gap"]
+
+    # Construct returns
+    fillvalues = gapfill["modelvalues"] - bias
+    msg = gapfill.apply(
+        lambda x: f"Modelvalue: {x['modelvalues']:.2f} corrected with a {bias:.2f} bias.",
+        axis=1,
+    )
+
+    return fillvalues, msg
+
+
+def _diurnal_debias_modeldata_fill(filldf):
+    # subset to anchors and gap
+    anchordf = filldf[filldf["fill_method"] != "gap"]
+    gapdf = filldf[filldf["fill_method"] == "gap"]
+    anchordf = anchordf.reset_index().set_index("datetime")[
+        ["modelvalues", "obsvalues"]
+    ]
+    gapdf = gapdf.reset_index().set_index("datetime")[["modelvalues", "obsvalues"]]
+
+    # create diurnal record categories (hour, minutes, seconds)
+    anchordf = _add_diurnal_timestamps(anchordf)
+    gapdf = _add_diurnal_timestamps(gapdf)
+
+    # calculate instantanious biases
+    anchordf["diff"] = anchordf["modelvalues"] - anchordf["obsvalues"]
+    # calculate biases for dirunal records
+    diurnalbias = (
+        anchordf.groupby(["hour", "minutes", "seconds"]).agg("mean").reset_index()
+    )
+
+    # merge biases to the gapdf
+    gapdf = gapdf.reset_index().merge(
+        diurnalbias[["diff", "hour", "minutes", "seconds"]],
+        how="left",
+        on=["hour", "minutes", "seconds"],
+    )
+
+    # revert the multiindex
+    gapdf["name"] = filldf.index.get_level_values("name")[0]
+    gapdf = gapdf.set_index(["name", "datetime"])
+
+    # Construct returns
+    fillvalues = gapdf["modelvalues"] - gapdf["diff"]  # correct with diurnal biasses
+    msg = gapdf.apply(
+        lambda x: f"Modelvalue: {x['modelvalues']:.2f} corrected with a {x['diff']:.2f} bias.",
+        axis=1,
+    )
+
+    return fillvalues, msg
+
+
+def _weighted_diurnal_debias_modeldata(filldf):
+
+    # subset to leading, trailing and gap
+    leadingdf = filldf[filldf["fill_method"] == "leading period"]
+    trailingdf = filldf[filldf["fill_method"] == "trailing period"]
+    gapdf = filldf[filldf["fill_method"] == "gap"]
+
[…]
