+++ metobs_toolkit/dataset.py
@@ -11,6 +11,7 @@ import os
 import sys
 import copy
 from datetime import timedelta
+import itertools
 import pytz
 import logging
 import pandas as pd
@@ -59,19 +60,20 @@ from metobs_toolkit.qc_checks import (
 from metobs_toolkit.qc_statistics import get_freq_statistics
 from metobs_toolkit.writing_files import write_dataset_to_csv
 
-from metobs_toolkit.missingobs import Missingob_collection
+# from metobs_toolkit.missingobs import Missingob_collection
 
 from metobs_toolkit.gap import (
     Gap,
-    remove_gaps_from_obs,
-    remove_gaps_from_outliers,
-    missing_timestamp_and_gap_check,
-    get_gaps_indx_in_obs_space,
-    get_station_gaps,
-    apply_interpolate_gaps,
-    make_gapfill_df,
-    apply_debias_era5_gapfill,
-    gaps_to_df,
+    find_gaps,
+    # remove_gaps_from_obs,
+    # remove_gaps_from_outliers,
+    # missing_timestamp_and_gap_check,
+    # get_gaps_indx_in_obs_space,
+    # get_station_gaps,
+    # apply_interpolate_gaps,
+    # make_gapfill_df,
+    # apply_debias_era5_gapfill,
+    # gaps_to_df,
 )
 
 
@@ -167,7 +169,6 @@ class Dataset:
     *{n_obs_tot} observation records \n \
     *{n_outl} records labeled as outliers \n \
     *{len(self.gaps)} gaps \n \
-    *{self.missing_obs.series.shape[0]} missing observations \n \
     *records range: {startdt} --> {enddt} (total duration:  {enddt - startdt}) \n \
     *time zone of the records: {self.settings.time_settings['timezone']} \n "
             + add_info
@@ -3601,11 +3602,17 @@ class Dataset:
         # update df
         self.df = df
 
+        # Since the gap period is defined by a start and end, the gap-records are
+        # depending (which form the index of gapdf) depends on the observation frequency.
+        # Thus the gapdf attributs must be updated of all gaps
+        for gap in self.gaps:
+            gap._initiate_gapdf(Dataset=self)
+
         # Remove gaps and missing from the observatios
         # most gaps and missing are already removed but when increasing timeres,
         # some records should be removed as well.
-        self.df = remove_gaps_from_obs(gaplist=self.gaps, obsdf=self.df)
-        self.df = self.missing_obs.remove_missing_from_obs(obsdf=self.df)
+        # self.df = remove_gaps_from_obs(gaplist=self.gaps, obsdf=self.df)
+        # self.df = self.missing_obs.remove_missing_from_obs(obsdf=self.df)
 
     def sync_observations(
         self,
@@ -4191,9 +4198,52 @@ station with the default name: {self.settings.app["default_name"]}.'
 
         self.metadf["dataset_resolution"] = freq_series
 
+        # initialize the gapdf attributes
+        for gap in self.gaps:
+            gap._initiate_gapdf(Dataset=self)
+
         # Remove gaps and missing from the observations AFTER timecoarsening
-        self.df = remove_gaps_from_obs(gaplist=self.gaps, obsdf=self.df)
-        self.df = self.missing_obs.remove_missing_from_obs(obsdf=self.df)
+
+    #     self.df = remove_gaps_from_obs(gaplist=self.gaps, obsdf=self.df)
+    #     self.df = self.missing_obs.remove_missing_from_obs(obsdf=self.df)
+
+    # def _remove_gaps_from_obs(gaplist, obsdf):
+    #     """
+    #     Remove station - datetime records that are in the gaps from the obsdf.
+
+    #     (Usefull when filling timestamps to a df, and if you whant to remove the
+    #       gaps.)
+
+    #     Parameters
+    #     ----------
[…]

[…]
