### 2024-02-20 #5916 @Zeitsperre
Hi @vergauwenthomas

I've confirmed that the Google Earth Engine features do indeed function. I did notice however that some of the steps I needed to follow appear to be different from what is mentioned in the documentation, specifically:
- When running the [`dataset.get_lcz()`](https://vergauwenthomas.github.io/MetObs_toolkit/topics/gee_authentication.html#test-your-gee-access) step, there doesn't seem to be a way to easily specify that the compute time should be associated with a certain project. If this is only true for a project with financial information on file, feel free to ignore (the dataset download works, and usage was accurately tracked to my account)

The only other comments I have are:
- The [introduction exercise](https://vergauwenthomas.github.io/MetObs_toolkit/examples/doc_example.html#Introduction-exercise) link at the bottom appears to be broken.
- The paper is great. There are a handful of very small typos that I'll address in a PR right now.

Beyond that, the documentation was very easy to follow, examples are solid and confirmed working, and the new API layout is great! Very impressive library.
