### 2024-02-21 4d5a359
Update paper.md (#438)
Minor grammar and typo fixes
