+++ metobs_toolkit/gap.py
@@ -135,6 +135,9 @@ class Gap:
         print("---- Gap Data Frame -----")
         print(self.gapdf)
 
+    # =============================================================================
+    # Helpers
+    # =============================================================================
     def _get_gapfill_status(self):
         if not self.does_gap_holds_missing_records():
             return "Gap does not exist in observation space"
@@ -175,6 +178,10 @@ class Gap:
         self.gapdf[f"{self.obstype.name}_fill"] = np.nan  # fill with nan values
         self.gapdf["fill_method"] = None  # fill with None
         self.gapdf["msg"] = None  # Nonedataset
+
+    # =============================================================================
+    # Gap --- Records methods
+    # =============================================================================
 
     def get_missing_records(self, Dataset):
         """Find the missing records of a gap in observation-space.
@@ -623,6 +630,7 @@ class Gap:
         max_trail_to_gap_distance=None,
     ):
 
+        obsname = self.obstype.name
         # 1. Get leading and trailing info
         # get leading record, check validity and add to the gapfilldf
         (_, lead_dt, lead_val, lead_msg) = self.get_leading_record(
@@ -640,24 +648,53 @@ class Gap:
 
         anchor_df = pd.DataFrame(
             data={
-                self.obstype.name: [lead_val, trail_val],
+                obsname: [lead_val, trail_val],
                 "fill_method": ["leading", "trailing"],
                 "msg": [lead_msg, trail_msg],
             },
             index=idx,
         )
+        # Update attributes
         self.anchordf = anchor_df
+        self.gapdf["fill_method"] = f"{method}-interpolation"
+
+        # check if gapfill can proceed
+        if (lead_msg != "ok") & (trail_msg != "ok"):
+            logger.warning(
+                f"Cannot fill {self}, because leading and trailing records are not valid."
+            )
+            print(
+                f"Warning! Cannot fill {self}, because leading and trailing periods are not valid."
+            )
+
+            self.gapdf[f"{obsname}_fill"] = np.nan
+            self.gapdf["msg"] = f"{lead_msg} and {trail_msg}"
+            return
+        elif (lead_msg != "ok") & (trail_msg == "ok"):
+            logger.warning(f"Cannot fill {self}, because leading record is not valid.")
+            print(f"Warning! Cannot fill {self}, because leading record is not valid.")
+
+            self.gapdf[f"{obsname}_fill"] = np.nan
+            self.gapdf["msg"] = f"{lead_msg}"
+            return
+        elif (lead_msg == "ok") & (trail_msg != "ok"):
+            logger.warning(f"Cannot fill {self}, because trailing record is not valid.")
+            print(f"Warning! Cannot fill {self}, because trailing record is not valid.")
+
+            self.gapdf[f"{obsname}_fill"] = np.nan
+            self.gapdf["msg"] = f"{trail_msg}"
+            return
+        else:
+            pass
 
         # 3. Combine the anchors with the observations
-        tofilldf = pd.concat(
-            [self.gapdf[[self.obstype.name]], self.anchordf[[self.obstype.name]]]
-        )
+        tofilldf = pd.concat([self.gapdf[[obsname]], self.anchordf[[obsname]]])
 
         # 4. Replace the NaN's (GAPFILLING)
         # make shure only datetimes are in the index, and sorted by datetimes
         tofilldf = tofilldf.reset_index().set_index("datetime").sort_index()
         # Interpolate series
-        tofilldf[self.obstype.name].interpolate(
+        tofilldf[obsname].interpolate(
             method=method,
[…]
