+++ metobs_toolkit/dataset.py
@@ -65,6 +65,7 @@ from metobs_toolkit.writing_files import write_dataset_to_csv
 from metobs_toolkit.gap import (
     Gap,
     find_gaps,
+    create_gaps_overview_df,
     # remove_gaps_from_obs,
     # remove_gaps_from_outliers,
     # missing_timestamp_and_gap_check,
@@ -123,8 +124,8 @@ class Dataset:
         self.missing_obs = None  # becomes a Missingob_collection after import
         self.gaps = None  # becomes a list of gaps
 
-        self.gapfilldf = init_multiindexdf()
-        self.missing_fill_df = init_multiindexdf()
+        # self.gapfilldf = init_multiindexdf()
+        # self.missing_fill_df = init_multiindexdf()
 
         # Dataset with metadata (static)
         self.metadf = pd.DataFrame()
@@ -702,29 +703,16 @@ class Dataset:
         except KeyError:
             sta_outliers = init_triple_multiindexdf()
 
-        sta_gaps = get_station_gaps(self.gaps, stationname)
-        sta_missingobs = self.missing_obs.get_station_missingobs(stationname)
-
-        try:
-            sta_gapfill = self.gapfilldf.xs(stationname, level="name", drop_level=False)
-        except KeyError:
-            sta_gapfill = init_multiindexdf()
-
-        try:
-            sta_missingfill = self.missing_fill_df.xs(
-                stationname, level="name", drop_level=False
-            )
-        except KeyError:
-            sta_missingfill = init_multiindexdf()
+        sta_gaps = [gap for gap in self.gaps if gap.name == stationname]
 
         return Station(
             name=stationname,
             df=sta_df,
             outliersdf=sta_outliers,
             gaps=sta_gaps,
-            missing_obs=sta_missingobs,
-            gapfilldf=sta_gapfill,
-            missing_fill_df=sta_missingfill,
+            # missing_obs=sta_missingobs,
+            # gapfilldf=sta_gapfill,
+            # missing_fill_df=sta_missingfill,
             metadf=sta_metadf,
             obstypes=self.obstypes,
             data_template=self.data_template,
@@ -1706,7 +1694,7 @@ class Dataset:
         comb_df = concat_save([filldf_interp, filldf_debias])
 
         # update attr
-        self.gapfilldf = comb_df
+        # self.gapfilldf = comb_df
 
         return comb_df
 
@@ -1827,12 +1815,12 @@ class Dataset:
         )
 
         # get gapfilldf
-        gapfilldf = make_gapfill_df(self.gaps)
+        # gapfilldf = make_gapfill_df(self.gaps)
 
         # update attr
-        self.gapfilldf = gapfilldf
+        # self.gapfilldf = gapfilldf
 
-        return gapfilldf
+        # return gapfilldf
 
     def fill_missing_obs_linear(self, obstype="temp"):
         """Interpolate missing observations.
@@ -2269,7 +2257,7 @@ class Dataset:
             sys.exit(f"{method} not implemented yet")
 
         # update attribute
-        self.gapfilldf = filldf
+        # self.gapfilldf = filldf
 
         return filldf
 
@@ -3262,32 +3250,54 @@ class Dataset:
         # =============================================================================
         # Stack gaps
         # =============================================================================
-        # add gapfill and remove the filled records from gaps
-        gapsfilldf = self.gapfilldf.copy()
 
-        # to triple index
-        gapsfilldf = value_labeled_doubleidxdf_to_triple_idxdf(
-            gapsfilldf, known_obstypes=list(self.obstypes.keys())
-        )
-        gapsfilldf["toolkit_representation"] = "gap fill"
+        gapsdf = create_gaps_overview_df(gapslist=self.gaps)
+        # Format columns
[…]
