+++ docs/topics/gee_authentication.rst
@@ -1,3 +1,5 @@
+.. _Using_gee:
+
 ***************************
 Using Google Earth Engine
 ***************************
