+++ setup.cfg
@@ -4,7 +4,8 @@ exclude =
 	docs,
 	build,
 	.eggs,
-	tests
+	tests,
+	docs/examples
 #ignore =
 #	AZ100,
 #	AZ200,
