+++ setup.cfg
@@ -5,7 +5,7 @@ exclude =
 	build,
 	.eggs,
 	tests,
-	docs/examples
+	docs/examples,
 #ignore =
 #	AZ100,
 #	AZ200,
