+++ setup.cfg
@@ -5,7 +5,7 @@ exclude =
 	build,
 	.eggs,
 	tests,
-	docs/examples,
+	docs/examples, #blankklines in md cells before itimization are removed ruining the itemize format, thus skip them.
 #ignore =
 #	AZ100,
 #	AZ200,
