### 2024-02-21 d53a4e7
Fix broken links in doc (#440)
* fix links and code references in documentation

* add _static css formatters to repo for building the docs

* Saved new PDF of paper

* blankline before itemize

* fix itemize format

* add comma

* Saved new PDF of paper

* fix review

* Saved new PDF of paper

---------

Co-authored-by: vergauwenthomas <vergauwenthomas@users.noreply.github.com>
