+++ setup.cfg
@@ -4,7 +4,8 @@ exclude =
 	docs,
 	build,
 	.eggs,
-	tests
+	tests,
+	docs/examples, #blankklines in md cells before itimization are removed ruining the itemize format, thus skip them.
 #ignore =
 #	AZ100,
 #	AZ200,
