### 2024-02-21 1c60587
trigger only when paper has been adapted (#441)
* trigger only when paper has been adapted

* trigger paperchange

* revert trigger in paper

* Saved new PDF of paper

---------

Co-authored-by: vergauwenthomas <vergauwenthomas@users.noreply.github.com>
