### 2024-02-21 vergauwenthomas/MetObs_toolkit#442 @ashwinvis
[JOSS] GEE: onboarding issues
## Possibly typos / mistakes in the docs

*My comments in italics*.

https://vergauwenthomas.github.io/MetObs_toolkit/topics/gee_authentication.html#setup-a-cloud-project-on-your-developer-account

- The API’s that are used by the toolkit has quite a lot of **free credentials**

_Free credits?_

_Also are you implying that we need to rely on the $300 USD free credits that new users get, or is it that the API itself is free, or is it that the toolkit supplies the necessary credentials to use this API? I am confused._

https://vergauwenthomas.github.io/MetObs_toolkit/topics/gee_authentication.html#enable-api-s-on-your-project

- Search for the ‘Google Earth Engine API’, **click on it –> hit ENABLE**

_There is no such ENABLE button to click. It says API ENABLED, so perhaps they automatically enabled it?_

![image](https://github.com/vergauwenthomas/MetObs_toolkit/assets/9155111/a65c3dcd-1877-45f6-bd36-79702ae893fd)

https://vergauwenthomas.github.io/MetObs_toolkit/topics/gee_authentication.html#test-your-gee-access

_This seems to imply that a dependency was missing from the installation instructions, but the only thing I had to do was to upgrade `pip install earthengine-api --upgrade`__

```py
>>> dataset.get_lcz()
Fetching credentials using gcloud
FileNotFoundError                         Traceback (most recent call last)
File ~/test-metobs/lib/python3.11/site-packages/ee/oauth.py:317, in _load_app_default_credentials(run_gcloud, scopes, quiet)
    316 try:
--> 317   subprocess.run(command, check=True)
    318 except FileNotFoundError as e:

...
File ~/test-metobs/lib/python3.11/site-packages/ee/oauth.py:320, in _load_app_default_credentials(run_gcloud, scopes, quiet)
    318 except FileNotFoundError as e:
    319   tip = 'Please ensure that gcloud is installed.' + more_info
--> 320   raise Exception('gcloud command not found. ' + tip) from e  # pylint:disable=broad-exception-raised
    321 except subprocess.CalledProcessError as e:
    322   tip = ('Please check for any errors above.\n*Possible fixes:'
    323          ' If you loaded a page with a "redirect_uri_mismatch" error,'
    324          ' run earthengine authenticate with the --quiet flag;'
    325          ' if the error page says "invalid_request", be sure to run the'
    326          ' entire gcloud auth command that is shown.' + more_info)

Exception: gcloud command not found. Please ensure that gcloud is installed.
More information: https://developers.google.com/earth-engine/guides/python_install
```

And then upon upgrading, and following the authentication link

```py
~/test-metobs/lib/python3.11/site-packages/metobs_toolkit/dataset.py:2886, in Dataset.get_lcz(self)
   2872 """Extract local climate zones for all stations.
   2873
   2874 Function to extract the Local CLimate zones (LCZ) from the
   (...)
   2883
   2884 """
   2885 # connect to gee
-> 2886 connect_to_gee()
   2888 # Extract LCZ for all stations
   2889 lcz_series = lcz_extractor(
   2890     metadf=self.metadf,
   2891     mapinfo=self.settings.gee["gee_dataset_info"]["global_lcz_map"],
   2892 )

...

File ~/test-metobs/lib/python3.11/site-packages/ee/data.py:404, in _execute_cloud_call(call, num_retries)
    402   return call.execute(num_retries=num_retries)
    403 except googleapiclient.errors.HttpError as e:
--> 404   raise _translate_cloud_exception(e)

EEException: Not signed up for Earth Engine or project is not registered. Visit https://developers.google.com/earth-engine/guides/access
```

Right now I am stuck here. How do I fix this?
