### 2024-02-21 #5916 @ashwinvis
Apologies for not replying sooner. I finally got around trying GEE and ran into a few problems with the process:

https://github.com/vergauwenthomas/MetObs_toolkit/issues/442
