### 2024-02-21 vergauwenthomas/MetObs_toolkit#442 @ashwinvis
I tried to move `~/.config/earthengine/credentials` file to redo the process but I end up with the same EEException again. Strange, since I can see a project with the API enabled:

![image](https://github.com/vergauwenthomas/MetObs_toolkit/assets/9155111/5224812e-83f4-4af6-b12f-20a351d4b008)
