### 2024-02-21 vergauwenthomas/MetObs_toolkit#442 @ashwinvis
```
EEException: Not signed up for Earth Engine or project is not registered. Visit https://developers.google.com/earth-engine/guides/access
```

Same issue as before :tired_face:
