### 2024-02-21 vergauwenthomas/MetObs_toolkit#442 @vergauwenthomas
Did you manage to enable the google earth engine (red arrow) and then you need to register it (green arrow) ?

![Screenshot from 2024-02-21 16-24-17](https://github.com/vergauwenthomas/MetObs_toolkit/assets/82087298/5c9367e2-4678-46b9-ac26-8ad4c6d2a539)
