### 2024-02-21 vergauwenthomas/MetObs_toolkit#442 @vergauwenthomas
Could you click on this link: [code.earthengine.google.com/client-auth](https://code.earthengine.google.com/client-auth)

This link should bring you to an authentication window, where you can select the project to use the API for.
