### 2024-02-21 vergauwenthomas/MetObs_toolkit#442 @ashwinvis
> Did you manage to enable the google earth engine (red arrow)

Not entirely sure, but I guess so.

> and then you need to register it (green arrow) ?

Yes

![image](https://github.com/vergauwenthomas/MetObs_toolkit/assets/9155111/1ddc3d36-af1f-4ebb-a7fe-d3602c6f2994)
