### 2024-02-21 vergauwenthomas/MetObs_toolkit#442 @ashwinvis
> Could you click on this link: [code.earthengine.google.com/client-auth](https://code.earthengine.google.com/client-auth)
>
> This link should bring you to an authentication window, where you can select the project to use the API for.

I can see it, but if I click on GENERATE TOKEN, I get "Invalid request"

![image](https://github.com/vergauwenthomas/MetObs_toolkit/assets/9155111/837a667e-35b6-4f2c-904e-4a30708e4dc6)
