### 2024-02-22 vergauwenthomas/MetObs_toolkit#442 @vergauwenthomas
Hi @ashwinvis, i made a short instruction video on how to set up the project. You can find it here: https://youtu.be/h_8kX_oJulw

To test the API and your project, i used this authentication method:

```
import ee
ee.Authenticate(force=True) #force to create new credentials
ee.Initialize()
```

Can you reproduce these steps and let me know if this works, or where is does not work?
