### 2024-02-23 vergauwenthomas/MetObs_toolkit#442 @ashwinvis
I see what went wrong. I was always trying from IPython console, and then authentication process is very different from the Notebook Authenticator. If I try that method, the `~/.config/earthengine/credentials` which gets saved, contains no `client_id`  and `client_secret` fields.

Now that I tried it via JupyterLab a correct credentials file is saved.
