### 2024-02-24 vergauwenthomas/MetObs_toolkit#442 @ashwinvis
I think you either try to do exception handling and / or note this known issue in the documentation. Recommend to use Jupyter instead.
