### 2024-02-26 d915f4f
refactoring gap defenition
