+++ metobs_toolkit/gap.py
@@ -226,7 +226,9 @@ class Gap:
     #  Gapfill anchor methods
     # =============================================================================
     def get_leading_period(
-        self, Dataset, leading_period_duration="24H", min_leading_records_total=60
+        self,
+        Dataset,
+        leading_period_duration="24H",
     ):
         """TODO update docstring
 
@@ -248,7 +250,7 @@ class Gap:
 
         Returns
         -------
-        tuple : (name, timestamp, obs-value, status-message)
+        tuple : (name, timestamp, obs-value)
             All information on the leading record stored in a tuple.
 
         """
@@ -268,19 +270,7 @@ class Gap:
                     self.name,
                     pd.DatetimeIndex([]),
                     np.nan,
-                    "no leading record candidates could be found.",
-                ]
-            )
-
-        elif (
-            leading_period.shape[0] < min_leading_records_total
-        ):  # not enouggh (total) records found
-            return tuple(
-                [
-                    self.name,
-                    leading_period,
-                    sta_obs.loc[leading_period][self.obstype.name].to_list(),
-                    f"minimum records ({min_leading_records_total}) for leading period not met.",
+                    # "no leading record candidates could be found.",
                 ]
             )
 
@@ -290,13 +280,11 @@ class Gap:
                     self.name,
                     leading_period,
                     sta_obs.loc[leading_period][self.obstype.name].to_list(),
-                    "ok",
+                    # "ok",
                 ]
             )
 
-    def get_trailing_period(
-        self, Dataset, trailing_period_duration="24H", min_trailing_records_total=60
-    ):
+    def get_trailing_period(self, Dataset, trailing_period_duration="24H"):
         """TODO update docstring
 
 
@@ -317,7 +305,7 @@ class Gap:
 
         Returns
         -------
-        tuple : (name, timestamp, obs-value, status-message)
+        tuple : (name, timestamp, obs-value)
             All information on the leading record stored in a tuple.
 
         """
@@ -337,19 +325,7 @@ class Gap:
                     self.name,
                     pd.DatetimeIndex([]),
                     [],
-                    "no trailing record candidates could be found.",
-                ]
-            )
-
-        elif (
-            trailing_period.shape[0] < min_trailing_records_total
-        ):  # not enouggh (total) records found
-            return tuple(
-                [
-                    self.name,
-                    trailing_period,
-                    sta_obs.loc[trailing_period][self.obstype.name].to_list(),
-                    f"minimum records ({min_trailing_records_total}) for trailing period not met.",
+                    # "no trailing record candidates could be found.",
                 ]
             )
 
@@ -359,7 +335,7 @@ class Gap:
                     self.name,
                     trailing_period,
                     sta_obs.loc[trailing_period][self.obstype.name].to_list(),
-                    "ok",
+                    # "ok",
                 ]
             )
 
@@ -471,124 +447,8 @@ class Gap:
     # Fill methods
     # =============================================================================
 
-    def model_gapfill(
-        self,
-        Dataset,
-        Modeldata,
-        method="debias_modeldata",
-        leading_period_duration="24H",
-        min_leading_records_total=60,
-        trailing_period_duration="24H",
-        min_trailing_records_total=60,
-    ):
-
-        # check attributes
-        assert method in [
-            "raw_modeldata",
[…]
