+++ metobs_toolkit/dataset.py
@@ -50,7 +50,7 @@ from metobs_toolkit.qc_checks import (
     duplicate_timestamp_check,
     step_check,
     window_variation_check,
-    invalid_input_check,
+    # invalid_input_check,
     toolkit_buddy_check,
     titan_buddy_check,
     titan_sct_resistant_check,
@@ -121,7 +121,7 @@ class Dataset:
         # Dataset with outlier observations
         self.outliersdf = init_triple_multiindexdf()
 
-        self.missing_obs = None  # becomes a Missingob_collection after import
+        # self.missing_obs = None  # becomes a Missingob_collection after import
         self.gaps = None  # becomes a list of gaps
 
         # self.gapfilldf = init_multiindexdf()
@@ -220,12 +220,6 @@ class Dataset:
         # their is a coarsening allied on either of them.
         new.gaps = []
 
-        # ---------- missing ---------
-        # Missing observations have to be recaluculated using a frequency assumtion from the
-        # combination of self.df and other.df, thus NOT the native frequency if
-        # their is a coarsening allied on either of them.
-        new.missing_obs = None
-
         # ---------- metadf -----------
         # Use the metadf from self and add new rows if they are present in other
         new.metadf = concat_save([self.metadf, other.metadf])
@@ -257,15 +251,7 @@ class Dataset:
         # that are dependand on the frequency (since the freq of the .df is used,
         # which is not the naitive frequency if coarsening is applied on either. )
 
-        # missing and gap check
-        if gapsize is None:
-            gapsize = new.settings.gap["gaps_settings"]["gaps_finder"]["gapsize_n"]
-
-        # note gapsize is now defined on the frequency of self
-        new.missing_obs, new.gaps = missing_timestamp_and_gap_check(
-            df=new.df,
-            gapsize_n=self.settings.gap["gaps_settings"]["gaps_finder"]["gapsize_n"],
-        )
+        new.gaps = find_gaps(new.df, known_obstypes=new.obstypes)
 
         # duplicate check
         new.df, dup_outl_df = duplicate_timestamp_check(
@@ -1380,160 +1366,202 @@ class Dataset:
         )
         return Modl
 
-    def update_gaps_and_missing_from_outliers(self, obstype="temp", n_gapsize=None):
-        """Interpret the outliers as missing observations.
+    def convert_outliers_to_gaps(self):
+        # TODO : docstrinc + alles van de update gaps ... vervangend door deze methode
 
-        If there is a sequence
-        of these outliers for a station, larger than n_gapsize than this will
-        be interpreted as a gap.
+        if self.outliersdf.empty:
+            print("Warning! No outliers are found to convert!")
+            return
 
-        The outliers are not removed.
+        if bool(self.gaps):
+            print("Warning! The current gaps will be removed and new gaps are formed!")
 
-        Parameters
-        ----------
-        obstype : str, optional
-            Use the outliers on this observation type to update the gaps and
-            missing timestamps. The default is 'temp'.
-        n_gapsize : int, optional
-            The minimum number of consecutive missing observations to define
-            as a gap. If None, n_gapsize is taken from the settings defenition
-            of gaps. The default is None.
+        if (
+            self.metadf["assumed_import_frequency"] != self.metadf["dataset_resolution"]
+        ).any():
+            print(
+                "Warning! The current gaps are defined at a resolution different from the assumed import resolution!"
+            )
 
-        Returns
-        -------
-        None.
+        # Get start and end of timeseries (be shure to include gaps and outliers to determine start en endpoints)
+        combdf = self.combine_all_to_obsspace()
 
-        Note
-        -------
-        Gaps and missing observations resulting from an outlier on a specific
[…]

[…]
