### 2024-02-26 6cbd18d
pass kwargs to gee authenticate and update docstring
