### 2024-02-26 vergauwenthomas/MetObs_toolkit#442 @vergauwenthomas
@ashwinvis, This issue is addressed in #444 where
- The user can now specify which authentication method is used. (Which solves the invalid credential file + the `gcloud` dependency)
- The authentication function is callable for the user (`metobs_toolkit.connect_to_gee()`) with kwargs passed to the `ee.Authenticate()`.
- This issue, and proposed solution is addressed in the documentation in the *Using google earth engine* page as a warning banner (see screenshot)
- Typo/explenation on free credits is fixed in the documentation.

![Screenshot from 2024-02-26 13-21-23](https://github.com/vergauwenthomas/MetObs_toolkit/assets/82087298/ebae659d-a581-4577-be5d-5c3090036b33)

@ashwinvis let me know if you agree that #444 solves this issue, so i can merge.
