### 2024-02-27 #5916 @ashwinvis
@vergauwenthomas I am now fairly confident that the software works as you have described.

I have a few requests regarding the article, particularly the section Technical implementation.

## Further reading and citations needed here

Since MetObs-Toolkit implements quite a few methods, it would be important to link to other sources which may describe them in detail. If you think that these methods might be subject to change, it might be sufficient to cite the references in the online documentation and link this software paper to the documentation instead. See how [scikit-image](https://scikit-image.org/docs/stable/api/skimage.filters.html#skimage.filters.butterworth)  does it.

> “55 to examine data types, irregular timestamps, max-min thresholds, repetitions criteria, spike tests, allowed variation in time windows and spatial tests” ([Vergauwen et al., 2024, p. 3](zotero://select/library/items/7QB9C9DM)) ([pdf](zotero://open-pdf/library/items/KNJNTHL2?page=3&annotation=E9BL426K))

These, IMHO, are quite involved methods. What are the formal basis behind these QC methods or are they based on intuition? Some reuses TITAN (titanlib) while the rest are implemented from the scratch (which means citing TITAN alone will not be sufficient). .

> “59 ... gap filling is applied by using interpolation methods” ([Vergauwen et al., 2024, p. 3](zotero://select/library/items/7QB9C9DM)) ([pdf](zotero://open-pdf/library/items/KNJNTHL2?page=3&annotation=276GQ5F3))

Gap filling and bias correction is another involved topic. What is the basis behind the methods used? Further reading and citations needed here.

## Minor corrections

> “48 ... data for analysis by making intensive use of the pandas (The pandas development team, 2023) and geopandas (Jordahl et al., 2020) functionalities” ([Vergauwen et al., 2024, p. 2](zotero://select/library/items/7QB9C9DM)) ([pdf](zotero://open-pdf/library/items/KNJNTHL2?page=2&annotation=YDVDSJND))

*use of the functionalities from pandas (..) and geopandas (..) libraries.

> “52 data is imported into the Toolkit Dataset” ([Vergauwen et al., 2024, p. 3](zotero://select/library/items/7QB9C9DM)) ([pdf](zotero://open-pdf/library/items/KNJNTHL2?page=3&annotation=S5WAVRH5))

into a Toolkit Dataset class instance

>“61 Modeldata” ([Vergauwen et al., 2024, p. 3](zotero://select/library/items/7QB9C9DM)) ([pdf](zotero://open-pdf/library/items/KNJNTHL2?page=3&annotation=3XCPJ3CZ))

Modeldata class instance

Finally could you make the typography consistent between
- Toolkit Dataset
- Toolkit `Modeldata`
- `Toolkit-Analysis`
?
