### 2024-02-27 59590a7
Document/fix the gee authentication issue  (#444)
* pass kwargs to gee authenticate and update docstring

* update the documentation

* fix identation
