### 2024-02-28 b4dbd8b
init for rtd hosting of the docs
