### 2024-02-28 e8eaeea
add poetry install in yaml
