+++ .readthedocs.yaml
@@ -3,21 +3,23 @@
 # See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
 
 # # Required
-# version: 2
+version: 2
 
 # # Set the version of Python and other tools you might need
-# build:
-#   os: ubuntu-22.04
-#   tools: {python: "3.10"}
-#   jobs:
-#     pre_create_environment:
-#       - asdf plugin add poetry
-#       - asdf install poetry latest
-#       - asdf global poetry latest
-#       - poetry config virtualenvs.create false
-#       - poetry build
-#     post_install:
-#       - poetry install
+build:
+  os: ubuntu-22.04
+  tools: {python: "3.10"}
+  jobs:
+    pre_create_environment:
+      - asdf plugin add poetry
+      - asdf install poetry latest
+      - asdf global poetry latest
+      - poetry config virtualenvs.create false
+      - poetry build
+      - poetry update
+      - poetry install --all-extras
+    post_install:
+      - poetry install
 
 # # Build documentation in the docs/ directory with Sphinx
 # sphinx:
@@ -30,13 +32,13 @@
 # See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
 
 # Required
-version: 2
+#version: 2
 
 # Set the OS, Python version and other tools you might need
-build:
-  os: ubuntu-22.04
-  tools:
-    python: "3.12"
+# build:
+#   os: ubuntu-22.04
+#   tools:
+#     python: "3.12"
     # You can also specify other tool versions:
     # nodejs: "19"
     # rust: "1.64"
