### 2024-02-28 4eb40d8
get_duration method implemented
