+++ docs/doc_environment.yml
@@ -1,0 +1,89 @@
+name: doctest
+channels:
+  - conda-forge
+  - defaults
+dependencies:
+  - _libgcc_mutex=0.1=conda_forge
+  - _openmp_mutex=4.5=2_gnu
+  - bzip2=1.0.8=hd590300_5
+  - ca-certificates=2024.2.2=hbcca054_0
+  - ld_impl_linux-64=2.40=h41732ed_0
+  - libexpat=2.5.0=hcb278e6_1
+  - libffi=3.4.2=h7f98852_5
+  - libgcc-ng=13.2.0=h807b86a_5
+  - libgomp=13.2.0=h807b86a_5
+  - libnsl=2.0.1=hd590300_0
+  - libsqlite=3.45.1=h2797004_0
+  - libuuid=2.38.1=h0b41bf4_0
+  - libzlib=1.2.13=hd590300_5
+  - ncurses=6.4=h59595ed_2
+  - openssl=3.2.1=hd590300_0
+  - pip=24.0=pyhd8ed1ab_0
+  - python=3.12.0=hab00c5b_0_cpython
+  - readline=8.2=h8228510_1
+  - setuptools=69.1.1=pyhd8ed1ab_0
+  - tk=8.6.13=noxft_h4845f30_101
+  - tzdata=2024a=h0c530f3_0
+  - wheel=0.42.0=pyhd8ed1ab_0
+  - xz=5.2.6=h166bdaf_0
+  - pip:
+      - accessible-pygments==0.0.4
+      - alabaster==0.7.16
+      - attrs==23.2.0
+      - babel==2.14.0
+      - beautifulsoup4==4.12.3
+      - bleach==6.1.0
+      - certifi==2024.2.2
+      - charset-normalizer==3.3.2
+      - defusedxml==0.7.1
+      - docutils==0.20.1
+      - fastjsonschema==2.19.1
+      - idna==3.6
+      - imagesize==1.4.1
+      - jinja2==3.1.3
+      - jsonschema==4.21.1
+      - jsonschema-specifications==2023.12.1
+      - jupyter-client==8.6.0
+      - jupyter-core==5.7.1
+      - jupyterlab-pygments==0.3.0
+      - markdown-it-py==3.0.0
+      - markupsafe==2.1.5
+      - mdit-py-plugins==0.4.0
+      - mdurl==0.1.2
+      - mistune==3.0.2
+      - myst-parser==2.0.0
+      - nbclient==0.9.0
+      - nbconvert==7.16.1
+      - nbformat==5.9.2
+      - nbsphinx==0.9.3
+      - packaging==23.2
+      - pandocfilters==1.5.1
+      - platformdirs==4.2.0
+      - pydata-sphinx-theme==0.15.2
+      - pygments==2.17.2
+      - python-dateutil==2.8.2
+      - pyyaml==6.0.1
+      - pyzmq==25.1.2
+      - referencing==0.33.0
+      - requests==2.31.0
+      - rpds-py==0.18.0
+      - six==1.16.0
+      - snowballstemmer==2.2.0
+      - soupsieve==2.5
+      - sphinx==7.2.6
+      - sphinx-copybutton==0.5.2
+      - sphinx-rtd-theme==2.0.0
+      - sphinxcontrib-applehelp==1.0.8
+      - sphinxcontrib-devhelp==1.0.6
+      - sphinxcontrib-htmlhelp==2.0.5
+      - sphinxcontrib-jquery==4.1
+      - sphinxcontrib-jsmath==1.0.1
+      - sphinxcontrib-qthelp==1.0.7
+      - sphinxcontrib-serializinghtml==1.1.10
+      - tinycss2==1.2.1
+      - tornado==6.4
[…]
