+++ .readthedocs.yaml
@@ -58,4 +58,4 @@ sphinx:
 # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
 python:
     install:
-    - requirements: docs/doc_environment.yml
+    - requirements: docs/requirements.txt
