### 2024-02-28 7870f45
no references to __version__ in the conf file
