+++ .readthedocs.yaml
@@ -1,38 +1,5 @@
-
-# Read the Docs configuration file
-# See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
-
-# # Required
-version: 2
-
-# # Set the version of Python and other tools you might need
-# build:
-#   os: ubuntu-22.04
-#   tools: {python: "3.10"}
-#   jobs:
-#     pre_create_environment:
-#       - asdf plugin add poetry
-#       - asdf install poetry latest
-#       - asdf global poetry latest
-#       - poetry config virtualenvs.create false
-#       - poetry build
-#       - poetry update
-#       - poetry install --all-extras
-#     post_install:
-#       - poetry install
-
-# # Build documentation in the docs/ directory with Sphinx
-# sphinx:
-#   configuration: docs/conf.py
-#   fail_on_warning: true
-
-
-# .readthedocs.yaml
-# Read the Docs configuration file
-# See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
-
 # Required
-#version: 2
+version: 2
 
 #Set the OS, Python version and other tools you might need
 build:
