### 2024-02-28 feb6ba8
remove versioning refernces of pydata theme
