+++ docs/full_module_summaries/missingobs.rst
@@ -1,6 +1,0 @@
-Missing Observations module
-----------------------------
-.. automodule:: metobs_toolkit.missingobs
-   :members:
-   :undoc-members:
-   :show-inheritance:
