+++ docs/full_module_summaries/settings.rst
@@ -1,7 +1,0 @@
-settings module
--------------------
-
-.. automodule:: metobs_toolkit.settings
-   :members:
-   :undoc-members:
-   :show-inheritance:
