+++ docs/full_module_summaries/dataset.rst
@@ -1,18 +1,0 @@
-dataset module
--------------------
-
-The dataset module holds the Dataset class.
-
-.. automodule:: metobs_toolkit.dataset
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-dataset module (settings updater)
-----------------------------------
-The dataset_settings_updater module is an extension of the Dataset class to include methods for updating the settings.
-
-.. automodule:: metobs_toolkit.dataset_settings_updater
-   :members:
-   :undoc-members:
-   :show-inheritance:
