+++ docs/full_module_summaries/writing.rst
@@ -1,9 +1,0 @@
-writing_files module
-----------------------
-
-This module contains functions to write datafiles.
-
-.. automodule:: metobs_toolkit.writing_files
-   :members:
-   :undoc-members:
-   :show-inheritance:
