+++ docs/full_module_summaries/printing.rst
@@ -1,7 +1,0 @@
-printing module
--------------------
-
-.. automodule:: metobs_toolkit.printing
-   :members:
-   :undoc-members:
-   :show-inheritance:
