### 2024-02-28 935ca24
Move to rtd (#446)
* init for rtd hosting of the docs

* add poetry install in yaml

* test conf without inmporting the toolkit

* moc all dependencies and creat sepereate doc build env for rtd support

* test with requirements file

* no references to __version__ in the conf file

* remove versioning refernces of pydata theme

* minor codereview
