+++ docs/full_module_summaries/plotting.rst
@@ -1,8 +1,0 @@
-
-plotting module
--------------------
-
-.. automodule:: metobs_toolkit.plotting_functions
-   :members:
-   :undoc-members:
-   :show-inheritance:
