+++ docs/full_module_summaries/df_helpers.rst
@@ -1,10 +1,0 @@
-df_helpers module
--------------------
-
-This module contains a collection of (random) functions that are usefull for
-multiple modules. None of these functions should be used by a regular user.
-
-.. automodule:: metobs_toolkit.df_helpers
-   :members:
-   :undoc-members:
-   :show-inheritance:
