+++ docs/full_module_summaries/qc.rst
@@ -1,16 +1,0 @@
-Quality control module
-----------------------
-
-.. automodule:: metobs_toolkit.qc_checks
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
-Quality control statistics module
------------------------------------
-
-.. automodule:: metobs_toolkit.qc_statistics
-   :members:
-   :undoc-members:
-   :show-inheritance:
