+++ docs/full_module_summaries/modeldata.rst
@@ -1,7 +1,0 @@
-modeldata module
--------------------
-
-.. automodule:: metobs_toolkit.modeldata
-   :members:
-   :undoc-members:
-   :show-inheritance:
