+++ docs/requirements.txt
@@ -1,0 +1,7 @@
+# Defining the exact version will make sure things don't break
+sphinx==7.2.6
+sphinx_rtd_theme==2.0.0
+myst-parser==2.0.0
+nbsphinx==0.9.3
+pydata-sphinx-theme==0.15.2
+sphinx-copybutton==0.5.2
