+++ docs/full_module_summaries/data_import.rst
@@ -1,10 +1,0 @@
-
-data_import module
--------------------
-
-This module contains functions to read datafiles.
-
-.. automodule:: metobs_toolkit.data_import
-   :members:
-   :undoc-members:
-   :show-inheritance:
