+++ docs/full_module_summaries/template_build.rst
@@ -1,17 +1,0 @@
-
-template build prompt module
-------------------------------
-
-.. automodule:: metobs_toolkit.data_templates.template_build_prompt
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
-import template module
-------------------------------
-
-.. automodule:: metobs_toolkit.data_templates.import_templates
-   :members:
-   :undoc-members:
-   :show-inheritance:
