+++ .readthedocs.yaml
@@ -1,24 +1,28 @@
-# Read the Docs configuration file
-# See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
-
 # Required
 version: 2
 
-# Set the version of Python and other tools you might need
+#Set the OS, Python version and other tools you might need
 build:
   os: ubuntu-22.04
-  tools: {python: "3.10"}
-  jobs:
-    pre_create_environment:
-      - asdf plugin add poetry
-      - asdf install poetry latest
-      - asdf global poetry latest
-      - poetry config virtualenvs.create false
-      - poetry build
-    post_install:
-      - poetry install
+  tools:
+    python: "3.12"
+   # You can also specify other tool versions:
+   # nodejs: "19"
+   # rust: "1.64"
+   # golang: "1.19"
 
-# Build documentation in the docs/ directory with Sphinx
+# Build documentation in the "docs/" directory with Sphinx
 sphinx:
   configuration: docs/conf.py
-  fail_on_warning: true
+
+# Optionally build your docs in additional formats such as PDF and ePub
+# formats:
+#    - pdf
+#    - epub
+
+# Optional but recommended, declare the Python requirements required
+# to build your documentation
+# See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
+python:
+    install:
+    - requirements: docs/requirements.txt
