+++ docs/paper/paper.md
@@ -77,13 +77,13 @@ Sophisticated software such as TITAN [@titan2020] and CrowdQC+ [@CrowdQC] were d
 
 # Technical implementation
 
-The MetObs-toolkit provides a comprehensive framework for scientists to process raw meteorological data for analysis by making intensive use of the pandas [@pandas] and geopandas [@geopandas] libraries. The process consists of the following steps, visualized in the \autoref{fig:overview_fig}.
+The MetObs-toolkit provides a comprehensive framework for scientists to process raw meteorological data for analysis by making intensive use of the pandas [@pandas] and GeoPandas [@geopandas] libraries. The process consists of the following steps, visualized in the \autoref{fig:overview_fig}.
 
 Firstly, the raw data is mapped to the toolkit standards by use of a template. Once the raw data is imported into A Toolkit Dataset class instance, missing observations are identified and methods to resample and synchronize observations can be used.
 
 Next, quality control is performed in the form of a series of checks. These checks are designed to examine data types, irregular timestamps, max-min thresholds, repetitions criteria, spike tests, allowed variation in time windows and spatial tests. Advanced quality control methods are available through the implementation of TITAN into the toolkit. The user can choose to keep the outliers or convert them to missing records (which can then be filled).
 
-Finally, gap filling is applied by using interpolation methods and/or importing ERA5 reanalysis [@era5] time series to replace missing records. The latter is stored as a Toolkit `Modeldata` class instance, which has a set of methods to directly import the required time series through the use of the Google Earth Engine API. The user then obtains a cleaned-up dataset ready for analysis. A set of typical analysis techniques such as filters, aggregation schemes, and landcover correlation estimates are implemented in the `Toolkit-Analysis` class.
+Finally, gap filling is applied by using interpolation methods and/or importing ERA5 reanalysis [@era5] time series to replace missing records. The latter is stored as a Toolkit *Modeldata* class instance, which has a set of methods to directly import the required time series through the use of the Google Earth Engine API. The user then obtains a cleaned-up dataset ready for analysis. A set of typical analysis techniques such as filters, aggregation schemes, and landcover correlation estimates are implemented in the `Toolkit-Analysis` class.
 
 \autoref{fig:overview_fig} gives an overview of the main framework of the MetObs-toolkit, but it is an evolving project that responds to the community's needs and input. As an example, the development of a graphical user interface (GUI) for the toolkit is planned. A GUI would increase the ease of use by enabling users to create templates, adjust QC settings, and plot data interactively.
 
@@ -93,7 +93,7 @@ The authors would like to thank all participants of the [COST FAIRNESS](https://
 
 No specific funding has been obtained to build the MetObs-toolkit, but the authors have been supported by different Belgian and Flemish scientific grants.
 
-FWO: Sara (fellowship 1270723N), Wout (fellowship 1157523N) and Amber(fellowship 11PBN24N).
+FWO: Sara (fellowship 1270723N), Wout (fellowship 1157523N) and Amber (fellowship 11PBN24N).
 
 BELSPO: Kobe (B2/223/P1/CORDEX.be II), Thomas (B2/202/P1/CS-MASK), Michiel (B2/212/P2/CLIMPACTH) and Steven (FED-tWIN Prf-2020-018_AURA)
 
