### 2024-02-29 #5916 @vergauwenthomas
> ## Further reading and citations needed here
> Since MetObs-Toolkit implements quite a few methods, it would be important to link to other sources which may describe them in detail. If you think that these methods might be subject to change, it might be sufficient to cite the references in the online documentation and link this software paper to the documentation instead. See how [scikit-image](https://scikit-image.org/docs/stable/api/skimage.filters.html#skimage.filters.butterworth) does it.
>
> > “55 to examine data types, irregular timestamps, max-min thresholds, repetitions criteria, spike tests, allowed variation in time windows and spatial tests” (Vergauwen et al., 2024, p. 3) (pdf)

> > “59 ... gap filling is applied by using interpolation methods” (Vergauwen et al., 2024, p. 3) (pdf)
>
> Gap filling and bias correction is another involved topic. What is the basis behind the methods used? Further reading and citations needed here.
>

The gap-filling methods implemented are indeed quite evolved techniques that could benefit from more explanation and references. The most advanced methods, specifically the debiasing methods using ERA5, have been thoroughly tested and described by a co-author ( @amberj99 ), who has submitted a scientific publication on them. Once this publication is accepted, I can refer to it in the documentation, similar to the scikit-image example.

To provide users with more information, I will extend the documentation on these methods. Explicitly referring to the documentation in the paper for additional details seems self-evident, so I would avoid doing so.
