### 2024-02-29 #5916 @vergauwenthomas
> > “55 to examine data types, irregular timestamps, max-min thresholds, repetitions criteria, spike tests, allowed variation in time windows and spatial tests” (Vergauwen et al., 2024, p. 3) (pdf)
>
> These, IMHO, are quite involved methods. What are the formal basis behind these QC methods or are they based on intuition? Some reuses TITAN (titanlib) while the rest are implemented from the scratch (which means citing TITAN alone will not be sufficient).
>
Most of these QC checks are well-known and often used (repetitions, spike, min-max, ... ) and are implemented from scratch. (*Similar to implementing your own function to calculate the RMSE.*)

The spatial check and window/spike check are a bit less known and could benefit from more explainable documentation. I will extend the documentation on these methods as well.
