### 2024-02-29 fceff96
step by step explenation of the buddy check
