### 2024-03-02 #5916 @ashwinvis
@vergauwenthomas thanks. I was trying to think in the shoes of a hobbyist user that you are targeting. From my perspective, even after working in the same domain, I have never heard the term "spike check", for example. It is great that you are documenting it.
