### 2024-03-04 5abf2c8
update docstrings on qc and gapfill methods
