### 2024-03-04 df96b42
updating and spellcheck on docstring
