+++ metobs_toolkit/dataset.py
@@ -1559,7 +1559,7 @@ class Dataset:
         max_interpolation_duration, the linear interpolation method is applied
         else the debiased modeldata method.
 
-        For detailed description of these methods, we refere to the
+        For a detailed description of these methods, we refer to the
         corresponding metobs_toolkit.Dataset.fill_gaps_linear() and
         metobs_toolkit.Dataset.fill_gaps_era5().
 
@@ -1727,12 +1727,12 @@ class Dataset:
 
         Notes
         -----
-        A schematic description on the linear gap fill:
+        A schematic description of the linear gap fill:
 
             1. Iterate over all gaps.
-            2. The gap is converted into a set of missing records (depends on time resolution of the observations).
+            2. The gap is converted into a set of missing records (depending on the time resolution of the observations).
             3. Find a leading (the last observations before the gap) record and a trailing record (the last observation after the gap).
-            4. By using the leading and trailing record an interpolation is applied to fill the missing records. A maximum consicutive fill threshold is applied, if exeeded the fill values are Nan's.
+            4. By using the leading and trailing record an interpolation is applied to fill the missing records. A maximum consecutive fill threshold is applied, if exceeded the fill values are Nan's.
             5. The gap is updated with the interpolated values (metobs_toolkit.Gap.gapfill_df)
 
         Examples
@@ -1856,12 +1856,12 @@ class Dataset:
 
         Notes
         -----
-        A schematic description on the linear fill of missing observations:
+        A schematic description of the linear fill of missing observations:
 
             1. Iterate over all missing observations.
-            2. The missing observations are converted into a set of missing records (depends on time resolution of the observations).
+            2. The missing observations are converted into a set of missing records (depending on the time resolution of the observations).
             3. Find a leading (the last observations before the missing observation) record and a trailing record (the last observation after the missing observation).
-            4. By using the leading and trailing record, an interpolation is applied to fill the missing records.
+            4. By using the leading and trailing records, interpolation is applied to fill the missing records.
             5. The missing record is updated with the interpolated values (metobs_toolkit.Gap.gapfill_df).
 
         Examples
@@ -2210,29 +2210,28 @@ class Dataset:
 
         Notes
         -----
-        A schematic description on the fill_gaps_era5 method:
+        A schematic description of the fill_gaps_era5 method:
 
             1. Modeldata is converted to the timezone of the observations.
             2. Iterate over all gaps.
-                * The gap is converted into a set of missing records (depends on time resolution of the observations).
+                * The gap is converted into a set of missing records (depending on the time resolution of the observations).
                 * Find a leading and trailing period. These periods are a subset
                  of observations respectively before and after the gap. The size
-                 of these subset is set by a target size (in records) and a minimum
-                 size (in records). If the subset of observations is smalller than
+                 of these subsets is set by a target size (in records) and a minimum
+                 size (in records). If the subset of observations is smaller than
                  the corresponding minimum size, the gap cannot be filled.
                 * Modeldata, for the corresponding station and observation type, is extracted for the leading and trailing period.
-                * By comparing with the leading and trailing observations,and by
                 * By comparing the model data with the observations of the
-                 leading and trailing period, and grouping all records to their
-                 timestamp (i.g. diurnal categories), biasses are computed.
+                  leading and trailing period, and grouping all records to their
+                  timestamp (i.g. diurnal categories), biasses are computed.
                 * Modeldata for the missing records is extracted.
-                * Weight ([0;1]) are computed for each gap record, representing
-                 the normelized distance (in time), to the beginning and end of
+                * Weights ([0;1]) are computed for each gap record, representing
+                 the normalized distance (in time), to the beginning and end of
                  the gap.
-                * The modeldata at the missing records is than corrected by
[…]
