### 2024-03-04 c5290ed
Extending docstrings with more in-dept explenations (#457)
* step by step explenation of the buddy check

* update docstrings on qc and gapfill methods

* updating and spellcheck on docstring

* fix identations

* import toolkit only for local doc builds
