### 2024-03-04 #5916 @vergauwenthomas
Hi @ashwinvis, the quality control methods and gapfill methods do now contain a step-by-step description in the docstrings.

Here are two examples of how this looks in the documentation:

- https://metobs-toolkit.readthedocs.io/en/latest/reference/api/metobs_toolkit.Dataset.fill_gaps_era5.html#metobs_toolkit.Dataset.fill_gaps_era5
- https://metobs-toolkit.readthedocs.io/en/latest/reference/api/metobs_toolkit.Dataset.apply_quality_control.html

I hope this answers your latest comment, if not let me know.
