+++ metobs_toolkit/dataset.py
@@ -1567,431 +1567,6 @@ class Dataset:
     #   Gap Filling
     # =============================================================================
 
-    def fill_gaps_automatic(
-        self,
-        modeldata,
-        obstype="temp",
-        max_interpolate_duration_str=None,
-        overwrite_fill=False,
-    ):
-        """Fill the gaps by using linear interpolation or debiased modeldata.
-
-        The method that is applied to perform the gapfill will be determined by
-        the duration of the gap.
-
-        When the duration of a gap is smaller or equal than
-        max_interpolation_duration, the linear interpolation method is applied
-        else the debiased modeldata method.
-
-
-        Parameters
-        ----------
-        modeldata : metobs_toolkit.Modeldata
-            The modeldata to use for the gapfill. This model data should the required
-            timeseries to fill all gaps present in the dataset.
-        obstype : String, optional
-            Name of the observationtype you want to apply gap filling on. The
-            modeldata must contain this observation type as well. The
-            default is 'temp'.
-        max_interpolate_duration_str : Timedelta or str, optional
-            Maximum duration to apply interpolation for gapfill when using the
-            automatic gapfill method. Gaps with longer durations will be filled
-            using debiased modeldata. The default is None.
-        overwrite_fill: bool, optional
-            If a gap has already filled values, the interpolation of this gap
-            is skipped if overwrite_fill is False. If set to True, the gapfill
-            values and info will be overwitten. The default is False.
-
-        Returns
-        -------
-        comb_df : TYPE
-            gapfilldf : pandas.DataFrame
-                A dataframe containing all the filled records.
-
-        Examples
-        --------
-        .. code-block:: python
-
-            import metobs_toolkit
-
-            your_dataset = metobs_toolkit.Dataset()
-            your_dataset.update_settings(
-                input_data_file=metobs_toolkit.demo_datafile, # path to the data file
-                input_metadata_file=metobs_toolkit.demo_metadatafile,
-                template_file=metobs_toolkit.demo_template,
-            )
-            # Specify the gap defenition
-            your_dataset.update_qc_settings(gapsize_in_records = 20)
-
-            #Update the gapsize BEFORE importing the data
-            your_dataset.import_data_from_file()
-
-            #Update the settings (definition of the period to calculate biases for)
-            your_dataset.update_gap_and_missing_fill_settings(
-                                                              gap_debias_prefered_leading_period_hours=24,
-                                                              gap_debias_prefered_trailing_period_hours=24,
-                                                              gap_debias_minimum_leading_period_hours=6,
-                                                              gap_debias_minimum_trailing_period_hours=6,
-                                                              )
-            #(As a demonstration, we will fill the gaps of a single station. The following functions can also be
-            # directly applied to the dataset.)
-            your_station = your_dataset.get_station('vlinder05')
-
-
-            #Get ERA5 modeldata at the location of your stations and period.
-            ERA5_modeldata = your_station.get_modeldata(modelname='ERA5_hourly',
-                                                        obstype='temp')
-
-            #Use the debias method to fill the gaps
-            gapfill_df = your_station.fill_gaps_automatic(modeldata=ERA5_modeldata,
-                                                          max_interpolate_duration_str='6H', # <6 hours will be interpolated
[…]
