+++ metobs_toolkit/dataset_gapfill.py
@@ -1,0 +1,570 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Tue Mar  5 11:46:29 2024
+
+@author: thoverga
+"""
+
+import logging
+import pandas as pd
+
+
+from metobs_toolkit import Dataset
+from metobs_toolkit.df_helpers import xs_save
+
+logger = logging.getLogger(__name__)
+
+
+class Dataset(Dataset):
+    """Extension on the metobs_toolkit.Dataset class with gapfill techniques"""
+
+    def get_gaps_fill_df(self):
+        """
+        Create a dataframe with information on how all gaps are filled.
+
+        Returns
+        -------
+        pandas.DataFrame
+            A DataFrame with ['name', 'datetime', 'obstype'] as index, ['fill',
+            'fill_method', 'msg'] as columns. Nan and None values are used
+            for records that have not been gapfilled.
+
+        Examples
+        --------
+        .. code-block:: python
+
+            >>> import metobs_toolkit
+            >>>
+            >>> # Import data into a Dataset
+            >>> dataset = metobs_toolkit.Dataset()
+            >>> dataset.update_settings(
+            ...                         input_data_file=metobs_toolkit.demo_datafile,
+            ...                         input_metadata_file=metobs_toolkit.demo_metadatafile,
+            ...                         template_file=metobs_toolkit.demo_template,
+            ...                         )
+            >>> dataset.import_data_from_file()
+            >>> dataset.coarsen_time_resolution(freq='1H')
+            >>>
+            >>> # Apply quality control on the temperature observations
+            >>> dataset.apply_quality_control(obstype='temp') #Using the default QC settings
+            >>>
+            >>> # Interpret the outliers as missing/gaps
+            >>> dataset.update_gaps_and_missing_from_outliers(obstype='temp')
+            >>> dataset
+            Dataset instance containing:
+                 *28 stations
+                 *['temp', 'humidity', 'radiation_temp', 'pressure', 'pressure_at_sea_level', 'precip', 'precip_sum', 'wind_speed', 'wind_gust', 'wind_direction'] observation types
+                 *10080 observation records
+                 *235 records labeled as outliers
+                 *2 gaps
+                 *1473 missing observations
+                 *records range: 2022-09-01 00:00:00+00:00 --> 2022-09-15 23:00:00+00:00 (total duration:  14 days 23:00:00)
+                 *time zone of the records: UTC
+                 *Coordinates are available for all stations.
+            >>> dataset.get_gaps_df()
+                                      start_gap                   end_gap        duration
+            name
+            vlinder05 2022-09-06 21:00:00+00:00 2022-09-13 06:00:00+00:00 6 days 09:00:00
+            vlinder05 2022-09-13 20:00:00+00:00 2022-09-15 23:00:00+00:00 2 days 03:00:00
+
+
+        """
+        gaps_infodf_list = []
+        for gap in self.gaps:
+            gapfilldf = pd.concat([gap.gapdf, gap.anchordf])
+            # Create a triple index
+            gapfilldf["obstype"] = gap.obstype.name
+            gapfilldf = gapfilldf.rename(columns={f"{gap.obstype.name}_fill": "fill"})
+            gapfilldf = gapfilldf.drop(columns=[gap.obstype.name])
+            gapfilldf = gapfilldf.reset_index().set_index(
+                ["name", "datetime", "obstype"]
+            )
+            # add a gapid column
+            gapfilldf["gap_ID"] = f"{gap.name};{gap.startdt};{gap.enddt}"
+
+            gaps_infodf_list.append(gapfilldf)
+
[…]
