+++ docs/paper/paper.md
@@ -81,9 +81,9 @@ The MetObs-toolkit provides a comprehensive framework for scientists to process 
 
 Firstly, the raw data is mapped to the toolkit standards by use of a template. Once the raw data is imported into a toolkit *Dataset class* instance, missing observations are identified and methods to resample and synchronize observations can be used.
 
-Next, quality control is performed in the form of a series of checks. These checks are designed to examine data types, irregular timestamps, max-min thresholds, repetitions criteria, spike tests, allowed variation in time windows and spatial tests. Advanced quality control methods are available through the implementation of TITAN into the toolkit. The user can choose to keep the outliers or convert them to missing records (which can then be filled).
+Next, quality control is performed in the form of a series of checks [@metobs_qc]. These checks are designed to examine data types, irregular timestamps, max-min thresholds, repetitions criteria, spike tests, allowed variation in time windows and spatial tests [@metobs_buddy]. Advanced quality control methods are available through the implementation of TITAN into the toolkit. The user can choose to keep the outliers or convert them to missing records (which can then be filled).
 
-Finally, gap filling is applied by using interpolation methods and/or importing ERA5 reanalysis [@era5] time series to replace missing records. The latter is stored as a toolkit *Modeldata class* instance, which has a set of methods to directly import the required time series through the use of the Google Earth Engine API. The user then obtains a cleaned-up dataset ready for analysis. A set of typical analysis techniques such as filters, aggregation schemes, and landcover correlation estimates are implemented in the toolkit *Analysis class*.
+Finally, gap filling is applied by using interpolation methods [@metobs_interpolation] and/or importing ERA5 reanalysis [@era5] time series to replace missing records [@metobs_gapfill]. The latter is stored as a toolkit *Modeldata class* instance, which has a set of methods to directly import the required time series through the use of the Google Earth Engine API. The user then obtains a cleaned-up dataset ready for analysis. A set of typical analysis techniques such as filters, aggregation schemes, and landcover correlation estimates are implemented in the toolkit *Analysis class*.
 
 \autoref{fig:overview_fig} gives an overview of the main framework of the MetObs-toolkit, but it is an evolving project that responds to the community's needs and input. As an example, the development of a graphical user interface (GUI) for the toolkit is planned. A GUI would increase the ease of use by enabling users to create templates, adjust QC settings, and plot data interactively.
 
