+++ docs/paper/paper.bib
@@ -1,4 +1,4 @@
-@Article{titan2020,
+s_qc@Article{titan2020,
     AUTHOR = {B{\aa}serud, L. and Lussana, C. and Nipen, T. N. and Seierstad, I. A. and Oram, L. and Aspelien, T.},
     TITLE = {TITAN automatic spatial quality control of meteorological in-situ observations},
     JOURNAL = {Advances in Science and Research},
@@ -156,27 +156,27 @@ year = {2020}
 }
 
 @software{metobs_qc,
-  title  = {Traditional quality control checks implemented in the MetObs toolkit.}
-  author = {Thomas Vergauwen, Michiel Vieijra}
-  year = {2023}
+  title  = {Traditional quality control checks implemented in the MetObs toolkit.},
+  author = {Thomas Vergauwen, Michiel Vieijra},
+  year = {2023},
   url = {https://metobs-toolkit.readthedocs.io/en/joss-version/reference/api/metobs_toolkit.Dataset.apply_quality_control.html}
 }
 @software{metobs_buddy,
-  title  = {Traditional spatial buddy check implemented in the MetObs toolkit.}
-  author = {Thomas Vergauwen}
-  year = {2023}
+  title  = {Traditional spatial buddy check implemented in the MetObs toolkit.},
+  author = {Thomas Vergauwen},
+  year = {2023},
   url = {https://metobs-toolkit.readthedocs.io/en/joss-version/reference/api/metobs_toolkit.Dataset.apply_buddy_check.html}
 }
 @software{metobs_interpolation,
-  title  = {Traditional gap interpolation implemented in the MetObs toolkit.}
-  author = {Thomas Vergauwen}
-  year = {2023}
+  title  = {Traditional gap interpolation implemented in the MetObs toolkit.},
+  author = {Thomas Vergauwen},
+  year = {2023},
   url = {https://metobs-toolkit.readthedocs.io/en/joss-version/reference/api/metobs_toolkit.Dataset.fill_gaps_linear.html}
 }
 @software{metobs_gapfill,
-  title  = {Debiased gap filling using ERA5 model data implemented in the MetObs toolkit}
-  author = {Amber Jacobs, Thomas Vergauwen}
-  year = {2023}
+  title  = {Debiased gap filling using ERA5 model data implemented in the MetObs toolkit.},
+  author = {Amber Jacobs, Thomas Vergauwen},
+  year = {2023},
   url = {https://metobs-toolkit.readthedocs.io/en/joss-version/reference/api/metobs_toolkit.Dataset.fill_gaps_era5.html}
 }
 
