+++ docs/paper/paper.bib
@@ -155,31 +155,6 @@ year = {2020}
  url           = {https://doi.org/10.1038/s41586-020-2649-2}
 }
 
-@software{metobs_qc,
-  title  = {Traditional quality control checks implemented in the MetObs toolkit.},
-  author = {Thomas Vergauwen, Michiel Vieijra},
-  year = {2023},
-  url = {https://metobs-toolkit.readthedocs.io/en/joss-version/reference/api/metobs_toolkit.Dataset.apply_quality_control.html}
-}
-@software{metobs_buddy,
-  title  = {Traditional spatial buddy check implemented in the MetObs toolkit.},
-  author = {Thomas Vergauwen},
-  year = {2023},
-  url = {https://metobs-toolkit.readthedocs.io/en/joss-version/reference/api/metobs_toolkit.Dataset.apply_buddy_check.html}
-}
-@software{metobs_interpolation,
-  title  = {Traditional gap interpolation implemented in the MetObs toolkit.},
-  author = {Thomas Vergauwen},
-  year = {2023},
-  url = {https://metobs-toolkit.readthedocs.io/en/joss-version/reference/api/metobs_toolkit.Dataset.fill_gaps_linear.html}
-}
-@software{metobs_gapfill,
-  title  = {Debiased gap filling using ERA5 model data implemented in the MetObs toolkit.},
-  author = {Amber Jacobs, Thomas Vergauwen},
-  year = {2023},
-  url = {https://metobs-toolkit.readthedocs.io/en/joss-version/reference/api/metobs_toolkit.Dataset.fill_gaps_era5.html}
-}
-
 
 
 @software{geopandas,
