+++ docs/paper/paper.md
@@ -81,7 +81,7 @@ The MetObs-toolkit provides a comprehensive framework for scientists to process 
 
 Firstly, the raw data is mapped to the toolkit standards by use of a template. Once the raw data is imported into a toolkit *Dataset class* instance, missing observations are identified and methods to resample and synchronize observations can be used.
 
-Next, quality control is performed in the form of a [series of checks]([https://www.fairness-ca20108.eu/](https://metobs-toolkit.readthedocs.io/en/joss-version/reference/api/metobs_toolkit.Dataset.apply_quality_control.html)). These checks are designed to examine data types, irregular timestamps, max-min thresholds, repetitions criteria, spike tests, allowed variation in time windows and [spatial tests](https://metobs-toolkit.readthedocs.io/en/joss-version/reference/api/metobs_toolkit.Dataset.apply_buddy_check.html). Advanced quality control methods are available through the implementation of TITAN into the toolkit. The user can choose to keep the outliers or convert them to missing records (which can then be filled).
+Next, quality control is performed in the form of a [series of checks](https://metobs-toolkit.readthedocs.io/en/joss-version/reference/api/metobs_toolkit.Dataset.apply_quality_control.html). These checks are designed to examine data types, irregular timestamps, max-min thresholds, repetitions criteria, spike tests, allowed variation in time windows and [spatial tests](https://metobs-toolkit.readthedocs.io/en/joss-version/reference/api/metobs_toolkit.Dataset.apply_buddy_check.html). Advanced quality control methods are available through the implementation of TITAN into the toolkit. The user can choose to keep the outliers or convert them to missing records (which can then be filled).
 
 Finally, gap filling is applied by using interpolation methods and/or importing ERA5 reanalysis [@era5] time series to [replace missing records](https://metobs-toolkit.readthedocs.io/en/joss-version/reference/api/metobs_toolkit.Dataset.fill_gaps_era5.html). The latter is stored as a toolkit *Modeldata class* instance, which has a set of methods to directly import the required time series through the use of the Google Earth Engine API. The user then obtains a cleaned-up dataset ready for analysis. A set of typical analysis techniques such as filters, aggregation schemes, and landcover correlation estimates are implemented in the toolkit *Analysis class*.
 
