+++ docs/paper/paper.bib
@@ -1,4 +1,4 @@
-s_qc@Article{titan2020,
+@Article{titan2020,
     AUTHOR = {B{\aa}serud, L. and Lussana, C. and Nipen, T. N. and Seierstad, I. A. and Oram, L. and Aspelien, T.},
     TITLE = {TITAN automatic spatial quality control of meteorological in-situ observations},
     JOURNAL = {Advances in Science and Research},
@@ -154,7 +154,6 @@ year = {2020}
  publisher     = {Springer Science and Business Media {LLC}},
  url           = {https://doi.org/10.1038/s41586-020-2649-2}
 }
-
 
 
 @software{geopandas,
