### 2024-03-06 #5916 @vergauwenthomas
@ashwinvis, the links are added. They point to the `JOSS version` of the documentation, so changes to these QC, gapfilling and documentation-structure should not break the hyperlinks.
