### 2024-03-06 b21263c
remove gapsettings in the Setings class
