+++ metobs_toolkit/missingobs.py
@@ -1,363 +1,363 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-This module contains the Missingob_collection class and all its methods.
+# #!/usr/bin/env python3
+# # -*- coding: utf-8 -*-
+# """
+# This module contains the Missingob_collection class and all its methods.
 
-A Missingob_collection holds all missing observations and methods on how to
-fill them.
-"""
+# A Missingob_collection holds all missing observations and methods on how to
+# fill them.
+# """
 
 
-import pandas as pd
-import numpy as np
-from datetime import timedelta
-import logging
+# import pandas as pd
+# import numpy as np
+# from datetime import timedelta
+# import logging
 
-from metobs_toolkit.df_helpers import _find_closes_occuring_date, xs_save, concat_save
+# from metobs_toolkit.df_helpers import _find_closes_occuring_date, xs_save, concat_save
 
-logger = logging.getLogger(__name__)
+# logger = logging.getLogger(__name__)
 
-# =============================================================================
-# Missing observation
-
-# a missing observation is a missing timestamp
-# =============================================================================
-
-
-class Missingob_collection:
-    """Class object handling a set of missing observations."""
-
-    def __init__(self, missing_obs_series):
-        """Init missing observations."""
-        missing_obs_series.name = "datetime"
-        missing_obs_series.index.name = "name"
-
-        missing_obs_df = missing_obs_series.reset_index()  # needed to find duplicates
-        missing_obs_df = missing_obs_df.drop_duplicates()
-        missing_obs_series = missing_obs_df.set_index("name")["datetime"]
-        missing_obs_series = missing_obs_series.sort_index()
-
-        missing_idx = missing_obs_series.reset_index()
-        missing_idx = missing_idx.set_index(["name", "datetime"])
-
-        self.series = missing_obs_series
-        self.idx = missing_idx.index
-
-        # gap fill (only for conventional saving)
-        self.fill_df = pd.DataFrame()
-        self.fill_technique = None
-
-    def __add__(self, other):
-        """Append two collections of missing observations."""
-
-        comb_series = concat_save([self.series, other.series])
-        # drop duplicates and sort
-        comb_df = comb_series.reset_index()  # needed to find duplicates
-        comb_df = comb_df.drop_duplicates()
-        comb_series = comb_df.set_index("name")["datetime"]
-        comb_series = comb_series.sort_index()
-
-        self.series = comb_series
-        comb_idx = comb_series.reset_index()
-        comb_idx = comb_idx.set_index(["name", "datetime"])
-        self.idx = comb_idx.index
-        return self
-
-    def __len__(self):
-        """Get number of missing observations."""
-        return self.series.shape[0]
-
-    def __str__(self):
-        """Print overview info of missing observations."""
-        if self.series.empty:
-            return "Empty missing observations."
-
-        if not self.fill_df.empty:
-            return f"Missing observations with filled ({self.fill_technique}) \
-                values: \n {self.fill_df} \n Original missing observations on import: \n {self.idx}"
-
-        return f"Missing observations: \n {self.series}"
-
-    def __repr__(self):
-        """Print overview info of missing observations."""
-        return self.__str__()
-
-    def get_info(self, max_disp_list=7):
-        """Print out detailed information on the missing observations.
-
-        Parameters
-        ----------
-        max_disp_list : int, optional
-            Max size of lists to print out. If listsize is larger, the length of
-            the list is printed. The default is 7.
-
-        Returns
-        -------
-        None.
-
-        """
-        print("-------- Missing observations info --------")
-        if self.series.empty:
-            print("Empty missing observations.")
-            return
-        print(
-            "(Note: missing observations are defined on the frequency estimation of the native dataset.)"
-        )
-
[…]
