+++ metobs_toolkit/settings.py
@@ -31,7 +31,7 @@ class Settings:
         self.app = {}
         self.qc = {}
         self.gap = {}
-        self.missing_obs = {}
+        # self.missing_obs = {}
         self.templates = {}
         self.gee = {}
         self.IO = {
@@ -155,21 +155,12 @@ class Settings:
         """
         logger.debug("Updating gap settings.")
         from .settings_files.gaps_and_missing_settings import (
-            gaps_settings,
             gaps_info,
-            gaps_fill_settings,
             gaps_fill_info,
-            missing_obs_fill_settings,
-            missing_obs_fill_info,
         )
 
-        self.gap["gaps_settings"] = gaps_settings
         self.gap["gaps_info"] = gaps_info
-        self.gap["gaps_fill_settings"] = gaps_fill_settings
         self.gap["gaps_fill_info"] = gaps_fill_info
-
-        self.missing_obs["missing_obs_fill_settings"] = missing_obs_fill_settings
-        self.missing_obs["missing_obs_fill_info"] = missing_obs_fill_info
 
     def _update_templates(self):
         """
@@ -333,7 +324,7 @@ class Settings:
             "app",
             "qc",
             "gap",
-            "missing_obs",
+            # "missing_obs",
             "templates",
             "gee",
         ]

[…]
