### 2024-03-06 #5916 @ashwinvis
@hugoledoux Everything looks good from my perspective and I recommend the article for acceptance into JOSS. Congrats to @vergauwenthomas and co. for developing this package and meticulously documenting and improving it!
