### 2024-03-07 #5916 @hugoledoux
> (editorialbot does not build the pdf from `master`, so i have merged the `paper` branch with the ´master`.)

you could also have changed the branch in the setup above, now it's "paper"... I guess useful for your next submission :)
