### 2024-03-07 #5916 @hugoledoux
It took some time but in the end it's the results that count (right?) and I'm happy with the thorough discussions and feedback that @ashwinvis and @Zeitsperre gave 👍

I have just proof-read the paper and I found no typos/mistakes, and as it is mentioned above, it is very well written and clear. I wish all submissions were like this one!

I checked the website and docs and perhaps one tiny comment from my side: you put everywhere `pip3` but Python 2.7 died a long time ago and I can't imagine people still using it... `pip3` or `python3` is actually confusing for our students here since it's not always in their path and then they ask us teachers why they don't have it. So I personally stopped using it. Just a small thing, not super important...

@vergauwenthomas at this point could you:
