+++ README.md
@@ -22,24 +22,24 @@ Documentation can be found [here](https://metobs-toolkit.readthedocs.io/en/lates
 Install the package using pip:
 
 ```bash
-pip3 install metobs-toolkit
+pip install metobs-toolkit
  ```
 To install the PyPi version of the toolkit. To install the github versions one can use these commands:
 
 ```bash
 #main versions
-pip3 install git+https://github.com/vergauwenthomas/MetObs_toolkit.git
+pip install git+https://github.com/vergauwenthomas/MetObs_toolkit.git
 
 #development version
-pip3 install git+https://github.com/vergauwenthomas/MetObs_toolkit.git@dev
+pip install git+https://github.com/vergauwenthomas/MetObs_toolkit.git@dev
 
 #specific release from github
-pip3 install git+https://github.com/vergauwenthomas/MetObs_toolkit.git@v0.1.1
+pip install git+https://github.com/vergauwenthomas/MetObs_toolkit.git@v0.1.3-joss
  ```
 For some advanced quality control methods, the [Titanlib package](https://github.com/metno/titanlib) is used. Since the installation of titanlib requires a c++ compiler, we have chosen not to include it in the toolkit. If you want to use the Titanlib functionality you must install both the toolkit and Titanlib:
 
 ```bash
-pip3 install metobs-toolkit titanlib
+pip install metobs-toolkit titanlib
  ```
 
 To use the package, import it in Python:
