### 2024-03-07 #5916 @vergauwenthomas
@hugoledoux, Great!

> It took some time but in the end it's the results that count (right?) and I'm happy with the thorough discussions and feedback that @ashwinvis and @Zeitsperre gave 👍
>
Fully agree, the toolkit has gained a lot from their remarks and suggestions ( and I learned a lot as well).

> I have just proof-read the paper and I found no typos/mistakes, and as it is mentioned above, it is very well written and clear. I wish all submissions were like this one!
>
> I checked the website and docs and perhaps one tiny comment from my side: you put everywhere `pip3` but Python 2.7 died a long time ago and I can't imagine people still using it... `pip3` or `python3` is actually confusing for our students here since it's not always in their path and then they ask us teachers why they don't have it. So I personally stopped using it. Just a small thing, not super important...
>
I have replaced all mentions of `pip3` and `Python3` with `pip` and `Python`.

> @vergauwenthomas at this point could you:
>
> * [ ]  Make a tagged release of your software, and list the version tag of the archived version here.

The release (*JOSS release*) comes with the version tag `v0.1.3-joss`.

> * [ ]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)

Done.

> * [ ]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.

Done. (I listed the same authors as the paper.)

> * [ ]  Please list the DOI of the archived version here.

10.5281/zenodo.10792385

> I can then move forward with recommending acceptance of the submission.
