### 2024-03-07 #5916 @hugoledoux
for the version it's ultimately up to you, but I think it might be an idea to bump to a minor version with all the changes made during the JOSS process. I would suggest v0.2.0, this makes it clearer.

For the authors, you added some categories and that mangles the reference and only lists you an as author, can you remove those categories? Researchers + Supervisor.

I'm glad you enjoyed JOSS! Consider adding your name to the potential list of reviewers also: https://reviewers.joss.theoj.org/

(I didn't know it was possible, so I learned that today).
