### 2024-03-07 #5916 @vergauwenthomas
I have bumped it to `v0.2.0`.

I had to create a new DOI (because there was a reference to `v0.1.3-joss' in the file name, which i could not change). You can find it here: **10.5281/zenodo.10794210**

And as @hugoledoux pointed out, i have changed the co-authors and the citations seems oke now:

Vergauwen, T., Vieijra, M., Covaci, A., Jacobs, A., Top, S., Dewettinck, W., Vandelanotte, K., Hellebosch, I., & Caluwaerts, S. (2024). MetObs - a Python toolkit for using non-traditional meteorological observations (v0.2.0). Zenodo. https://doi.org/10.5281/zenodo.10794210
