+++ metobs_toolkit/__init__.py
@@ -72,4 +72,4 @@ from metobs_toolkit.dataset_settings_updater import Dataset
 
 # DO not change this manually!
 
-__version__ = "0.1.3a"
+__version__ = "0.2.0"
