+++ _sources/reference/api/metobs_toolkit.Dataset.import_dataset.rst.txt
@@ -1,0 +1,6 @@
+metobs\_toolkit.Dataset.import\_dataset
+=======================================
+
+.. currentmodule:: metobs_toolkit
+
+.. automethod:: Dataset.import_dataset
