+++ _sources/paper/index.rst.txt
@@ -1,0 +1,21 @@
+###########################
+  JOSS publication
+###########################
+
+About JOSS
+-----------
+The `Journal of Open Source Software <https://joss.theoj.org/>`_ is a developer friendly, open access journal for research software packages.
+
+
+JOSS paper
+-------------------------
+A `MetObs-toolkit publication <https://joss.theoj.org/papers/ffa3a79315bdf4c4793992a1de41193d>`_ has been submitted and is currently under `review <https://github.com/openjournals/joss-reviews/issues/5916>`_. A draft version of the paper can be found in `here <https://github.com/vergauwenthomas/MetObs_toolkit/blob/master/docs/paper/paper.pdf>`_.
+
+
+Additionally, we add the script for creating the figures that are used in the publication.
+
+
+.. toctree::
+   :maxdepth: 1
+
+   paper_figures.ipynb
