+++ _sources/reference/api/metobs_toolkit.Obstype.get_orig_name.rst.txt
@@ -1,0 +1,6 @@
+metobs\_toolkit.Obstype.get\_orig\_name
+=======================================
+
+.. currentmodule:: metobs_toolkit
+
+.. automethod:: Obstype.get_orig_name
