+++ _sources/reference/api/metobs_toolkit.Obstype.set_original_name.rst.txt
@@ -1,0 +1,6 @@
+metobs\_toolkit.Obstype.set\_original\_name
+===========================================
+
+.. currentmodule:: metobs_toolkit
+
+.. automethod:: Obstype.set_original_name
