+++ _sources/reference/api/metobs_toolkit.Modeldata.save_modeldata.rst.txt
@@ -1,0 +1,6 @@
+metobs\_toolkit.Modeldata.save\_modeldata
+=========================================
+
+.. currentmodule:: metobs_toolkit
+
+.. automethod:: Modeldata.save_modeldata
