+++ _sources/reference/api/metobs_toolkit.ModelObstype_Vectorfield.has_mapped_band.rst.txt
@@ -1,0 +1,6 @@
+metobs\_toolkit.ModelObstype\_Vectorfield.has\_mapped\_band
+===========================================================
+
+.. currentmodule:: metobs_toolkit
+
+.. automethod:: ModelObstype_Vectorfield.has_mapped_band
