+++ _sources/reference/api/metobs_toolkit.ModelObstype_Vectorfield.rst.txt
@@ -1,0 +1,41 @@
+metobs\_toolkit.ModelObstype\_Vectorfield
+=========================================
+
+.. currentmodule:: metobs_toolkit
+
+.. autoclass:: ModelObstype_Vectorfield
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~ModelObstype_Vectorfield.__init__
+      ~ModelObstype_Vectorfield.add_new_band
+      ~ModelObstype_Vectorfield.add_unit
+      ~ModelObstype_Vectorfield.convert_to_standard_units
+      ~ModelObstype_Vectorfield.get_all_units
+      ~ModelObstype_Vectorfield.get_bandname_mapper
+      ~ModelObstype_Vectorfield.get_description
+      ~ModelObstype_Vectorfield.get_info
+      ~ModelObstype_Vectorfield.get_mapped_datasets
+      ~ModelObstype_Vectorfield.get_modelunit
+      ~ModelObstype_Vectorfield.get_orig_name
+      ~ModelObstype_Vectorfield.get_plot_y_label
+      ~ModelObstype_Vectorfield.get_standard_unit
+      ~ModelObstype_Vectorfield.get_u_column
+      ~ModelObstype_Vectorfield.get_v_column
+      ~ModelObstype_Vectorfield.has_mapped_band
+      ~ModelObstype_Vectorfield.set_description
+      ~ModelObstype_Vectorfield.set_original_name
+      ~ModelObstype_Vectorfield.set_original_unit
+      ~ModelObstype_Vectorfield.test_if_unit_is_known
+   
+   
+
+   
+   
+   
