+++ _sources/reference/api/metobs_toolkit.analysis.Analysis.apply_filter.rst.txt
@@ -1,0 +1,6 @@
+metobs\_toolkit.analysis.Analysis.apply\_filter
+===============================================
+
+.. currentmodule:: metobs_toolkit.analysis
+
+.. automethod:: Analysis.apply_filter
