+++ _sources/reference/api/metobs_toolkit.Obstype.get_standard_unit.rst.txt
@@ -1,0 +1,6 @@
+metobs\_toolkit.Obstype.get\_standard\_unit
+===========================================
+
+.. currentmodule:: metobs_toolkit
+
+.. automethod:: Obstype.get_standard_unit
