### 2024-03-08 1b14022
working version of the gapfill example
